feat: button to report issues
