--- ovos_skill_config/static/assets/index-sHOxXLxz.js
@@ -1,121 +1,0 @@
-function tp(s,a){for(var u=0;u<a.length;u++){const f=a[u];if(typeof f!="string"&&!Array.isArray(f)){for(const m in f)if(m!=="default"&&!(m in s)){const h=Object.getOwnPropertyDescriptor(f,m);h&&Object.defineProperty(s,m,h.get?h:{enumerable:!0,get:()=>f[m]})}}}return Object.freeze(Object.defineProperty(s,Symbol.toStringTag,{value:"Module"}))}(function(){const a=document.createElement("link").relList;if(a&&a.supports&&a.supports("modulepreload"))return;for(const m of document.querySelectorAll('link[rel="modulepreload"]'))f(m);new MutationObserver(m=>{for(const h of m)if(h.type==="childList")for(const k of h.addedNodes)k.tagName==="LINK"&&k.rel==="modulepreload"&&f(k)}).observe(document,{childList:!0,subtree:!0});function u(m){const h={};return m.integrity&&(h.integrity=m.integrity),m.referrerPolicy&&(h.referrerPolicy=m.referrerPolicy),m.crossOrigin==="use-credentials"?h.credentials="include":m.crossOrigin==="anonymous"?h.credentials="omit":h.credentials="same-origin",h}function f(m){if(m.ep)return;m.ep=!0;const h=u(m);fetch(m.href,h)}})();function np(s){return s&&s.__esModule&&Object.prototype.hasOwnProperty.call(s,"default")?s.default:s}var Gi={exports:{}},Rr={},Ki={exports:{}},te={};/**
- * @license React
- * react.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var sc;function rp(){if(sc)return te;sc=1;var s=Symbol.for("react.element"),a=Symbol.for("react.portal"),u=Symbol.for("react.fragment"),f=Symbol.for("react.strict_mode"),m=Symbol.for("react.profiler"),h=Symbol.for("react.provider"),k=Symbol.for("react.context"),x=Symbol.for("react.forward_ref"),C=Symbol.for("react.suspense"),j=Symbol.for("react.memo"),I=Symbol.for("react.lazy"),A=Symbol.iterator;function F(v){return v===null||typeof v!="object"?null:(v=A&&v[A]||v["@@iterator"],typeof v=="function"?v:null)}var Q={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},z=Object.assign,M={};function R(v,N,ee){this.props=v,this.context=N,this.refs=M,this.updater=ee||Q}R.prototype.isReactComponent={},R.prototype.setState=function(v,N){if(typeof v!="object"&&typeof v!="function"&&v!=null)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,v,N,"setState")},R.prototype.forceUpdate=function(v){this.updater.enqueueForceUpdate(this,v,"forceUpdate")};function U(){}U.prototype=R.prototype;function L(v,N,ee){this.props=v,this.context=N,this.refs=M,this.updater=ee||Q}var b=L.prototype=new U;b.constructor=L,z(b,R.prototype),b.isPureReactComponent=!0;var J=Array.isArray,le=Object.prototype.hasOwnProperty,de={current:null},xe={key:!0,ref:!0,__self:!0,__source:!0};function _e(v,N,ee){var ne,ie={},se=null,fe=null;if(N!=null)for(ne in N.ref!==void 0&&(fe=N.ref),N.key!==void 0&&(se=""+N.key),N)le.call(N,ne)&&!xe.hasOwnProperty(ne)&&(ie[ne]=N[ne]);var ae=arguments.length-2;if(ae===1)ie.children=ee;else if(1<ae){for(var ye=Array(ae),et=0;et<ae;et++)ye[et]=arguments[et+2];ie.children=ye}if(v&&v.defaultProps)for(ne in ae=v.defaultProps,ae)ie[ne]===void 0&&(ie[ne]=ae[ne]);return{$$typeof:s,type:v,key:se,ref:fe,props:ie,_owner:de.current}}function De(v,N){return{$$typeof:s,type:v.type,key:N,ref:v.ref,props:v.props,_owner:v._owner}}function Fe(v){return typeof v=="object"&&v!==null&&v.$$typeof===s}function ft(v){var N={"=":"=0",":":"=2"};return"$"+v.replace(/[=:]/g,function(ee){return N[ee]})}var re=/\/+/g;function Ue(v,N){return typeof v=="object"&&v!==null&&v.key!=null?ft(""+v.key):N.toString(36)}function Ve(v,N,ee,ne,ie){var se=typeof v;(se==="undefined"||se==="boolean")&&(v=null);var fe=!1;if(v===null)fe=!0;else switch(se){case"string":case"number":fe=!0;break;case"object":switch(v.$$typeof){case s:case a:fe=!0}}if(fe)return fe=v,ie=ie(fe),v=ne===""?"."+Ue(fe,0):ne,J(ie)?(ee="",v!=null&&(ee=v.replace(re,"$&/")+"/"),Ve(ie,N,ee,"",function(et){return et})):ie!=null&&(Fe(ie)&&(ie=De(ie,ee+(!ie.key||fe&&fe.key===ie.key?"":(""+ie.key).replace(re,"$&/")+"/")+v)),N.push(ie)),1;if(fe=0,ne=ne===""?".":ne+":",J(v))for(var ae=0;ae<v.length;ae++){se=v[ae];var ye=ne+Ue(se,ae);fe+=Ve(se,N,ee,ye,ie)}else if(ye=F(v),typeof ye=="function")for(v=ye.call(v),ae=0;!(se=v.next()).done;)se=se.value,ye=ne+Ue(se,ae++),fe+=Ve(se,N,ee,ye,ie);else if(se==="object")throw N=String(v),Error("Objects are not valid as a React child (found: "+(N==="[object Object]"?"object with keys {"+Object.keys(v).join(", ")+"}":N)+"). If you meant to render a collection of children, use an array instead.");return fe}function qe(v,N,ee){if(v==null)return v;var ne=[],ie=0;return Ve(v,ne,"","",function(se){return N.call(ee,se,ie++)}),ne}function Re(v){if(v._status===-1){var N=v._result;N=N(),N.then(function(ee){(v._status===0||v._status===-1)&&(v._status=1,v._result=ee)},function(ee){(v._status===0||v._status===-1)&&(v._status=2,v._result=ee)}),v._status===-1&&(v._status=0,v._result=N)}if(v._status===1)return v._result.default;throw v._result}var me={current:null},D={transition:null},K={ReactCurrentDispatcher:me,ReactCurrentBatchConfig:D,ReactCurrentOwner:de};function $(){throw Error("act(...) is not supported in production builds of React.")}return te.Children={map:qe,forEach:function(v,N,ee){qe(v,function(){N.apply(this,arguments)},ee)},count:function(v){var N=0;return qe(v,function(){N++}),N},toArray:function(v){return qe(v,function(N){return N})||[]},only:function(v){if(!Fe(v))throw Error("React.Children.only expected to receive a single React element child.");return v}},te.Component=R,te.Fragment=u,te.Profiler=m,te.PureComponent=L,te.StrictMode=f,te.Suspense=C,te.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=K,te.act=$,te.cloneElement=function(v,N,ee){if(v==null)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+v+".");var ne=z({},v.props),ie=v.key,se=v.ref,fe=v._owner;if(N!=null){if(N.ref!==void 0&&(se=N.ref,fe=de.current),N.key!==void 0&&(ie=""+N.key),v.type&&v.type.defaultProps)var ae=v.type.defaultProps;for(ye in N)le.call(N,ye)&&!xe.hasOwnProperty(ye)&&(ne[ye]=N[ye]===void 0&&ae!==void 0?ae[ye]:N[ye])}var ye=arguments.length-2;if(ye===1)ne.children=ee;else if(1<ye){ae=Array(ye);for(var et=0;et<ye;et++)ae[et]=arguments[et+2];ne.children=ae}return{$$typeof:s,type:v.type,key:ie,ref:se,props:ne,_owner:fe}},te.createContext=function(v){return v={$$typeof:k,_currentValue:v,_currentValue2:v,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null},v.Provider={$$typeof:h,_context:v},v.Consumer=v},te.createElement=_e,te.createFactory=function(v){var N=_e.bind(null,v);return N.type=v,N},te.createRef=function(){return{current:null}},te.forwardRef=function(v){return{$$typeof:x,render:v}},te.isValidElement=Fe,te.lazy=function(v){return{$$typeof:I,_payload:{_status:-1,_result:v},_init:Re}},te.memo=function(v,N){return{$$typeof:j,type:v,compare:N===void 0?null:N}},te.startTransition=function(v){var N=D.transition;D.transition={};try{v()}finally{D.transition=N}},te.unstable_act=$,te.useCallback=function(v,N){return me.current.useCallback(v,N)},te.useContext=function(v){return me.current.useContext(v)},te.useDebugValue=function(){},te.useDeferredValue=function(v){return me.current.useDeferredValue(v)},te.useEffect=function(v,N){return me.current.useEffect(v,N)},te.useId=function(){return me.current.useId()},te.useImperativeHandle=function(v,N,ee){return me.current.useImperativeHandle(v,N,ee)},te.useInsertionEffect=function(v,N){return me.current.useInsertionEffect(v,N)},te.useLayoutEffect=function(v,N){return me.current.useLayoutEffect(v,N)},te.useMemo=function(v,N){return me.current.useMemo(v,N)},te.useReducer=function(v,N,ee){return me.current.useReducer(v,N,ee)},te.useRef=function(v){return me.current.useRef(v)},te.useState=function(v){return me.current.useState(v)},te.useSyncExternalStore=function(v,N,ee){return me.current.useSyncExternalStore(v,N,ee)},te.useTransition=function(){return me.current.useTransition()},te.version="18.3.1",te}var uc;function ns(){return uc||(uc=1,Ki.exports=rp()),Ki.exports}/**
- * @license React
- * react-jsx-runtime.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var ac;function op(){if(ac)return Rr;ac=1;var s=ns(),a=Symbol.for("react.element"),u=Symbol.for("react.fragment"),f=Object.prototype.hasOwnProperty,m=s.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,h={key:!0,ref:!0,__self:!0,__source:!0};function k(x,C,j){var I,A={},F=null,Q=null;j!==void 0&&(F=""+j),C.key!==void 0&&(F=""+C.key),C.ref!==void 0&&(Q=C.ref);for(I in C)f.call(C,I)&&!h.hasOwnProperty(I)&&(A[I]=C[I]);if(x&&x.defaultProps)for(I in C=x.defaultProps,C)A[I]===void 0&&(A[I]=C[I]);return{$$typeof:a,type:x,key:F,ref:Q,props:A,_owner:m.current}}return Rr.Fragment=u,Rr.jsx=k,Rr.jsxs=k,Rr}var cc;function lp(){return cc||(cc=1,Gi.exports=op()),Gi.exports}var S=lp(),O=ns();const je=np(O),ip=tp({__proto__:null,default:je},[O]);var Ko={},Yi={exports:{}},be={},Xi={exports:{}},Zi={};/**
- * @license React
- * scheduler.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var dc;function sp(){return dc||(dc=1,function(s){function a(D,K){var $=D.length;D.push(K);e:for(;0<$;){var v=$-1>>>1,N=D[v];if(0<m(N,K))D[v]=K,D[$]=N,$=v;else break e}}function u(D){return D.length===0?null:D[0]}function f(D){if(D.length===0)return null;var K=D[0],$=D.pop();if($!==K){D[0]=$;e:for(var v=0,N=D.length,ee=N>>>1;v<ee;){var ne=2*(v+1)-1,ie=D[ne],se=ne+1,fe=D[se];if(0>m(ie,$))se<N&&0>m(fe,ie)?(D[v]=fe,D[se]=$,v=se):(D[v]=ie,D[ne]=$,v=ne);else if(se<N&&0>m(fe,$))D[v]=fe,D[se]=$,v=se;else break e}}return K}function m(D,K){var $=D.sortIndex-K.sortIndex;return $!==0?$:D.id-K.id}if(typeof performance=="object"&&typeof performance.now=="function"){var h=performance;s.unstable_now=function(){return h.now()}}else{var k=Date,x=k.now();s.unstable_now=function(){return k.now()-x}}var C=[],j=[],I=1,A=null,F=3,Q=!1,z=!1,M=!1,R=typeof setTimeout=="function"?setTimeout:null,U=typeof clearTimeout=="function"?clearTimeout:null,L=typeof setImmediate<"u"?setImmediate:null;typeof navigator<"u"&&navigator.scheduling!==void 0&&navigator.scheduling.isInputPending!==void 0&&navigator.scheduling.isInputPending.bind(navigator.scheduling);function b(D){for(var K=u(j);K!==null;){if(K.callback===null)f(j);else if(K.startTime<=D)f(j),K.sortIndex=K.expirationTime,a(C,K);else break;K=u(j)}}function J(D){if(M=!1,b(D),!z)if(u(C)!==null)z=!0,Re(le);else{var K=u(j);K!==null&&me(J,K.startTime-D)}}function le(D,K){z=!1,M&&(M=!1,U(_e),_e=-1),Q=!0;var $=F;try{for(b(K),A=u(C);A!==null&&(!(A.expirationTime>K)||D&&!ft());){var v=A.callback;if(typeof v=="function"){A.callback=null,F=A.priorityLevel;var N=v(A.expirationTime<=K);K=s.unstable_now(),typeof N=="function"?A.callback=N:A===u(C)&&f(C),b(K)}else f(C);A=u(C)}if(A!==null)var ee=!0;else{var ne=u(j);ne!==null&&me(J,ne.startTime-K),ee=!1}return ee}finally{A=null,F=$,Q=!1}}var de=!1,xe=null,_e=-1,De=5,Fe=-1;function ft(){return!(s.unstable_now()-Fe<De)}function re(){if(xe!==null){var D=s.unstable_now();Fe=D;var K=!0;try{K=xe(!0,D)}finally{K?Ue():(de=!1,xe=null)}}else de=!1}var Ue;if(typeof L=="function")Ue=function(){L(re)};else if(typeof MessageChannel<"u"){var Ve=new MessageChannel,qe=Ve.port2;Ve.port1.onmessage=re,Ue=function(){qe.postMessage(null)}}else Ue=function(){R(re,0)};function Re(D){xe=D,de||(de=!0,Ue())}function me(D,K){_e=R(function(){D(s.unstable_now())},K)}s.unstable_IdlePriority=5,s.unstable_ImmediatePriority=1,s.unstable_LowPriority=4,s.unstable_NormalPriority=3,s.unstable_Profiling=null,s.unstable_UserBlockingPriority=2,s.unstable_cancelCallback=function(D){D.callback=null},s.unstable_continueExecution=function(){z||Q||(z=!0,Re(le))},s.unstable_forceFrameRate=function(D){0>D||125<D?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):De=0<D?Math.floor(1e3/D):5},s.unstable_getCurrentPriorityLevel=function(){return F},s.unstable_getFirstCallbackNode=function(){return u(C)},s.unstable_next=function(D){switch(F){case 1:case 2:case 3:var K=3;break;default:K=F}var $=F;F=K;try{return D()}finally{F=$}},s.unstable_pauseExecution=function(){},s.unstable_requestPaint=function(){},s.unstable_runWithPriority=function(D,K){switch(D){case 1:case 2:case 3:case 4:case 5:break;default:D=3}var $=F;F=D;try{return K()}finally{F=$}},s.unstable_scheduleCallback=function(D,K,$){var v=s.unstable_now();switch(typeof $=="object"&&$!==null?($=$.delay,$=typeof $=="number"&&0<$?v+$:v):$=v,D){case 1:var N=-1;break;case 2:N=250;break;case 5:N=1073741823;break;case 4:N=1e4;break;default:N=5e3}return N=$+N,D={id:I++,callback:K,priorityLevel:D,startTime:$,expirationTime:N,sortIndex:-1},$>v?(D.sortIndex=$,a(j,D),u(C)===null&&D===u(j)&&(M?(U(_e),_e=-1):M=!0,me(J,$-v))):(D.sortIndex=N,a(C,D),z||Q||(z=!0,Re(le))),D},s.unstable_shouldYield=ft,s.unstable_wrapCallback=function(D){var K=F;return function(){var $=F;F=K;try{return D.apply(this,arguments)}finally{F=$}}}}(Zi)),Zi}var fc;function up(){return fc||(fc=1,Xi.exports=sp()),Xi.exports}/**
- * @license React
- * react-dom.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var pc;function ap(){if(pc)return be;pc=1;var s=ns(),a=up();function u(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var f=new Set,m={};function h(e,t){k(e,t),k(e+"Capture",t)}function k(e,t){for(m[e]=t,e=0;e<t.length;e++)f.add(t[e])}var x=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),C=Object.prototype.hasOwnProperty,j=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,I={},A={};function F(e){return C.call(A,e)?!0:C.call(I,e)?!1:j.test(e)?A[e]=!0:(I[e]=!0,!1)}function Q(e,t,n,r){if(n!==null&&n.type===0)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return r?!1:n!==null?!n.acceptsBooleans:(e=e.toLowerCase().slice(0,5),e!=="data-"&&e!=="aria-");default:return!1}}function z(e,t,n,r){if(t===null||typeof t>"u"||Q(e,t,n,r))return!0;if(r)return!1;if(n!==null)switch(n.type){case 3:return!t;case 4:return t===!1;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}function M(e,t,n,r,o,l,i){this.acceptsBooleans=t===2||t===3||t===4,this.attributeName=r,this.attributeNamespace=o,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=l,this.removeEmptyString=i}var R={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach(function(e){R[e]=new M(e,0,!1,e,null,!1,!1)}),[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach(function(e){var t=e[0];R[t]=new M(t,1,!1,e[1],null,!1,!1)}),["contentEditable","draggable","spellCheck","value"].forEach(function(e){R[e]=new M(e,2,!1,e.toLowerCase(),null,!1,!1)}),["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach(function(e){R[e]=new M(e,2,!1,e,null,!1,!1)}),"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach(function(e){R[e]=new M(e,3,!1,e.toLowerCase(),null,!1,!1)}),["checked","multiple","muted","selected"].forEach(function(e){R[e]=new M(e,3,!0,e,null,!1,!1)}),["capture","download"].forEach(function(e){R[e]=new M(e,4,!1,e,null,!1,!1)}),["cols","rows","size","span"].forEach(function(e){R[e]=new M(e,6,!1,e,null,!1,!1)}),["rowSpan","start"].forEach(function(e){R[e]=new M(e,5,!1,e.toLowerCase(),null,!1,!1)});var U=/[\-:]([a-z])/g;function L(e){return e[1].toUpperCase()}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach(function(e){var t=e.replace(U,L);R[t]=new M(t,1,!1,e,null,!1,!1)}),"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach(function(e){var t=e.replace(U,L);R[t]=new M(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)}),["xml:base","xml:lang","xml:space"].forEach(function(e){var t=e.replace(U,L);R[t]=new M(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)}),["tabIndex","crossOrigin"].forEach(function(e){R[e]=new M(e,1,!1,e.toLowerCase(),null,!1,!1)}),R.xlinkHref=new M("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1),["src","href","action","formAction"].forEach(function(e){R[e]=new M(e,1,!1,e.toLowerCase(),null,!0,!0)});function b(e,t,n,r){var o=R.hasOwnProperty(t)?R[t]:null;(o!==null?o.type!==0:r||!(2<t.length)||t[0]!=="o"&&t[0]!=="O"||t[1]!=="n"&&t[1]!=="N")&&(z(t,n,o,r)&&(n=null),r||o===null?F(t)&&(n===null?e.removeAttribute(t):e.setAttribute(t,""+n)):o.mustUseProperty?e[o.propertyName]=n===null?o.type===3?!1:"":n:(t=o.attributeName,r=o.attributeNamespace,n===null?e.removeAttribute(t):(o=o.type,n=o===3||o===4&&n===!0?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}var J=s.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,le=Symbol.for("react.element"),de=Symbol.for("react.portal"),xe=Symbol.for("react.fragment"),_e=Symbol.for("react.strict_mode"),De=Symbol.for("react.profiler"),Fe=Symbol.for("react.provider"),ft=Symbol.for("react.context"),re=Symbol.for("react.forward_ref"),Ue=Symbol.for("react.suspense"),Ve=Symbol.for("react.suspense_list"),qe=Symbol.for("react.memo"),Re=Symbol.for("react.lazy"),me=Symbol.for("react.offscreen"),D=Symbol.iterator;function K(e){return e===null||typeof e!="object"?null:(e=D&&e[D]||e["@@iterator"],typeof e=="function"?e:null)}var $=Object.assign,v;function N(e){if(v===void 0)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);v=t&&t[1]||""}return`
-`+v+e}var ee=!1;function ne(e,t){if(!e||ee)return"";ee=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),typeof Reflect=="object"&&Reflect.construct){try{Reflect.construct(t,[])}catch(w){var r=w}Reflect.construct(e,[],t)}else{try{t.call()}catch(w){r=w}e.call(t.prototype)}else{try{throw Error()}catch(w){r=w}e()}}catch(w){if(w&&r&&typeof w.stack=="string"){for(var o=w.stack.split(`
-`),l=r.stack.split(`
-`),i=o.length-1,c=l.length-1;1<=i&&0<=c&&o[i]!==l[c];)c--;for(;1<=i&&0<=c;i--,c--)if(o[i]!==l[c]){if(i!==1||c!==1)do if(i--,c--,0>c||o[i]!==l[c]){var d=`
-`+o[i].replace(" at new "," at ");return e.displayName&&d.includes("<anonymous>")&&(d=d.replace("<anonymous>",e.displayName)),d}while(1<=i&&0<=c);break}}}finally{ee=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?N(e):""}function ie(e){switch(e.tag){case 5:return N(e.type);case 16:return N("Lazy");case 13:return N("Suspense");case 19:return N("SuspenseList");case 0:case 2:case 15:return e=ne(e.type,!1),e;case 11:return e=ne(e.type.render,!1),e;case 1:return e=ne(e.type,!0),e;default:return""}}function se(e){if(e==null)return null;if(typeof e=="function")return e.displayName||e.name||null;if(typeof e=="string")return e;switch(e){case xe:return"Fragment";case de:return"Portal";case De:return"Profiler";case _e:return"StrictMode";case Ue:return"Suspense";case Ve:return"SuspenseList"}if(typeof e=="object")switch(e.$$typeof){case ft:return(e.displayName||"Context")+".Consumer";case Fe:return(e._context.displayName||"Context")+".Provider";case re:var t=e.render;return e=e.displayName,e||(e=t.displayName||t.name||"",e=e!==""?"ForwardRef("+e+")":"ForwardRef"),e;case qe:return t=e.displayName||null,t!==null?t:se(e.type)||"Memo";case Re:t=e._payload,e=e._init;try{return se(e(t))}catch{}}return null}function fe(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=t.render,e=e.displayName||e.name||"",t.displayName||(e!==""?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return se(t);case 8:return t===_e?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if(typeof t=="function")return t.displayName||t.name||null;if(typeof t=="string")return t}return null}function ae(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":return e;case"object":return e;default:return""}}function ye(e){var t=e.type;return(e=e.nodeName)&&e.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function et(e){var t=ye(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&typeof n<"u"&&typeof n.get=="function"&&typeof n.set=="function"){var o=n.get,l=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return o.call(this)},set:function(i){r=""+i,l.call(this,i)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(i){r=""+i},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}function Ar(e){e._valueTracker||(e._valueTracker=et(e))}function ps(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=ye(e)?e.checked?"true":"false":e.value),e=r,e!==n?(t.setValue(e),!0):!1}function Dr(e){if(e=e||(typeof document<"u"?document:void 0),typeof e>"u")return null;try{return e.activeElement||e.body}catch{return e.body}}function qo(e,t){var n=t.checked;return $({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:n??e._wrapperState.initialChecked})}function ms(e,t){var n=t.defaultValue==null?"":t.defaultValue,r=t.checked!=null?t.checked:t.defaultChecked;n=ae(t.value!=null?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:t.type==="checkbox"||t.type==="radio"?t.checked!=null:t.value!=null}}function hs(e,t){t=t.checked,t!=null&&b(e,"checked",t,!1)}function el(e,t){hs(e,t);var n=ae(t.value),r=t.type;if(n!=null)r==="number"?(n===0&&e.value===""||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if(r==="submit"||r==="reset"){e.removeAttribute("value");return}t.hasOwnProperty("value")?tl(e,t.type,n):t.hasOwnProperty("defaultValue")&&tl(e,t.type,ae(t.defaultValue)),t.checked==null&&t.defaultChecked!=null&&(e.defaultChecked=!!t.defaultChecked)}function vs(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!(r!=="submit"&&r!=="reset"||t.value!==void 0&&t.value!==null))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}n=e.name,n!==""&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,n!==""&&(e.name=n)}function tl(e,t,n){(t!=="number"||Dr(e.ownerDocument)!==e)&&(n==null?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var Hn=Array.isArray;function gn(e,t,n,r){if(e=e.options,t){t={};for(var o=0;o<n.length;o++)t["$"+n[o]]=!0;for(n=0;n<e.length;n++)o=t.hasOwnProperty("$"+e[n].value),e[n].selected!==o&&(e[n].selected=o),o&&r&&(e[n].defaultSelected=!0)}else{for(n=""+ae(n),t=null,o=0;o<e.length;o++){if(e[o].value===n){e[o].selected=!0,r&&(e[o].defaultSelected=!0);return}t!==null||e[o].disabled||(t=e[o])}t!==null&&(t.selected=!0)}}function nl(e,t){if(t.dangerouslySetInnerHTML!=null)throw Error(u(91));return $({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function gs(e,t){var n=t.value;if(n==null){if(n=t.children,t=t.defaultValue,n!=null){if(t!=null)throw Error(u(92));if(Hn(n)){if(1<n.length)throw Error(u(93));n=n[0]}t=n}t==null&&(t=""),n=t}e._wrapperState={initialValue:ae(n)}}function ys(e,t){var n=ae(t.value),r=ae(t.defaultValue);n!=null&&(n=""+n,n!==e.value&&(e.value=n),t.defaultValue==null&&e.defaultValue!==n&&(e.defaultValue=n)),r!=null&&(e.defaultValue=""+r)}function ws(e){var t=e.textContent;t===e._wrapperState.initialValue&&t!==""&&t!==null&&(e.value=t)}function xs(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function rl(e,t){return e==null||e==="http://www.w3.org/1999/xhtml"?xs(t):e==="http://www.w3.org/2000/svg"&&t==="foreignObject"?"http://www.w3.org/1999/xhtml":e}var Fr,ks=function(e){return typeof MSApp<"u"&&MSApp.execUnsafeLocalFunction?function(t,n,r,o){MSApp.execUnsafeLocalFunction(function(){return e(t,n,r,o)})}:e}(function(e,t){if(e.namespaceURI!=="http://www.w3.org/2000/svg"||"innerHTML"in e)e.innerHTML=t;else{for(Fr=Fr||document.createElement("div"),Fr.innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=Fr.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}});function Qn(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&n.nodeType===3){n.nodeValue=t;return}}e.textContent=t}var Gn={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},ld=["Webkit","ms","Moz","O"];Object.keys(Gn).forEach(function(e){ld.forEach(function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),Gn[t]=Gn[e]})});function Ss(e,t,n){return t==null||typeof t=="boolean"||t===""?"":n||typeof t!="number"||t===0||Gn.hasOwnProperty(e)&&Gn[e]?(""+t).trim():t+"px"}function Cs(e,t){e=e.style;for(var n in t)if(t.hasOwnProperty(n)){var r=n.indexOf("--")===0,o=Ss(n,t[n],r);n==="float"&&(n="cssFloat"),r?e.setProperty(n,o):e[n]=o}}var id=$({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function ol(e,t){if(t){if(id[e]&&(t.children!=null||t.dangerouslySetInnerHTML!=null))throw Error(u(137,e));if(t.dangerouslySetInnerHTML!=null){if(t.children!=null)throw Error(u(60));if(typeof t.dangerouslySetInnerHTML!="object"||!("__html"in t.dangerouslySetInnerHTML))throw Error(u(61))}if(t.style!=null&&typeof t.style!="object")throw Error(u(62))}}function ll(e,t){if(e.indexOf("-")===-1)return typeof t.is=="string";switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var il=null;function sl(e){return e=e.target||e.srcElement||window,e.correspondingUseElement&&(e=e.correspondingUseElement),e.nodeType===3?e.parentNode:e}var ul=null,yn=null,wn=null;function Es(e){if(e=mr(e)){if(typeof ul!="function")throw Error(u(280));var t=e.stateNode;t&&(t=so(t),ul(e.stateNode,e.type,t))}}function Ns(e){yn?wn?wn.push(e):wn=[e]:yn=e}function _s(){if(yn){var e=yn,t=wn;if(wn=yn=null,Es(e),t)for(e=0;e<t.length;e++)Es(t[e])}}function Ps(e,t){return e(t)}function js(){}var al=!1;function Rs(e,t,n){if(al)return e(t,n);al=!0;try{return Ps(e,t,n)}finally{al=!1,(yn!==null||wn!==null)&&(js(),_s())}}function Kn(e,t){var n=e.stateNode;if(n===null)return null;var r=so(n);if(r===null)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(e=e.type,r=!(e==="button"||e==="input"||e==="select"||e==="textarea")),e=!r;break e;default:e=!1}if(e)return null;if(n&&typeof n!="function")throw Error(u(231,t,typeof n));return n}var cl=!1;if(x)try{var Yn={};Object.defineProperty(Yn,"passive",{get:function(){cl=!0}}),window.addEventListener("test",Yn,Yn),window.removeEventListener("test",Yn,Yn)}catch{cl=!1}function sd(e,t,n,r,o,l,i,c,d){var w=Array.prototype.slice.call(arguments,3);try{t.apply(n,w)}catch(_){this.onError(_)}}var Xn=!1,Ur=null,Vr=!1,dl=null,ud={onError:function(e){Xn=!0,Ur=e}};function ad(e,t,n,r,o,l,i,c,d){Xn=!1,Ur=null,sd.apply(ud,arguments)}function cd(e,t,n,r,o,l,i,c,d){if(ad.apply(this,arguments),Xn){if(Xn){var w=Ur;Xn=!1,Ur=null}else throw Error(u(198));Vr||(Vr=!0,dl=w)}}function rn(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do t=e,t.flags&4098&&(n=t.return),e=t.return;while(e)}return t.tag===3?n:null}function zs(e){if(e.tag===13){var t=e.memoizedState;if(t===null&&(e=e.alternate,e!==null&&(t=e.memoizedState)),t!==null)return t.dehydrated}return null}function Ts(e){if(rn(e)!==e)throw Error(u(188))}function dd(e){var t=e.alternate;if(!t){if(t=rn(e),t===null)throw Error(u(188));return t!==e?null:e}for(var n=e,r=t;;){var o=n.return;if(o===null)break;var l=o.alternate;if(l===null){if(r=o.return,r!==null){n=r;continue}break}if(o.child===l.child){for(l=o.child;l;){if(l===n)return Ts(o),e;if(l===r)return Ts(o),t;l=l.sibling}throw Error(u(188))}if(n.return!==r.return)n=o,r=l;else{for(var i=!1,c=o.child;c;){if(c===n){i=!0,n=o,r=l;break}if(c===r){i=!0,r=o,n=l;break}c=c.sibling}if(!i){for(c=l.child;c;){if(c===n){i=!0,n=l,r=o;break}if(c===r){i=!0,r=l,n=o;break}c=c.sibling}if(!i)throw Error(u(189))}}if(n.alternate!==r)throw Error(u(190))}if(n.tag!==3)throw Error(u(188));return n.stateNode.current===n?e:t}function Ms(e){return e=dd(e),e!==null?Ls(e):null}function Ls(e){if(e.tag===5||e.tag===6)return e;for(e=e.child;e!==null;){var t=Ls(e);if(t!==null)return t;e=e.sibling}return null}var Os=a.unstable_scheduleCallback,Is=a.unstable_cancelCallback,fd=a.unstable_shouldYield,pd=a.unstable_requestPaint,Ee=a.unstable_now,md=a.unstable_getCurrentPriorityLevel,fl=a.unstable_ImmediatePriority,As=a.unstable_UserBlockingPriority,$r=a.unstable_NormalPriority,hd=a.unstable_LowPriority,Ds=a.unstable_IdlePriority,Wr=null,xt=null;function vd(e){if(xt&&typeof xt.onCommitFiberRoot=="function")try{xt.onCommitFiberRoot(Wr,e,void 0,(e.current.flags&128)===128)}catch{}}var pt=Math.clz32?Math.clz32:wd,gd=Math.log,yd=Math.LN2;function wd(e){return e>>>=0,e===0?32:31-(gd(e)/yd|0)|0}var Br=64,Hr=4194304;function Zn(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return e&4194240;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return e&130023424;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function Qr(e,t){var n=e.pendingLanes;if(n===0)return 0;var r=0,o=e.suspendedLanes,l=e.pingedLanes,i=n&268435455;if(i!==0){var c=i&~o;c!==0?r=Zn(c):(l&=i,l!==0&&(r=Zn(l)))}else i=n&~o,i!==0?r=Zn(i):l!==0&&(r=Zn(l));if(r===0)return 0;if(t!==0&&t!==r&&!(t&o)&&(o=r&-r,l=t&-t,o>=l||o===16&&(l&4194240)!==0))return t;if(r&4&&(r|=n&16),t=e.entangledLanes,t!==0)for(e=e.entanglements,t&=r;0<t;)n=31-pt(t),o=1<<n,r|=e[n],t&=~o;return r}function xd(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return-1;case 134217728:case 268435456:case 536870912:case 1073741824:return-1;default:return-1}}function kd(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,o=e.expirationTimes,l=e.pendingLanes;0<l;){var i=31-pt(l),c=1<<i,d=o[i];d===-1?(!(c&n)||c&r)&&(o[i]=xd(c,t)):d<=t&&(e.expiredLanes|=c),l&=~c}}function pl(e){return e=e.pendingLanes&-1073741825,e!==0?e:e&1073741824?1073741824:0}function Fs(){var e=Br;return Br<<=1,!(Br&4194240)&&(Br=64),e}function ml(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function Jn(e,t,n){e.pendingLanes|=t,t!==536870912&&(e.suspendedLanes=0,e.pingedLanes=0),e=e.eventTimes,t=31-pt(t),e[t]=n}function Sd(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var o=31-pt(n),l=1<<o;t[o]=0,r[o]=-1,e[o]=-1,n&=~l}}function hl(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-pt(n),o=1<<r;o&t|e[r]&t&&(e[r]|=t),n&=~o}}var ce=0;function Us(e){return e&=-e,1<e?4<e?e&268435455?16:536870912:4:1}var Vs,vl,$s,Ws,Bs,gl=!1,Gr=[],At=null,Dt=null,Ft=null,bn=new Map,qn=new Map,Ut=[],Cd="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function Hs(e,t){switch(e){case"focusin":case"focusout":At=null;break;case"dragenter":case"dragleave":Dt=null;break;case"mouseover":case"mouseout":Ft=null;break;case"pointerover":case"pointerout":bn.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":qn.delete(t.pointerId)}}function er(e,t,n,r,o,l){return e===null||e.nativeEvent!==l?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:l,targetContainers:[o]},t!==null&&(t=mr(t),t!==null&&vl(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,o!==null&&t.indexOf(o)===-1&&t.push(o),e)}function Ed(e,t,n,r,o){switch(t){case"focusin":return At=er(At,e,t,n,r,o),!0;case"dragenter":return Dt=er(Dt,e,t,n,r,o),!0;case"mouseover":return Ft=er(Ft,e,t,n,r,o),!0;case"pointerover":var l=o.pointerId;return bn.set(l,er(bn.get(l)||null,e,t,n,r,o)),!0;case"gotpointercapture":return l=o.pointerId,qn.set(l,er(qn.get(l)||null,e,t,n,r,o)),!0}return!1}function Qs(e){var t=on(e.target);if(t!==null){var n=rn(t);if(n!==null){if(t=n.tag,t===13){if(t=zs(n),t!==null){e.blockedOn=t,Bs(e.priority,function(){$s(n)});return}}else if(t===3&&n.stateNode.current.memoizedState.isDehydrated){e.blockedOn=n.tag===3?n.stateNode.containerInfo:null;return}}}e.blockedOn=null}function Kr(e){if(e.blockedOn!==null)return!1;for(var t=e.targetContainers;0<t.length;){var n=wl(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(n===null){n=e.nativeEvent;var r=new n.constructor(n.type,n);il=r,n.target.dispatchEvent(r),il=null}else return t=mr(n),t!==null&&vl(t),e.blockedOn=n,!1;t.shift()}return!0}function Gs(e,t,n){Kr(e)&&n.delete(t)}function Nd(){gl=!1,At!==null&&Kr(At)&&(At=null),Dt!==null&&Kr(Dt)&&(Dt=null),Ft!==null&&Kr(Ft)&&(Ft=null),bn.forEach(Gs),qn.forEach(Gs)}function tr(e,t){e.blockedOn===t&&(e.blockedOn=null,gl||(gl=!0,a.unstable_scheduleCallback(a.unstable_NormalPriority,Nd)))}function nr(e){function t(o){return tr(o,e)}if(0<Gr.length){tr(Gr[0],e);for(var n=1;n<Gr.length;n++){var r=Gr[n];r.blockedOn===e&&(r.blockedOn=null)}}for(At!==null&&tr(At,e),Dt!==null&&tr(Dt,e),Ft!==null&&tr(Ft,e),bn.forEach(t),qn.forEach(t),n=0;n<Ut.length;n++)r=Ut[n],r.blockedOn===e&&(r.blockedOn=null);for(;0<Ut.length&&(n=Ut[0],n.blockedOn===null);)Qs(n),n.blockedOn===null&&Ut.shift()}var xn=J.ReactCurrentBatchConfig,Yr=!0;function _d(e,t,n,r){var o=ce,l=xn.transition;xn.transition=null;try{ce=1,yl(e,t,n,r)}finally{ce=o,xn.transition=l}}function Pd(e,t,n,r){var o=ce,l=xn.transition;xn.transition=null;try{ce=4,yl(e,t,n,r)}finally{ce=o,xn.transition=l}}function yl(e,t,n,r){if(Yr){var o=wl(e,t,n,r);if(o===null)Al(e,t,r,Xr,n),Hs(e,r);else if(Ed(o,e,t,n,r))r.stopPropagation();else if(Hs(e,r),t&4&&-1<Cd.indexOf(e)){for(;o!==null;){var l=mr(o);if(l!==null&&Vs(l),l=wl(e,t,n,r),l===null&&Al(e,t,r,Xr,n),l===o)break;o=l}o!==null&&r.stopPropagation()}else Al(e,t,r,null,n)}}var Xr=null;function wl(e,t,n,r){if(Xr=null,e=sl(r),e=on(e),e!==null)if(t=rn(e),t===null)e=null;else if(n=t.tag,n===13){if(e=zs(t),e!==null)return e;e=null}else if(n===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return Xr=e,null}function Ks(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(md()){case fl:return 1;case As:return 4;case $r:case hd:return 16;case Ds:return 536870912;default:return 16}default:return 16}}var Vt=null,xl=null,Zr=null;function Ys(){if(Zr)return Zr;var e,t=xl,n=t.length,r,o="value"in Vt?Vt.value:Vt.textContent,l=o.length;for(e=0;e<n&&t[e]===o[e];e++);var i=n-e;for(r=1;r<=i&&t[n-r]===o[l-r];r++);return Zr=o.slice(e,1<r?1-r:void 0)}function Jr(e){var t=e.keyCode;return"charCode"in e?(e=e.charCode,e===0&&t===13&&(e=13)):e=t,e===10&&(e=13),32<=e||e===13?e:0}function br(){return!0}function Xs(){return!1}function tt(e){function t(n,r,o,l,i){this._reactName=n,this._targetInst=o,this.type=r,this.nativeEvent=l,this.target=i,this.currentTarget=null;for(var c in e)e.hasOwnProperty(c)&&(n=e[c],this[c]=n?n(l):l[c]);return this.isDefaultPrevented=(l.defaultPrevented!=null?l.defaultPrevented:l.returnValue===!1)?br:Xs,this.isPropagationStopped=Xs,this}return $(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var n=this.nativeEvent;n&&(n.preventDefault?n.preventDefault():typeof n.returnValue!="unknown"&&(n.returnValue=!1),this.isDefaultPrevented=br)},stopPropagation:function(){var n=this.nativeEvent;n&&(n.stopPropagation?n.stopPropagation():typeof n.cancelBubble!="unknown"&&(n.cancelBubble=!0),this.isPropagationStopped=br)},persist:function(){},isPersistent:br}),t}var kn={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},kl=tt(kn),rr=$({},kn,{view:0,detail:0}),jd=tt(rr),Sl,Cl,or,qr=$({},rr,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:Nl,button:0,buttons:0,relatedTarget:function(e){return e.relatedTarget===void 0?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==or&&(or&&e.type==="mousemove"?(Sl=e.screenX-or.screenX,Cl=e.screenY-or.screenY):Cl=Sl=0,or=e),Sl)},movementY:function(e){return"movementY"in e?e.movementY:Cl}}),Zs=tt(qr),Rd=$({},qr,{dataTransfer:0}),zd=tt(Rd),Td=$({},rr,{relatedTarget:0}),El=tt(Td),Md=$({},kn,{animationName:0,elapsedTime:0,pseudoElement:0}),Ld=tt(Md),Od=$({},kn,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),Id=tt(Od),Ad=$({},kn,{data:0}),Js=tt(Ad),Dd={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},Fd={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},Ud={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function Vd(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):(e=Ud[e])?!!t[e]:!1}function Nl(){return Vd}var $d=$({},rr,{key:function(e){if(e.key){var t=Dd[e.key]||e.key;if(t!=="Unidentified")return t}return e.type==="keypress"?(e=Jr(e),e===13?"Enter":String.fromCharCode(e)):e.type==="keydown"||e.type==="keyup"?Fd[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:Nl,charCode:function(e){return e.type==="keypress"?Jr(e):0},keyCode:function(e){return e.type==="keydown"||e.type==="keyup"?e.keyCode:0},which:function(e){return e.type==="keypress"?Jr(e):e.type==="keydown"||e.type==="keyup"?e.keyCode:0}}),Wd=tt($d),Bd=$({},qr,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),bs=tt(Bd),Hd=$({},rr,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:Nl}),Qd=tt(Hd),Gd=$({},kn,{propertyName:0,elapsedTime:0,pseudoElement:0}),Kd=tt(Gd),Yd=$({},qr,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),Xd=tt(Yd),Zd=[9,13,27,32],_l=x&&"CompositionEvent"in window,lr=null;x&&"documentMode"in document&&(lr=document.documentMode);var Jd=x&&"TextEvent"in window&&!lr,qs=x&&(!_l||lr&&8<lr&&11>=lr),eu=" ",tu=!1;function nu(e,t){switch(e){case"keyup":return Zd.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function ru(e){return e=e.detail,typeof e=="object"&&"data"in e?e.data:null}var Sn=!1;function bd(e,t){switch(e){case"compositionend":return ru(t);case"keypress":return t.which!==32?null:(tu=!0,eu);case"textInput":return e=t.data,e===eu&&tu?null:e;default:return null}}function qd(e,t){if(Sn)return e==="compositionend"||!_l&&nu(e,t)?(e=Ys(),Zr=xl=Vt=null,Sn=!1,e):null;switch(e){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return qs&&t.locale!=="ko"?null:t.data;default:return null}}var ef={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function ou(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t==="input"?!!ef[e.type]:t==="textarea"}function lu(e,t,n,r){Ns(r),t=oo(t,"onChange"),0<t.length&&(n=new kl("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var ir=null,sr=null;function tf(e){Cu(e,0)}function eo(e){var t=Pn(e);if(ps(t))return e}function nf(e,t){if(e==="change")return t}var iu=!1;if(x){var Pl;if(x){var jl="oninput"in document;if(!jl){var su=document.createElement("div");su.setAttribute("oninput","return;"),jl=typeof su.oninput=="function"}Pl=jl}else Pl=!1;iu=Pl&&(!document.documentMode||9<document.documentMode)}function uu(){ir&&(ir.detachEvent("onpropertychange",au),sr=ir=null)}function au(e){if(e.propertyName==="value"&&eo(sr)){var t=[];lu(t,sr,e,sl(e)),Rs(tf,t)}}function rf(e,t,n){e==="focusin"?(uu(),ir=t,sr=n,ir.attachEvent("onpropertychange",au)):e==="focusout"&&uu()}function of(e){if(e==="selectionchange"||e==="keyup"||e==="keydown")return eo(sr)}function lf(e,t){if(e==="click")return eo(t)}function sf(e,t){if(e==="input"||e==="change")return eo(t)}function uf(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}var mt=typeof Object.is=="function"?Object.is:uf;function ur(e,t){if(mt(e,t))return!0;if(typeof e!="object"||e===null||typeof t!="object"||t===null)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var o=n[r];if(!C.call(t,o)||!mt(e[o],t[o]))return!1}return!0}function cu(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function du(e,t){var n=cu(e);e=0;for(var r;n;){if(n.nodeType===3){if(r=e+n.textContent.length,e<=t&&r>=t)return{node:n,offset:t-e};e=r}e:{for(;n;){if(n.nextSibling){n=n.nextSibling;break e}n=n.parentNode}n=void 0}n=cu(n)}}function fu(e,t){return e&&t?e===t?!0:e&&e.nodeType===3?!1:t&&t.nodeType===3?fu(e,t.parentNode):"contains"in e?e.contains(t):e.compareDocumentPosition?!!(e.compareDocumentPosition(t)&16):!1:!1}function pu(){for(var e=window,t=Dr();t instanceof e.HTMLIFrameElement;){try{var n=typeof t.contentWindow.location.href=="string"}catch{n=!1}if(n)e=t.contentWindow;else break;t=Dr(e.document)}return t}function Rl(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&(t==="input"&&(e.type==="text"||e.type==="search"||e.type==="tel"||e.type==="url"||e.type==="password")||t==="textarea"||e.contentEditable==="true")}function af(e){var t=pu(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&fu(n.ownerDocument.documentElement,n)){if(r!==null&&Rl(n)){if(t=r.start,e=r.end,e===void 0&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if(e=(t=n.ownerDocument||document)&&t.defaultView||window,e.getSelection){e=e.getSelection();var o=n.textContent.length,l=Math.min(r.start,o);r=r.end===void 0?l:Math.min(r.end,o),!e.extend&&l>r&&(o=r,r=l,l=o),o=du(n,l);var i=du(n,r);o&&i&&(e.rangeCount!==1||e.anchorNode!==o.node||e.anchorOffset!==o.offset||e.focusNode!==i.node||e.focusOffset!==i.offset)&&(t=t.createRange(),t.setStart(o.node,o.offset),e.removeAllRanges(),l>r?(e.addRange(t),e.extend(i.node,i.offset)):(t.setEnd(i.node,i.offset),e.addRange(t)))}}for(t=[],e=n;e=e.parentNode;)e.nodeType===1&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for(typeof n.focus=="function"&&n.focus(),n=0;n<t.length;n++)e=t[n],e.element.scrollLeft=e.left,e.element.scrollTop=e.top}}var cf=x&&"documentMode"in document&&11>=document.documentMode,Cn=null,zl=null,ar=null,Tl=!1;function mu(e,t,n){var r=n.window===n?n.document:n.nodeType===9?n:n.ownerDocument;Tl||Cn==null||Cn!==Dr(r)||(r=Cn,"selectionStart"in r&&Rl(r)?r={start:r.selectionStart,end:r.selectionEnd}:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection(),r={anchorNode:r.anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset}),ar&&ur(ar,r)||(ar=r,r=oo(zl,"onSelect"),0<r.length&&(t=new kl("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=Cn)))}function to(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var En={animationend:to("Animation","AnimationEnd"),animationiteration:to("Animation","AnimationIteration"),animationstart:to("Animation","AnimationStart"),transitionend:to("Transition","TransitionEnd")},Ml={},hu={};x&&(hu=document.createElement("div").style,"AnimationEvent"in window||(delete En.animationend.animation,delete En.animationiteration.animation,delete En.animationstart.animation),"TransitionEvent"in window||delete En.transitionend.transition);function no(e){if(Ml[e])return Ml[e];if(!En[e])return e;var t=En[e],n;for(n in t)if(t.hasOwnProperty(n)&&n in hu)return Ml[e]=t[n];return e}var vu=no("animationend"),gu=no("animationiteration"),yu=no("animationstart"),wu=no("transitionend"),xu=new Map,ku="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function $t(e,t){xu.set(e,t),h(t,[e])}for(var Ll=0;Ll<ku.length;Ll++){var Ol=ku[Ll],df=Ol.toLowerCase(),ff=Ol[0].toUpperCase()+Ol.slice(1);$t(df,"on"+ff)}$t(vu,"onAnimationEnd"),$t(gu,"onAnimationIteration"),$t(yu,"onAnimationStart"),$t("dblclick","onDoubleClick"),$t("focusin","onFocus"),$t("focusout","onBlur"),$t(wu,"onTransitionEnd"),k("onMouseEnter",["mouseout","mouseover"]),k("onMouseLeave",["mouseout","mouseover"]),k("onPointerEnter",["pointerout","pointerover"]),k("onPointerLeave",["pointerout","pointerover"]),h("onChange","change click focusin focusout input keydown keyup selectionchange".split(" ")),h("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" ")),h("onBeforeInput",["compositionend","keypress","textInput","paste"]),h("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" ")),h("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" ")),h("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var cr="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),pf=new Set("cancel close invalid load scroll toggle".split(" ").concat(cr));function Su(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,cd(r,t,void 0,e),e.currentTarget=null}function Cu(e,t){t=(t&4)!==0;for(var n=0;n<e.length;n++){var r=e[n],o=r.event;r=r.listeners;e:{var l=void 0;if(t)for(var i=r.length-1;0<=i;i--){var c=r[i],d=c.instance,w=c.currentTarget;if(c=c.listener,d!==l&&o.isPropagationStopped())break e;Su(o,c,w),l=d}else for(i=0;i<r.length;i++){if(c=r[i],d=c.instance,w=c.currentTarget,c=c.listener,d!==l&&o.isPropagationStopped())break e;Su(o,c,w),l=d}}}if(Vr)throw e=dl,Vr=!1,dl=null,e}function he(e,t){var n=t[Wl];n===void 0&&(n=t[Wl]=new Set);var r=e+"__bubble";n.has(r)||(Eu(t,e,2,!1),n.add(r))}function Il(e,t,n){var r=0;t&&(r|=4),Eu(n,e,r,t)}var ro="_reactListening"+Math.random().toString(36).slice(2);function dr(e){if(!e[ro]){e[ro]=!0,f.forEach(function(n){n!=="selectionchange"&&(pf.has(n)||Il(n,!1,e),Il(n,!0,e))});var t=e.nodeType===9?e:e.ownerDocument;t===null||t[ro]||(t[ro]=!0,Il("selectionchange",!1,t))}}function Eu(e,t,n,r){switch(Ks(t)){case 1:var o=_d;break;case 4:o=Pd;break;default:o=yl}n=o.bind(null,t,n,e),o=void 0,!cl||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(o=!0),r?o!==void 0?e.addEventListener(t,n,{capture:!0,passive:o}):e.addEventListener(t,n,!0):o!==void 0?e.addEventListener(t,n,{passive:o}):e.addEventListener(t,n,!1)}function Al(e,t,n,r,o){var l=r;if(!(t&1)&&!(t&2)&&r!==null)e:for(;;){if(r===null)return;var i=r.tag;if(i===3||i===4){var c=r.stateNode.containerInfo;if(c===o||c.nodeType===8&&c.parentNode===o)break;if(i===4)for(i=r.return;i!==null;){var d=i.tag;if((d===3||d===4)&&(d=i.stateNode.containerInfo,d===o||d.nodeType===8&&d.parentNode===o))return;i=i.return}for(;c!==null;){if(i=on(c),i===null)return;if(d=i.tag,d===5||d===6){r=l=i;continue e}c=c.parentNode}}r=r.return}Rs(function(){var w=l,_=sl(n),P=[];e:{var E=xu.get(e);if(E!==void 0){var V=kl,B=e;switch(e){case"keypress":if(Jr(n)===0)break e;case"keydown":case"keyup":V=Wd;break;case"focusin":B="focus",V=El;break;case"focusout":B="blur",V=El;break;case"beforeblur":case"afterblur":V=El;break;case"click":if(n.button===2)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":V=Zs;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":V=zd;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":V=Qd;break;case vu:case gu:case yu:V=Ld;break;case wu:V=Kd;break;case"scroll":V=jd;break;case"wheel":V=Xd;break;case"copy":case"cut":case"paste":V=Id;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":V=bs}var H=(t&4)!==0,Ne=!H&&e==="scroll",g=H?E!==null?E+"Capture":null:E;H=[];for(var p=w,y;p!==null;){y=p;var T=y.stateNode;if(y.tag===5&&T!==null&&(y=T,g!==null&&(T=Kn(p,g),T!=null&&H.push(fr(p,T,y)))),Ne)break;p=p.return}0<H.length&&(E=new V(E,B,null,n,_),P.push({event:E,listeners:H}))}}if(!(t&7)){e:{if(E=e==="mouseover"||e==="pointerover",V=e==="mouseout"||e==="pointerout",E&&n!==il&&(B=n.relatedTarget||n.fromElement)&&(on(B)||B[Nt]))break e;if((V||E)&&(E=_.window===_?_:(E=_.ownerDocument)?E.defaultView||E.parentWindow:window,V?(B=n.relatedTarget||n.toElement,V=w,B=B?on(B):null,B!==null&&(Ne=rn(B),B!==Ne||B.tag!==5&&B.tag!==6)&&(B=null)):(V=null,B=w),V!==B)){if(H=Zs,T="onMouseLeave",g="onMouseEnter",p="mouse",(e==="pointerout"||e==="pointerover")&&(H=bs,T="onPointerLeave",g="onPointerEnter",p="pointer"),Ne=V==null?E:Pn(V),y=B==null?E:Pn(B),E=new H(T,p+"leave",V,n,_),E.target=Ne,E.relatedTarget=y,T=null,on(_)===w&&(H=new H(g,p+"enter",B,n,_),H.target=y,H.relatedTarget=Ne,T=H),Ne=T,V&&B)t:{for(H=V,g=B,p=0,y=H;y;y=Nn(y))p++;for(y=0,T=g;T;T=Nn(T))y++;for(;0<p-y;)H=Nn(H),p--;for(;0<y-p;)g=Nn(g),y--;for(;p--;){if(H===g||g!==null&&H===g.alternate)break t;H=Nn(H),g=Nn(g)}H=null}else H=null;V!==null&&Nu(P,E,V,H,!1),B!==null&&Ne!==null&&Nu(P,Ne,B,H,!0)}}e:{if(E=w?Pn(w):window,V=E.nodeName&&E.nodeName.toLowerCase(),V==="select"||V==="input"&&E.type==="file")var G=nf;else if(ou(E))if(iu)G=sf;else{G=of;var Y=rf}else(V=E.nodeName)&&V.toLowerCase()==="input"&&(E.type==="checkbox"||E.type==="radio")&&(G=lf);if(G&&(G=G(e,w))){lu(P,G,n,_);break e}Y&&Y(e,E,w),e==="focusout"&&(Y=E._wrapperState)&&Y.controlled&&E.type==="number"&&tl(E,"number",E.value)}switch(Y=w?Pn(w):window,e){case"focusin":(ou(Y)||Y.contentEditable==="true")&&(Cn=Y,zl=w,ar=null);break;case"focusout":ar=zl=Cn=null;break;case"mousedown":Tl=!0;break;case"contextmenu":case"mouseup":case"dragend":Tl=!1,mu(P,n,_);break;case"selectionchange":if(cf)break;case"keydown":case"keyup":mu(P,n,_)}var X;if(_l)e:{switch(e){case"compositionstart":var Z="onCompositionStart";break e;case"compositionend":Z="onCompositionEnd";break e;case"compositionupdate":Z="onCompositionUpdate";break e}Z=void 0}else Sn?nu(e,n)&&(Z="onCompositionEnd"):e==="keydown"&&n.keyCode===229&&(Z="onCompositionStart");Z&&(qs&&n.locale!=="ko"&&(Sn||Z!=="onCompositionStart"?Z==="onCompositionEnd"&&Sn&&(X=Ys()):(Vt=_,xl="value"in Vt?Vt.value:Vt.textContent,Sn=!0)),Y=oo(w,Z),0<Y.length&&(Z=new Js(Z,e,null,n,_),P.push({event:Z,listeners:Y}),X?Z.data=X:(X=ru(n),X!==null&&(Z.data=X)))),(X=Jd?bd(e,n):qd(e,n))&&(w=oo(w,"onBeforeInput"),0<w.length&&(_=new Js("onBeforeInput","beforeinput",null,n,_),P.push({event:_,listeners:w}),_.data=X))}Cu(P,t)})}function fr(e,t,n){return{instance:e,listener:t,currentTarget:n}}function oo(e,t){for(var n=t+"Capture",r=[];e!==null;){var o=e,l=o.stateNode;o.tag===5&&l!==null&&(o=l,l=Kn(e,n),l!=null&&r.unshift(fr(e,l,o)),l=Kn(e,t),l!=null&&r.push(fr(e,l,o))),e=e.return}return r}function Nn(e){if(e===null)return null;do e=e.return;while(e&&e.tag!==5);return e||null}function Nu(e,t,n,r,o){for(var l=t._reactName,i=[];n!==null&&n!==r;){var c=n,d=c.alternate,w=c.stateNode;if(d!==null&&d===r)break;c.tag===5&&w!==null&&(c=w,o?(d=Kn(n,l),d!=null&&i.unshift(fr(n,d,c))):o||(d=Kn(n,l),d!=null&&i.push(fr(n,d,c)))),n=n.return}i.length!==0&&e.push({event:t,listeners:i})}var mf=/\r\n?/g,hf=/\u0000|\uFFFD/g;function _u(e){return(typeof e=="string"?e:""+e).replace(mf,`
-`).replace(hf,"")}function lo(e,t,n){if(t=_u(t),_u(e)!==t&&n)throw Error(u(425))}function io(){}var Dl=null,Fl=null;function Ul(e,t){return e==="textarea"||e==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var Vl=typeof setTimeout=="function"?setTimeout:void 0,vf=typeof clearTimeout=="function"?clearTimeout:void 0,Pu=typeof Promise=="function"?Promise:void 0,gf=typeof queueMicrotask=="function"?queueMicrotask:typeof Pu<"u"?function(e){return Pu.resolve(null).then(e).catch(yf)}:Vl;function yf(e){setTimeout(function(){throw e})}function $l(e,t){var n=t,r=0;do{var o=n.nextSibling;if(e.removeChild(n),o&&o.nodeType===8)if(n=o.data,n==="/$"){if(r===0){e.removeChild(o),nr(t);return}r--}else n!=="$"&&n!=="$?"&&n!=="$!"||r++;n=o}while(n);nr(t)}function Wt(e){for(;e!=null;e=e.nextSibling){var t=e.nodeType;if(t===1||t===3)break;if(t===8){if(t=e.data,t==="$"||t==="$!"||t==="$?")break;if(t==="/$")return null}}return e}function ju(e){e=e.previousSibling;for(var t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="$"||n==="$!"||n==="$?"){if(t===0)return e;t--}else n==="/$"&&t++}e=e.previousSibling}return null}var _n=Math.random().toString(36).slice(2),kt="__reactFiber$"+_n,pr="__reactProps$"+_n,Nt="__reactContainer$"+_n,Wl="__reactEvents$"+_n,wf="__reactListeners$"+_n,xf="__reactHandles$"+_n;function on(e){var t=e[kt];if(t)return t;for(var n=e.parentNode;n;){if(t=n[Nt]||n[kt]){if(n=t.alternate,t.child!==null||n!==null&&n.child!==null)for(e=ju(e);e!==null;){if(n=e[kt])return n;e=ju(e)}return t}e=n,n=e.parentNode}return null}function mr(e){return e=e[kt]||e[Nt],!e||e.tag!==5&&e.tag!==6&&e.tag!==13&&e.tag!==3?null:e}function Pn(e){if(e.tag===5||e.tag===6)return e.stateNode;throw Error(u(33))}function so(e){return e[pr]||null}var Bl=[],jn=-1;function Bt(e){return{current:e}}function ve(e){0>jn||(e.current=Bl[jn],Bl[jn]=null,jn--)}function pe(e,t){jn++,Bl[jn]=e.current,e.current=t}var Ht={},$e=Bt(Ht),Ke=Bt(!1),ln=Ht;function Rn(e,t){var n=e.type.contextTypes;if(!n)return Ht;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var o={},l;for(l in n)o[l]=t[l];return r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=o),o}function Ye(e){return e=e.childContextTypes,e!=null}function uo(){ve(Ke),ve($e)}function Ru(e,t,n){if($e.current!==Ht)throw Error(u(168));pe($e,t),pe(Ke,n)}function zu(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,typeof r.getChildContext!="function")return n;r=r.getChildContext();for(var o in r)if(!(o in t))throw Error(u(108,fe(e)||"Unknown",o));return $({},n,r)}function ao(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||Ht,ln=$e.current,pe($e,e),pe(Ke,Ke.current),!0}function Tu(e,t,n){var r=e.stateNode;if(!r)throw Error(u(169));n?(e=zu(e,t,ln),r.__reactInternalMemoizedMergedChildContext=e,ve(Ke),ve($e),pe($e,e)):ve(Ke),pe(Ke,n)}var _t=null,co=!1,Hl=!1;function Mu(e){_t===null?_t=[e]:_t.push(e)}function kf(e){co=!0,Mu(e)}function Qt(){if(!Hl&&_t!==null){Hl=!0;var e=0,t=ce;try{var n=_t;for(ce=1;e<n.length;e++){var r=n[e];do r=r(!0);while(r!==null)}_t=null,co=!1}catch(o){throw _t!==null&&(_t=_t.slice(e+1)),Os(fl,Qt),o}finally{ce=t,Hl=!1}}return null}var zn=[],Tn=0,fo=null,po=0,lt=[],it=0,sn=null,Pt=1,jt="";function un(e,t){zn[Tn++]=po,zn[Tn++]=fo,fo=e,po=t}function Lu(e,t,n){lt[it++]=Pt,lt[it++]=jt,lt[it++]=sn,sn=e;var r=Pt;e=jt;var o=32-pt(r)-1;r&=~(1<<o),n+=1;var l=32-pt(t)+o;if(30<l){var i=o-o%5;l=(r&(1<<i)-1).toString(32),r>>=i,o-=i,Pt=1<<32-pt(t)+o|n<<o|r,jt=l+e}else Pt=1<<l|n<<o|r,jt=e}function Ql(e){e.return!==null&&(un(e,1),Lu(e,1,0))}function Gl(e){for(;e===fo;)fo=zn[--Tn],zn[Tn]=null,po=zn[--Tn],zn[Tn]=null;for(;e===sn;)sn=lt[--it],lt[it]=null,jt=lt[--it],lt[it]=null,Pt=lt[--it],lt[it]=null}var nt=null,rt=null,we=!1,ht=null;function Ou(e,t){var n=ct(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,t=e.deletions,t===null?(e.deletions=[n],e.flags|=16):t.push(n)}function Iu(e,t){switch(e.tag){case 5:var n=e.type;return t=t.nodeType!==1||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t,t!==null?(e.stateNode=t,nt=e,rt=Wt(t.firstChild),!0):!1;case 6:return t=e.pendingProps===""||t.nodeType!==3?null:t,t!==null?(e.stateNode=t,nt=e,rt=null,!0):!1;case 13:return t=t.nodeType!==8?null:t,t!==null?(n=sn!==null?{id:Pt,overflow:jt}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},n=ct(18,null,null,0),n.stateNode=t,n.return=e,e.child=n,nt=e,rt=null,!0):!1;default:return!1}}function Kl(e){return(e.mode&1)!==0&&(e.flags&128)===0}function Yl(e){if(we){var t=rt;if(t){var n=t;if(!Iu(e,t)){if(Kl(e))throw Error(u(418));t=Wt(n.nextSibling);var r=nt;t&&Iu(e,t)?Ou(r,n):(e.flags=e.flags&-4097|2,we=!1,nt=e)}}else{if(Kl(e))throw Error(u(418));e.flags=e.flags&-4097|2,we=!1,nt=e}}}function Au(e){for(e=e.return;e!==null&&e.tag!==5&&e.tag!==3&&e.tag!==13;)e=e.return;nt=e}function mo(e){if(e!==nt)return!1;if(!we)return Au(e),we=!0,!1;var t;if((t=e.tag!==3)&&!(t=e.tag!==5)&&(t=e.type,t=t!=="head"&&t!=="body"&&!Ul(e.type,e.memoizedProps)),t&&(t=rt)){if(Kl(e))throw Du(),Error(u(418));for(;t;)Ou(e,t),t=Wt(t.nextSibling)}if(Au(e),e.tag===13){if(e=e.memoizedState,e=e!==null?e.dehydrated:null,!e)throw Error(u(317));e:{for(e=e.nextSibling,t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="/$"){if(t===0){rt=Wt(e.nextSibling);break e}t--}else n!=="$"&&n!=="$!"&&n!=="$?"||t++}e=e.nextSibling}rt=null}}else rt=nt?Wt(e.stateNode.nextSibling):null;return!0}function Du(){for(var e=rt;e;)e=Wt(e.nextSibling)}function Mn(){rt=nt=null,we=!1}function Xl(e){ht===null?ht=[e]:ht.push(e)}var Sf=J.ReactCurrentBatchConfig;function hr(e,t,n){if(e=n.ref,e!==null&&typeof e!="function"&&typeof e!="object"){if(n._owner){if(n=n._owner,n){if(n.tag!==1)throw Error(u(309));var r=n.stateNode}if(!r)throw Error(u(147,e));var o=r,l=""+e;return t!==null&&t.ref!==null&&typeof t.ref=="function"&&t.ref._stringRef===l?t.ref:(t=function(i){var c=o.refs;i===null?delete c[l]:c[l]=i},t._stringRef=l,t)}if(typeof e!="string")throw Error(u(284));if(!n._owner)throw Error(u(290,e))}return e}function ho(e,t){throw e=Object.prototype.toString.call(t),Error(u(31,e==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function Fu(e){var t=e._init;return t(e._payload)}function Uu(e){function t(g,p){if(e){var y=g.deletions;y===null?(g.deletions=[p],g.flags|=16):y.push(p)}}function n(g,p){if(!e)return null;for(;p!==null;)t(g,p),p=p.sibling;return null}function r(g,p){for(g=new Map;p!==null;)p.key!==null?g.set(p.key,p):g.set(p.index,p),p=p.sibling;return g}function o(g,p){return g=qt(g,p),g.index=0,g.sibling=null,g}function l(g,p,y){return g.index=y,e?(y=g.alternate,y!==null?(y=y.index,y<p?(g.flags|=2,p):y):(g.flags|=2,p)):(g.flags|=1048576,p)}function i(g){return e&&g.alternate===null&&(g.flags|=2),g}function c(g,p,y,T){return p===null||p.tag!==6?(p=Vi(y,g.mode,T),p.return=g,p):(p=o(p,y),p.return=g,p)}function d(g,p,y,T){var G=y.type;return G===xe?_(g,p,y.props.children,T,y.key):p!==null&&(p.elementType===G||typeof G=="object"&&G!==null&&G.$$typeof===Re&&Fu(G)===p.type)?(T=o(p,y.props),T.ref=hr(g,p,y),T.return=g,T):(T=Uo(y.type,y.key,y.props,null,g.mode,T),T.ref=hr(g,p,y),T.return=g,T)}function w(g,p,y,T){return p===null||p.tag!==4||p.stateNode.containerInfo!==y.containerInfo||p.stateNode.implementation!==y.implementation?(p=$i(y,g.mode,T),p.return=g,p):(p=o(p,y.children||[]),p.return=g,p)}function _(g,p,y,T,G){return p===null||p.tag!==7?(p=vn(y,g.mode,T,G),p.return=g,p):(p=o(p,y),p.return=g,p)}function P(g,p,y){if(typeof p=="string"&&p!==""||typeof p=="number")return p=Vi(""+p,g.mode,y),p.return=g,p;if(typeof p=="object"&&p!==null){switch(p.$$typeof){case le:return y=Uo(p.type,p.key,p.props,null,g.mode,y),y.ref=hr(g,null,p),y.return=g,y;case de:return p=$i(p,g.mode,y),p.return=g,p;case Re:var T=p._init;return P(g,T(p._payload),y)}if(Hn(p)||K(p))return p=vn(p,g.mode,y,null),p.return=g,p;ho(g,p)}return null}function E(g,p,y,T){var G=p!==null?p.key:null;if(typeof y=="string"&&y!==""||typeof y=="number")return G!==null?null:c(g,p,""+y,T);if(typeof y=="object"&&y!==null){switch(y.$$typeof){case le:return y.key===G?d(g,p,y,T):null;case de:return y.key===G?w(g,p,y,T):null;case Re:return G=y._init,E(g,p,G(y._payload),T)}if(Hn(y)||K(y))return G!==null?null:_(g,p,y,T,null);ho(g,y)}return null}function V(g,p,y,T,G){if(typeof T=="string"&&T!==""||typeof T=="number")return g=g.get(y)||null,c(p,g,""+T,G);if(typeof T=="object"&&T!==null){switch(T.$$typeof){case le:return g=g.get(T.key===null?y:T.key)||null,d(p,g,T,G);case de:return g=g.get(T.key===null?y:T.key)||null,w(p,g,T,G);case Re:var Y=T._init;return V(g,p,y,Y(T._payload),G)}if(Hn(T)||K(T))return g=g.get(y)||null,_(p,g,T,G,null);ho(p,T)}return null}function B(g,p,y,T){for(var G=null,Y=null,X=p,Z=p=0,Oe=null;X!==null&&Z<y.length;Z++){X.index>Z?(Oe=X,X=null):Oe=X.sibling;var ue=E(g,X,y[Z],T);if(ue===null){X===null&&(X=Oe);break}e&&X&&ue.alternate===null&&t(g,X),p=l(ue,p,Z),Y===null?G=ue:Y.sibling=ue,Y=ue,X=Oe}if(Z===y.length)return n(g,X),we&&un(g,Z),G;if(X===null){for(;Z<y.length;Z++)X=P(g,y[Z],T),X!==null&&(p=l(X,p,Z),Y===null?G=X:Y.sibling=X,Y=X);return we&&un(g,Z),G}for(X=r(g,X);Z<y.length;Z++)Oe=V(X,g,Z,y[Z],T),Oe!==null&&(e&&Oe.alternate!==null&&X.delete(Oe.key===null?Z:Oe.key),p=l(Oe,p,Z),Y===null?G=Oe:Y.sibling=Oe,Y=Oe);return e&&X.forEach(function(en){return t(g,en)}),we&&un(g,Z),G}function H(g,p,y,T){var G=K(y);if(typeof G!="function")throw Error(u(150));if(y=G.call(y),y==null)throw Error(u(151));for(var Y=G=null,X=p,Z=p=0,Oe=null,ue=y.next();X!==null&&!ue.done;Z++,ue=y.next()){X.index>Z?(Oe=X,X=null):Oe=X.sibling;var en=E(g,X,ue.value,T);if(en===null){X===null&&(X=Oe);break}e&&X&&en.alternate===null&&t(g,X),p=l(en,p,Z),Y===null?G=en:Y.sibling=en,Y=en,X=Oe}if(ue.done)return n(g,X),we&&un(g,Z),G;if(X===null){for(;!ue.done;Z++,ue=y.next())ue=P(g,ue.value,T),ue!==null&&(p=l(ue,p,Z),Y===null?G=ue:Y.sibling=ue,Y=ue);return we&&un(g,Z),G}for(X=r(g,X);!ue.done;Z++,ue=y.next())ue=V(X,g,Z,ue.value,T),ue!==null&&(e&&ue.alternate!==null&&X.delete(ue.key===null?Z:ue.key),p=l(ue,p,Z),Y===null?G=ue:Y.sibling=ue,Y=ue);return e&&X.forEach(function(ep){return t(g,ep)}),we&&un(g,Z),G}function Ne(g,p,y,T){if(typeof y=="object"&&y!==null&&y.type===xe&&y.key===null&&(y=y.props.children),typeof y=="object"&&y!==null){switch(y.$$typeof){case le:e:{for(var G=y.key,Y=p;Y!==null;){if(Y.key===G){if(G=y.type,G===xe){if(Y.tag===7){n(g,Y.sibling),p=o(Y,y.props.children),p.return=g,g=p;break e}}else if(Y.elementType===G||typeof G=="object"&&G!==null&&G.$$typeof===Re&&Fu(G)===Y.type){n(g,Y.sibling),p=o(Y,y.props),p.ref=hr(g,Y,y),p.return=g,g=p;break e}n(g,Y);break}else t(g,Y);Y=Y.sibling}y.type===xe?(p=vn(y.props.children,g.mode,T,y.key),p.return=g,g=p):(T=Uo(y.type,y.key,y.props,null,g.mode,T),T.ref=hr(g,p,y),T.return=g,g=T)}return i(g);case de:e:{for(Y=y.key;p!==null;){if(p.key===Y)if(p.tag===4&&p.stateNode.containerInfo===y.containerInfo&&p.stateNode.implementation===y.implementation){n(g,p.sibling),p=o(p,y.children||[]),p.return=g,g=p;break e}else{n(g,p);break}else t(g,p);p=p.sibling}p=$i(y,g.mode,T),p.return=g,g=p}return i(g);case Re:return Y=y._init,Ne(g,p,Y(y._payload),T)}if(Hn(y))return B(g,p,y,T);if(K(y))return H(g,p,y,T);ho(g,y)}return typeof y=="string"&&y!==""||typeof y=="number"?(y=""+y,p!==null&&p.tag===6?(n(g,p.sibling),p=o(p,y),p.return=g,g=p):(n(g,p),p=Vi(y,g.mode,T),p.return=g,g=p),i(g)):n(g,p)}return Ne}var Ln=Uu(!0),Vu=Uu(!1),vo=Bt(null),go=null,On=null,Zl=null;function Jl(){Zl=On=go=null}function bl(e){var t=vo.current;ve(vo),e._currentValue=t}function ql(e,t,n){for(;e!==null;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,r!==null&&(r.childLanes|=t)):r!==null&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function In(e,t){go=e,Zl=On=null,e=e.dependencies,e!==null&&e.firstContext!==null&&(e.lanes&t&&(Xe=!0),e.firstContext=null)}function st(e){var t=e._currentValue;if(Zl!==e)if(e={context:e,memoizedValue:t,next:null},On===null){if(go===null)throw Error(u(308));On=e,go.dependencies={lanes:0,firstContext:e}}else On=On.next=e;return t}var an=null;function ei(e){an===null?an=[e]:an.push(e)}function $u(e,t,n,r){var o=t.interleaved;return o===null?(n.next=n,ei(t)):(n.next=o.next,o.next=n),t.interleaved=n,Rt(e,r)}function Rt(e,t){e.lanes|=t;var n=e.alternate;for(n!==null&&(n.lanes|=t),n=e,e=e.return;e!==null;)e.childLanes|=t,n=e.alternate,n!==null&&(n.childLanes|=t),n=e,e=e.return;return n.tag===3?n.stateNode:null}var Gt=!1;function ti(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function Wu(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function zt(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function Kt(e,t,n){var r=e.updateQueue;if(r===null)return null;if(r=r.shared,oe&2){var o=r.pending;return o===null?t.next=t:(t.next=o.next,o.next=t),r.pending=t,Rt(e,n)}return o=r.interleaved,o===null?(t.next=t,ei(r)):(t.next=o.next,o.next=t),r.interleaved=t,Rt(e,n)}function yo(e,t,n){if(t=t.updateQueue,t!==null&&(t=t.shared,(n&4194240)!==0)){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,hl(e,n)}}function Bu(e,t){var n=e.updateQueue,r=e.alternate;if(r!==null&&(r=r.updateQueue,n===r)){var o=null,l=null;if(n=n.firstBaseUpdate,n!==null){do{var i={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};l===null?o=l=i:l=l.next=i,n=n.next}while(n!==null);l===null?o=l=t:l=l.next=t}else o=l=t;n={baseState:r.baseState,firstBaseUpdate:o,lastBaseUpdate:l,shared:r.shared,effects:r.effects},e.updateQueue=n;return}e=n.lastBaseUpdate,e===null?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function wo(e,t,n,r){var o=e.updateQueue;Gt=!1;var l=o.firstBaseUpdate,i=o.lastBaseUpdate,c=o.shared.pending;if(c!==null){o.shared.pending=null;var d=c,w=d.next;d.next=null,i===null?l=w:i.next=w,i=d;var _=e.alternate;_!==null&&(_=_.updateQueue,c=_.lastBaseUpdate,c!==i&&(c===null?_.firstBaseUpdate=w:c.next=w,_.lastBaseUpdate=d))}if(l!==null){var P=o.baseState;i=0,_=w=d=null,c=l;do{var E=c.lane,V=c.eventTime;if((r&E)===E){_!==null&&(_=_.next={eventTime:V,lane:0,tag:c.tag,payload:c.payload,callback:c.callback,next:null});e:{var B=e,H=c;switch(E=t,V=n,H.tag){case 1:if(B=H.payload,typeof B=="function"){P=B.call(V,P,E);break e}P=B;break e;case 3:B.flags=B.flags&-65537|128;case 0:if(B=H.payload,E=typeof B=="function"?B.call(V,P,E):B,E==null)break e;P=$({},P,E);break e;case 2:Gt=!0}}c.callback!==null&&c.lane!==0&&(e.flags|=64,E=o.effects,E===null?o.effects=[c]:E.push(c))}else V={eventTime:V,lane:E,tag:c.tag,payload:c.payload,callback:c.callback,next:null},_===null?(w=_=V,d=P):_=_.next=V,i|=E;if(c=c.next,c===null){if(c=o.shared.pending,c===null)break;E=c,c=E.next,E.next=null,o.lastBaseUpdate=E,o.shared.pending=null}}while(!0);if(_===null&&(d=P),o.baseState=d,o.firstBaseUpdate=w,o.lastBaseUpdate=_,t=o.shared.interleaved,t!==null){o=t;do i|=o.lane,o=o.next;while(o!==t)}else l===null&&(o.shared.lanes=0);fn|=i,e.lanes=i,e.memoizedState=P}}function Hu(e,t,n){if(e=t.effects,t.effects=null,e!==null)for(t=0;t<e.length;t++){var r=e[t],o=r.callback;if(o!==null){if(r.callback=null,r=n,typeof o!="function")throw Error(u(191,o));o.call(r)}}}var vr={},St=Bt(vr),gr=Bt(vr),yr=Bt(vr);function cn(e){if(e===vr)throw Error(u(174));return e}function ni(e,t){switch(pe(yr,t),pe(gr,e),pe(St,vr),e=t.nodeType,e){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:rl(null,"");break;default:e=e===8?t.parentNode:t,t=e.namespaceURI||null,e=e.tagName,t=rl(t,e)}ve(St),pe(St,t)}function An(){ve(St),ve(gr),ve(yr)}function Qu(e){cn(yr.current);var t=cn(St.current),n=rl(t,e.type);t!==n&&(pe(gr,e),pe(St,n))}function ri(e){gr.current===e&&(ve(St),ve(gr))}var ke=Bt(0);function xo(e){for(var t=e;t!==null;){if(t.tag===13){var n=t.memoizedState;if(n!==null&&(n=n.dehydrated,n===null||n.data==="$?"||n.data==="$!"))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if(t.flags&128)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var oi=[];function li(){for(var e=0;e<oi.length;e++)oi[e]._workInProgressVersionPrimary=null;oi.length=0}var ko=J.ReactCurrentDispatcher,ii=J.ReactCurrentBatchConfig,dn=0,Se=null,ze=null,Me=null,So=!1,wr=!1,xr=0,Cf=0;function We(){throw Error(u(321))}function si(e,t){if(t===null)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!mt(e[n],t[n]))return!1;return!0}function ui(e,t,n,r,o,l){if(dn=l,Se=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,ko.current=e===null||e.memoizedState===null?Pf:jf,e=n(r,o),wr){l=0;do{if(wr=!1,xr=0,25<=l)throw Error(u(301));l+=1,Me=ze=null,t.updateQueue=null,ko.current=Rf,e=n(r,o)}while(wr)}if(ko.current=No,t=ze!==null&&ze.next!==null,dn=0,Me=ze=Se=null,So=!1,t)throw Error(u(300));return e}function ai(){var e=xr!==0;return xr=0,e}function Ct(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return Me===null?Se.memoizedState=Me=e:Me=Me.next=e,Me}function ut(){if(ze===null){var e=Se.alternate;e=e!==null?e.memoizedState:null}else e=ze.next;var t=Me===null?Se.memoizedState:Me.next;if(t!==null)Me=t,ze=e;else{if(e===null)throw Error(u(310));ze=e,e={memoizedState:ze.memoizedState,baseState:ze.baseState,baseQueue:ze.baseQueue,queue:ze.queue,next:null},Me===null?Se.memoizedState=Me=e:Me=Me.next=e}return Me}function kr(e,t){return typeof t=="function"?t(e):t}function ci(e){var t=ut(),n=t.queue;if(n===null)throw Error(u(311));n.lastRenderedReducer=e;var r=ze,o=r.baseQueue,l=n.pending;if(l!==null){if(o!==null){var i=o.next;o.next=l.next,l.next=i}r.baseQueue=o=l,n.pending=null}if(o!==null){l=o.next,r=r.baseState;var c=i=null,d=null,w=l;do{var _=w.lane;if((dn&_)===_)d!==null&&(d=d.next={lane:0,action:w.action,hasEagerState:w.hasEagerState,eagerState:w.eagerState,next:null}),r=w.hasEagerState?w.eagerState:e(r,w.action);else{var P={lane:_,action:w.action,hasEagerState:w.hasEagerState,eagerState:w.eagerState,next:null};d===null?(c=d=P,i=r):d=d.next=P,Se.lanes|=_,fn|=_}w=w.next}while(w!==null&&w!==l);d===null?i=r:d.next=c,mt(r,t.memoizedState)||(Xe=!0),t.memoizedState=r,t.baseState=i,t.baseQueue=d,n.lastRenderedState=r}if(e=n.interleaved,e!==null){o=e;do l=o.lane,Se.lanes|=l,fn|=l,o=o.next;while(o!==e)}else o===null&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function di(e){var t=ut(),n=t.queue;if(n===null)throw Error(u(311));n.lastRenderedReducer=e;var r=n.dispatch,o=n.pending,l=t.memoizedState;if(o!==null){n.pending=null;var i=o=o.next;do l=e(l,i.action),i=i.next;while(i!==o);mt(l,t.memoizedState)||(Xe=!0),t.memoizedState=l,t.baseQueue===null&&(t.baseState=l),n.lastRenderedState=l}return[l,r]}function Gu(){}function Ku(e,t){var n=Se,r=ut(),o=t(),l=!mt(r.memoizedState,o);if(l&&(r.memoizedState=o,Xe=!0),r=r.queue,fi(Zu.bind(null,n,r,e),[e]),r.getSnapshot!==t||l||Me!==null&&Me.memoizedState.tag&1){if(n.flags|=2048,Sr(9,Xu.bind(null,n,r,o,t),void 0,null),Le===null)throw Error(u(349));dn&30||Yu(n,t,o)}return o}function Yu(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},t=Se.updateQueue,t===null?(t={lastEffect:null,stores:null},Se.updateQueue=t,t.stores=[e]):(n=t.stores,n===null?t.stores=[e]:n.push(e))}function Xu(e,t,n,r){t.value=n,t.getSnapshot=r,Ju(t)&&bu(e)}function Zu(e,t,n){return n(function(){Ju(t)&&bu(e)})}function Ju(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!mt(e,n)}catch{return!0}}function bu(e){var t=Rt(e,1);t!==null&&wt(t,e,1,-1)}function qu(e){var t=Ct();return typeof e=="function"&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:kr,lastRenderedState:e},t.queue=e,e=e.dispatch=_f.bind(null,Se,e),[t.memoizedState,e]}function Sr(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},t=Se.updateQueue,t===null?(t={lastEffect:null,stores:null},Se.updateQueue=t,t.lastEffect=e.next=e):(n=t.lastEffect,n===null?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e)),e}function ea(){return ut().memoizedState}function Co(e,t,n,r){var o=Ct();Se.flags|=e,o.memoizedState=Sr(1|t,n,void 0,r===void 0?null:r)}function Eo(e,t,n,r){var o=ut();r=r===void 0?null:r;var l=void 0;if(ze!==null){var i=ze.memoizedState;if(l=i.destroy,r!==null&&si(r,i.deps)){o.memoizedState=Sr(t,n,l,r);return}}Se.flags|=e,o.memoizedState=Sr(1|t,n,l,r)}function ta(e,t){return Co(8390656,8,e,t)}function fi(e,t){return Eo(2048,8,e,t)}function na(e,t){return Eo(4,2,e,t)}function ra(e,t){return Eo(4,4,e,t)}function oa(e,t){if(typeof t=="function")return e=e(),t(e),function(){t(null)};if(t!=null)return e=e(),t.current=e,function(){t.current=null}}function la(e,t,n){return n=n!=null?n.concat([e]):null,Eo(4,4,oa.bind(null,t,e),n)}function pi(){}function ia(e,t){var n=ut();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&si(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function sa(e,t){var n=ut();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&si(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function ua(e,t,n){return dn&21?(mt(n,t)||(n=Fs(),Se.lanes|=n,fn|=n,e.baseState=!0),t):(e.baseState&&(e.baseState=!1,Xe=!0),e.memoizedState=n)}function Ef(e,t){var n=ce;ce=n!==0&&4>n?n:4,e(!0);var r=ii.transition;ii.transition={};try{e(!1),t()}finally{ce=n,ii.transition=r}}function aa(){return ut().memoizedState}function Nf(e,t,n){var r=Jt(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},ca(e))da(t,n);else if(n=$u(e,t,n,r),n!==null){var o=Ge();wt(n,e,r,o),fa(n,t,r)}}function _f(e,t,n){var r=Jt(e),o={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(ca(e))da(t,o);else{var l=e.alternate;if(e.lanes===0&&(l===null||l.lanes===0)&&(l=t.lastRenderedReducer,l!==null))try{var i=t.lastRenderedState,c=l(i,n);if(o.hasEagerState=!0,o.eagerState=c,mt(c,i)){var d=t.interleaved;d===null?(o.next=o,ei(t)):(o.next=d.next,d.next=o),t.interleaved=o;return}}catch{}finally{}n=$u(e,t,o,r),n!==null&&(o=Ge(),wt(n,e,r,o),fa(n,t,r))}}function ca(e){var t=e.alternate;return e===Se||t!==null&&t===Se}function da(e,t){wr=So=!0;var n=e.pending;n===null?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function fa(e,t,n){if(n&4194240){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,hl(e,n)}}var No={readContext:st,useCallback:We,useContext:We,useEffect:We,useImperativeHandle:We,useInsertionEffect:We,useLayoutEffect:We,useMemo:We,useReducer:We,useRef:We,useState:We,useDebugValue:We,useDeferredValue:We,useTransition:We,useMutableSource:We,useSyncExternalStore:We,useId:We,unstable_isNewReconciler:!1},Pf={readContext:st,useCallback:function(e,t){return Ct().memoizedState=[e,t===void 0?null:t],e},useContext:st,useEffect:ta,useImperativeHandle:function(e,t,n){return n=n!=null?n.concat([e]):null,Co(4194308,4,oa.bind(null,t,e),n)},useLayoutEffect:function(e,t){return Co(4194308,4,e,t)},useInsertionEffect:function(e,t){return Co(4,2,e,t)},useMemo:function(e,t){var n=Ct();return t=t===void 0?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=Ct();return t=n!==void 0?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=Nf.bind(null,Se,e),[r.memoizedState,e]},useRef:function(e){var t=Ct();return e={current:e},t.memoizedState=e},useState:qu,useDebugValue:pi,useDeferredValue:function(e){return Ct().memoizedState=e},useTransition:function(){var e=qu(!1),t=e[0];return e=Ef.bind(null,e[1]),Ct().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=Se,o=Ct();if(we){if(n===void 0)throw Error(u(407));n=n()}else{if(n=t(),Le===null)throw Error(u(349));dn&30||Yu(r,t,n)}o.memoizedState=n;var l={value:n,getSnapshot:t};return o.queue=l,ta(Zu.bind(null,r,l,e),[e]),r.flags|=2048,Sr(9,Xu.bind(null,r,l,n,t),void 0,null),n},useId:function(){var e=Ct(),t=Le.identifierPrefix;if(we){var n=jt,r=Pt;n=(r&~(1<<32-pt(r)-1)).toString(32)+n,t=":"+t+"R"+n,n=xr++,0<n&&(t+="H"+n.toString(32)),t+=":"}else n=Cf++,t=":"+t+"r"+n.toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},jf={readContext:st,useCallback:ia,useContext:st,useEffect:fi,useImperativeHandle:la,useInsertionEffect:na,useLayoutEffect:ra,useMemo:sa,useReducer:ci,useRef:ea,useState:function(){return ci(kr)},useDebugValue:pi,useDeferredValue:function(e){var t=ut();return ua(t,ze.memoizedState,e)},useTransition:function(){var e=ci(kr)[0],t=ut().memoizedState;return[e,t]},useMutableSource:Gu,useSyncExternalStore:Ku,useId:aa,unstable_isNewReconciler:!1},Rf={readContext:st,useCallback:ia,useContext:st,useEffect:fi,useImperativeHandle:la,useInsertionEffect:na,useLayoutEffect:ra,useMemo:sa,useReducer:di,useRef:ea,useState:function(){return di(kr)},useDebugValue:pi,useDeferredValue:function(e){var t=ut();return ze===null?t.memoizedState=e:ua(t,ze.memoizedState,e)},useTransition:function(){var e=di(kr)[0],t=ut().memoizedState;return[e,t]},useMutableSource:Gu,useSyncExternalStore:Ku,useId:aa,unstable_isNewReconciler:!1};function vt(e,t){if(e&&e.defaultProps){t=$({},t),e=e.defaultProps;for(var n in e)t[n]===void 0&&(t[n]=e[n]);return t}return t}function mi(e,t,n,r){t=e.memoizedState,n=n(r,t),n=n==null?t:$({},t,n),e.memoizedState=n,e.lanes===0&&(e.updateQueue.baseState=n)}var _o={isMounted:function(e){return(e=e._reactInternals)?rn(e)===e:!1},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=Ge(),o=Jt(e),l=zt(r,o);l.payload=t,n!=null&&(l.callback=n),t=Kt(e,l,o),t!==null&&(wt(t,e,o,r),yo(t,e,o))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=Ge(),o=Jt(e),l=zt(r,o);l.tag=1,l.payload=t,n!=null&&(l.callback=n),t=Kt(e,l,o),t!==null&&(wt(t,e,o,r),yo(t,e,o))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=Ge(),r=Jt(e),o=zt(n,r);o.tag=2,t!=null&&(o.callback=t),t=Kt(e,o,r),t!==null&&(wt(t,e,r,n),yo(t,e,r))}};function pa(e,t,n,r,o,l,i){return e=e.stateNode,typeof e.shouldComponentUpdate=="function"?e.shouldComponentUpdate(r,l,i):t.prototype&&t.prototype.isPureReactComponent?!ur(n,r)||!ur(o,l):!0}function ma(e,t,n){var r=!1,o=Ht,l=t.contextType;return typeof l=="object"&&l!==null?l=st(l):(o=Ye(t)?ln:$e.current,r=t.contextTypes,l=(r=r!=null)?Rn(e,o):Ht),t=new t(n,l),e.memoizedState=t.state!==null&&t.state!==void 0?t.state:null,t.updater=_o,e.stateNode=t,t._reactInternals=e,r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=o,e.__reactInternalMemoizedMaskedChildContext=l),t}function ha(e,t,n,r){e=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(n,r),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&_o.enqueueReplaceState(t,t.state,null)}function hi(e,t,n,r){var o=e.stateNode;o.props=n,o.state=e.memoizedState,o.refs={},ti(e);var l=t.contextType;typeof l=="object"&&l!==null?o.context=st(l):(l=Ye(t)?ln:$e.current,o.context=Rn(e,l)),o.state=e.memoizedState,l=t.getDerivedStateFromProps,typeof l=="function"&&(mi(e,t,l,n),o.state=e.memoizedState),typeof t.getDerivedStateFromProps=="function"||typeof o.getSnapshotBeforeUpdate=="function"||typeof o.UNSAFE_componentWillMount!="function"&&typeof o.componentWillMount!="function"||(t=o.state,typeof o.componentWillMount=="function"&&o.componentWillMount(),typeof o.UNSAFE_componentWillMount=="function"&&o.UNSAFE_componentWillMount(),t!==o.state&&_o.enqueueReplaceState(o,o.state,null),wo(e,n,o,r),o.state=e.memoizedState),typeof o.componentDidMount=="function"&&(e.flags|=4194308)}function Dn(e,t){try{var n="",r=t;do n+=ie(r),r=r.return;while(r);var o=n}catch(l){o=`
-Error generating stack: `+l.message+`
-`+l.stack}return{value:e,source:t,stack:o,digest:null}}function vi(e,t,n){return{value:e,source:null,stack:n??null,digest:t??null}}function gi(e,t){try{console.error(t.value)}catch(n){setTimeout(function(){throw n})}}var zf=typeof WeakMap=="function"?WeakMap:Map;function va(e,t,n){n=zt(-1,n),n.tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){Lo||(Lo=!0,Mi=r),gi(e,t)},n}function ga(e,t,n){n=zt(-1,n),n.tag=3;var r=e.type.getDerivedStateFromError;if(typeof r=="function"){var o=t.value;n.payload=function(){return r(o)},n.callback=function(){gi(e,t)}}var l=e.stateNode;return l!==null&&typeof l.componentDidCatch=="function"&&(n.callback=function(){gi(e,t),typeof r!="function"&&(Xt===null?Xt=new Set([this]):Xt.add(this));var i=t.stack;this.componentDidCatch(t.value,{componentStack:i!==null?i:""})}),n}function ya(e,t,n){var r=e.pingCache;if(r===null){r=e.pingCache=new zf;var o=new Set;r.set(t,o)}else o=r.get(t),o===void 0&&(o=new Set,r.set(t,o));o.has(n)||(o.add(n),e=Hf.bind(null,e,t,n),t.then(e,e))}function wa(e){do{var t;if((t=e.tag===13)&&(t=e.memoizedState,t=t!==null?t.dehydrated!==null:!0),t)return e;e=e.return}while(e!==null);return null}function xa(e,t,n,r,o){return e.mode&1?(e.flags|=65536,e.lanes=o,e):(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,n.tag===1&&(n.alternate===null?n.tag=17:(t=zt(-1,1),t.tag=2,Kt(n,t,1))),n.lanes|=1),e)}var Tf=J.ReactCurrentOwner,Xe=!1;function Qe(e,t,n,r){t.child=e===null?Vu(t,null,n,r):Ln(t,e.child,n,r)}function ka(e,t,n,r,o){n=n.render;var l=t.ref;return In(t,o),r=ui(e,t,n,r,l,o),n=ai(),e!==null&&!Xe?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~o,Tt(e,t,o)):(we&&n&&Ql(t),t.flags|=1,Qe(e,t,r,o),t.child)}function Sa(e,t,n,r,o){if(e===null){var l=n.type;return typeof l=="function"&&!Ui(l)&&l.defaultProps===void 0&&n.compare===null&&n.defaultProps===void 0?(t.tag=15,t.type=l,Ca(e,t,l,r,o)):(e=Uo(n.type,null,r,t,t.mode,o),e.ref=t.ref,e.return=t,t.child=e)}if(l=e.child,!(e.lanes&o)){var i=l.memoizedProps;if(n=n.compare,n=n!==null?n:ur,n(i,r)&&e.ref===t.ref)return Tt(e,t,o)}return t.flags|=1,e=qt(l,r),e.ref=t.ref,e.return=t,t.child=e}function Ca(e,t,n,r,o){if(e!==null){var l=e.memoizedProps;if(ur(l,r)&&e.ref===t.ref)if(Xe=!1,t.pendingProps=r=l,(e.lanes&o)!==0)e.flags&131072&&(Xe=!0);else return t.lanes=e.lanes,Tt(e,t,o)}return yi(e,t,n,r,o)}function Ea(e,t,n){var r=t.pendingProps,o=r.children,l=e!==null?e.memoizedState:null;if(r.mode==="hidden")if(!(t.mode&1))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},pe(Un,ot),ot|=n;else{if(!(n&1073741824))return e=l!==null?l.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,pe(Un,ot),ot|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=l!==null?l.baseLanes:n,pe(Un,ot),ot|=r}else l!==null?(r=l.baseLanes|n,t.memoizedState=null):r=n,pe(Un,ot),ot|=r;return Qe(e,t,o,n),t.child}function Na(e,t){var n=t.ref;(e===null&&n!==null||e!==null&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function yi(e,t,n,r,o){var l=Ye(n)?ln:$e.current;return l=Rn(t,l),In(t,o),n=ui(e,t,n,r,l,o),r=ai(),e!==null&&!Xe?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~o,Tt(e,t,o)):(we&&r&&Ql(t),t.flags|=1,Qe(e,t,n,o),t.child)}function _a(e,t,n,r,o){if(Ye(n)){var l=!0;ao(t)}else l=!1;if(In(t,o),t.stateNode===null)jo(e,t),ma(t,n,r),hi(t,n,r,o),r=!0;else if(e===null){var i=t.stateNode,c=t.memoizedProps;i.props=c;var d=i.context,w=n.contextType;typeof w=="object"&&w!==null?w=st(w):(w=Ye(n)?ln:$e.current,w=Rn(t,w));var _=n.getDerivedStateFromProps,P=typeof _=="function"||typeof i.getSnapshotBeforeUpdate=="function";P||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(c!==r||d!==w)&&ha(t,i,r,w),Gt=!1;var E=t.memoizedState;i.state=E,wo(t,r,i,o),d=t.memoizedState,c!==r||E!==d||Ke.current||Gt?(typeof _=="function"&&(mi(t,n,_,r),d=t.memoizedState),(c=Gt||pa(t,n,c,r,E,d,w))?(P||typeof i.UNSAFE_componentWillMount!="function"&&typeof i.componentWillMount!="function"||(typeof i.componentWillMount=="function"&&i.componentWillMount(),typeof i.UNSAFE_componentWillMount=="function"&&i.UNSAFE_componentWillMount()),typeof i.componentDidMount=="function"&&(t.flags|=4194308)):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=d),i.props=r,i.state=d,i.context=w,r=c):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),r=!1)}else{i=t.stateNode,Wu(e,t),c=t.memoizedProps,w=t.type===t.elementType?c:vt(t.type,c),i.props=w,P=t.pendingProps,E=i.context,d=n.contextType,typeof d=="object"&&d!==null?d=st(d):(d=Ye(n)?ln:$e.current,d=Rn(t,d));var V=n.getDerivedStateFromProps;(_=typeof V=="function"||typeof i.getSnapshotBeforeUpdate=="function")||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(c!==P||E!==d)&&ha(t,i,r,d),Gt=!1,E=t.memoizedState,i.state=E,wo(t,r,i,o);var B=t.memoizedState;c!==P||E!==B||Ke.current||Gt?(typeof V=="function"&&(mi(t,n,V,r),B=t.memoizedState),(w=Gt||pa(t,n,w,r,E,B,d)||!1)?(_||typeof i.UNSAFE_componentWillUpdate!="function"&&typeof i.componentWillUpdate!="function"||(typeof i.componentWillUpdate=="function"&&i.componentWillUpdate(r,B,d),typeof i.UNSAFE_componentWillUpdate=="function"&&i.UNSAFE_componentWillUpdate(r,B,d)),typeof i.componentDidUpdate=="function"&&(t.flags|=4),typeof i.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof i.componentDidUpdate!="function"||c===e.memoizedProps&&E===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||c===e.memoizedProps&&E===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=B),i.props=r,i.state=B,i.context=d,r=w):(typeof i.componentDidUpdate!="function"||c===e.memoizedProps&&E===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||c===e.memoizedProps&&E===e.memoizedState||(t.flags|=1024),r=!1)}return wi(e,t,n,r,l,o)}function wi(e,t,n,r,o,l){Na(e,t);var i=(t.flags&128)!==0;if(!r&&!i)return o&&Tu(t,n,!1),Tt(e,t,l);r=t.stateNode,Tf.current=t;var c=i&&typeof n.getDerivedStateFromError!="function"?null:r.render();return t.flags|=1,e!==null&&i?(t.child=Ln(t,e.child,null,l),t.child=Ln(t,null,c,l)):Qe(e,t,c,l),t.memoizedState=r.state,o&&Tu(t,n,!0),t.child}function Pa(e){var t=e.stateNode;t.pendingContext?Ru(e,t.pendingContext,t.pendingContext!==t.context):t.context&&Ru(e,t.context,!1),ni(e,t.containerInfo)}function ja(e,t,n,r,o){return Mn(),Xl(o),t.flags|=256,Qe(e,t,n,r),t.child}var xi={dehydrated:null,treeContext:null,retryLane:0};function ki(e){return{baseLanes:e,cachePool:null,transitions:null}}function Ra(e,t,n){var r=t.pendingProps,o=ke.current,l=!1,i=(t.flags&128)!==0,c;if((c=i)||(c=e!==null&&e.memoizedState===null?!1:(o&2)!==0),c?(l=!0,t.flags&=-129):(e===null||e.memoizedState!==null)&&(o|=1),pe(ke,o&1),e===null)return Yl(t),e=t.memoizedState,e!==null&&(e=e.dehydrated,e!==null)?(t.mode&1?e.data==="$!"?t.lanes=8:t.lanes=1073741824:t.lanes=1,null):(i=r.children,e=r.fallback,l?(r=t.mode,l=t.child,i={mode:"hidden",children:i},!(r&1)&&l!==null?(l.childLanes=0,l.pendingProps=i):l=Vo(i,r,0,null),e=vn(e,r,n,null),l.return=t,e.return=t,l.sibling=e,t.child=l,t.child.memoizedState=ki(n),t.memoizedState=xi,e):Si(t,i));if(o=e.memoizedState,o!==null&&(c=o.dehydrated,c!==null))return Mf(e,t,i,r,c,o,n);if(l){l=r.fallback,i=t.mode,o=e.child,c=o.sibling;var d={mode:"hidden",children:r.children};return!(i&1)&&t.child!==o?(r=t.child,r.childLanes=0,r.pendingProps=d,t.deletions=null):(r=qt(o,d),r.subtreeFlags=o.subtreeFlags&14680064),c!==null?l=qt(c,l):(l=vn(l,i,n,null),l.flags|=2),l.return=t,r.return=t,r.sibling=l,t.child=r,r=l,l=t.child,i=e.child.memoizedState,i=i===null?ki(n):{baseLanes:i.baseLanes|n,cachePool:null,transitions:i.transitions},l.memoizedState=i,l.childLanes=e.childLanes&~n,t.memoizedState=xi,r}return l=e.child,e=l.sibling,r=qt(l,{mode:"visible",children:r.children}),!(t.mode&1)&&(r.lanes=n),r.return=t,r.sibling=null,e!==null&&(n=t.deletions,n===null?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=r,t.memoizedState=null,r}function Si(e,t){return t=Vo({mode:"visible",children:t},e.mode,0,null),t.return=e,e.child=t}function Po(e,t,n,r){return r!==null&&Xl(r),Ln(t,e.child,null,n),e=Si(t,t.pendingProps.children),e.flags|=2,t.memoizedState=null,e}function Mf(e,t,n,r,o,l,i){if(n)return t.flags&256?(t.flags&=-257,r=vi(Error(u(422))),Po(e,t,i,r)):t.memoizedState!==null?(t.child=e.child,t.flags|=128,null):(l=r.fallback,o=t.mode,r=Vo({mode:"visible",children:r.children},o,0,null),l=vn(l,o,i,null),l.flags|=2,r.return=t,l.return=t,r.sibling=l,t.child=r,t.mode&1&&Ln(t,e.child,null,i),t.child.memoizedState=ki(i),t.memoizedState=xi,l);if(!(t.mode&1))return Po(e,t,i,null);if(o.data==="$!"){if(r=o.nextSibling&&o.nextSibling.dataset,r)var c=r.dgst;return r=c,l=Error(u(419)),r=vi(l,r,void 0),Po(e,t,i,r)}if(c=(i&e.childLanes)!==0,Xe||c){if(r=Le,r!==null){switch(i&-i){case 4:o=2;break;case 16:o=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:o=32;break;case 536870912:o=268435456;break;default:o=0}o=o&(r.suspendedLanes|i)?0:o,o!==0&&o!==l.retryLane&&(l.retryLane=o,Rt(e,o),wt(r,e,o,-1))}return Fi(),r=vi(Error(u(421))),Po(e,t,i,r)}return o.data==="$?"?(t.flags|=128,t.child=e.child,t=Qf.bind(null,e),o._reactRetry=t,null):(e=l.treeContext,rt=Wt(o.nextSibling),nt=t,we=!0,ht=null,e!==null&&(lt[it++]=Pt,lt[it++]=jt,lt[it++]=sn,Pt=e.id,jt=e.overflow,sn=t),t=Si(t,r.children),t.flags|=4096,t)}function za(e,t,n){e.lanes|=t;var r=e.alternate;r!==null&&(r.lanes|=t),ql(e.return,t,n)}function Ci(e,t,n,r,o){var l=e.memoizedState;l===null?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:o}:(l.isBackwards=t,l.rendering=null,l.renderingStartTime=0,l.last=r,l.tail=n,l.tailMode=o)}function Ta(e,t,n){var r=t.pendingProps,o=r.revealOrder,l=r.tail;if(Qe(e,t,r.children,n),r=ke.current,r&2)r=r&1|2,t.flags|=128;else{if(e!==null&&e.flags&128)e:for(e=t.child;e!==null;){if(e.tag===13)e.memoizedState!==null&&za(e,n,t);else if(e.tag===19)za(e,n,t);else if(e.child!==null){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;e.sibling===null;){if(e.return===null||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if(pe(ke,r),!(t.mode&1))t.memoizedState=null;else switch(o){case"forwards":for(n=t.child,o=null;n!==null;)e=n.alternate,e!==null&&xo(e)===null&&(o=n),n=n.sibling;n=o,n===null?(o=t.child,t.child=null):(o=n.sibling,n.sibling=null),Ci(t,!1,o,n,l);break;case"backwards":for(n=null,o=t.child,t.child=null;o!==null;){if(e=o.alternate,e!==null&&xo(e)===null){t.child=o;break}e=o.sibling,o.sibling=n,n=o,o=e}Ci(t,!0,n,null,l);break;case"together":Ci(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function jo(e,t){!(t.mode&1)&&e!==null&&(e.alternate=null,t.alternate=null,t.flags|=2)}function Tt(e,t,n){if(e!==null&&(t.dependencies=e.dependencies),fn|=t.lanes,!(n&t.childLanes))return null;if(e!==null&&t.child!==e.child)throw Error(u(153));if(t.child!==null){for(e=t.child,n=qt(e,e.pendingProps),t.child=n,n.return=t;e.sibling!==null;)e=e.sibling,n=n.sibling=qt(e,e.pendingProps),n.return=t;n.sibling=null}return t.child}function Lf(e,t,n){switch(t.tag){case 3:Pa(t),Mn();break;case 5:Qu(t);break;case 1:Ye(t.type)&&ao(t);break;case 4:ni(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,o=t.memoizedProps.value;pe(vo,r._currentValue),r._currentValue=o;break;case 13:if(r=t.memoizedState,r!==null)return r.dehydrated!==null?(pe(ke,ke.current&1),t.flags|=128,null):n&t.child.childLanes?Ra(e,t,n):(pe(ke,ke.current&1),e=Tt(e,t,n),e!==null?e.sibling:null);pe(ke,ke.current&1);break;case 19:if(r=(n&t.childLanes)!==0,e.flags&128){if(r)return Ta(e,t,n);t.flags|=128}if(o=t.memoizedState,o!==null&&(o.rendering=null,o.tail=null,o.lastEffect=null),pe(ke,ke.current),r)break;return null;case 22:case 23:return t.lanes=0,Ea(e,t,n)}return Tt(e,t,n)}var Ma,Ei,La,Oa;Ma=function(e,t){for(var n=t.child;n!==null;){if(n.tag===5||n.tag===6)e.appendChild(n.stateNode);else if(n.tag!==4&&n.child!==null){n.child.return=n,n=n.child;continue}if(n===t)break;for(;n.sibling===null;){if(n.return===null||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}},Ei=function(){},La=function(e,t,n,r){var o=e.memoizedProps;if(o!==r){e=t.stateNode,cn(St.current);var l=null;switch(n){case"input":o=qo(e,o),r=qo(e,r),l=[];break;case"select":o=$({},o,{value:void 0}),r=$({},r,{value:void 0}),l=[];break;case"textarea":o=nl(e,o),r=nl(e,r),l=[];break;default:typeof o.onClick!="function"&&typeof r.onClick=="function"&&(e.onclick=io)}ol(n,r);var i;n=null;for(w in o)if(!r.hasOwnProperty(w)&&o.hasOwnProperty(w)&&o[w]!=null)if(w==="style"){var c=o[w];for(i in c)c.hasOwnProperty(i)&&(n||(n={}),n[i]="")}else w!=="dangerouslySetInnerHTML"&&w!=="children"&&w!=="suppressContentEditableWarning"&&w!=="suppressHydrationWarning"&&w!=="autoFocus"&&(m.hasOwnProperty(w)?l||(l=[]):(l=l||[]).push(w,null));for(w in r){var d=r[w];if(c=o!=null?o[w]:void 0,r.hasOwnProperty(w)&&d!==c&&(d!=null||c!=null))if(w==="style")if(c){for(i in c)!c.hasOwnProperty(i)||d&&d.hasOwnProperty(i)||(n||(n={}),n[i]="");for(i in d)d.hasOwnProperty(i)&&c[i]!==d[i]&&(n||(n={}),n[i]=d[i])}else n||(l||(l=[]),l.push(w,n)),n=d;else w==="dangerouslySetInnerHTML"?(d=d?d.__html:void 0,c=c?c.__html:void 0,d!=null&&c!==d&&(l=l||[]).push(w,d)):w==="children"?typeof d!="string"&&typeof d!="number"||(l=l||[]).push(w,""+d):w!=="suppressContentEditableWarning"&&w!=="suppressHydrationWarning"&&(m.hasOwnProperty(w)?(d!=null&&w==="onScroll"&&he("scroll",e),l||c===d||(l=[])):(l=l||[]).push(w,d))}n&&(l=l||[]).push("style",n);var w=l;(t.updateQueue=w)&&(t.flags|=4)}},Oa=function(e,t,n,r){n!==r&&(t.flags|=4)};function Cr(e,t){if(!we)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;t!==null;)t.alternate!==null&&(n=t),t=t.sibling;n===null?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;n!==null;)n.alternate!==null&&(r=n),n=n.sibling;r===null?t||e.tail===null?e.tail=null:e.tail.sibling=null:r.sibling=null}}function Be(e){var t=e.alternate!==null&&e.alternate.child===e.child,n=0,r=0;if(t)for(var o=e.child;o!==null;)n|=o.lanes|o.childLanes,r|=o.subtreeFlags&14680064,r|=o.flags&14680064,o.return=e,o=o.sibling;else for(o=e.child;o!==null;)n|=o.lanes|o.childLanes,r|=o.subtreeFlags,r|=o.flags,o.return=e,o=o.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function Of(e,t,n){var r=t.pendingProps;switch(Gl(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return Be(t),null;case 1:return Ye(t.type)&&uo(),Be(t),null;case 3:return r=t.stateNode,An(),ve(Ke),ve($e),li(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),(e===null||e.child===null)&&(mo(t)?t.flags|=4:e===null||e.memoizedState.isDehydrated&&!(t.flags&256)||(t.flags|=1024,ht!==null&&(Ii(ht),ht=null))),Ei(e,t),Be(t),null;case 5:ri(t);var o=cn(yr.current);if(n=t.type,e!==null&&t.stateNode!=null)La(e,t,n,r,o),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(t.stateNode===null)throw Error(u(166));return Be(t),null}if(e=cn(St.current),mo(t)){r=t.stateNode,n=t.type;var l=t.memoizedProps;switch(r[kt]=t,r[pr]=l,e=(t.mode&1)!==0,n){case"dialog":he("cancel",r),he("close",r);break;case"iframe":case"object":case"embed":he("load",r);break;case"video":case"audio":for(o=0;o<cr.length;o++)he(cr[o],r);break;case"source":he("error",r);break;case"img":case"image":case"link":he("error",r),he("load",r);break;case"details":he("toggle",r);break;case"input":ms(r,l),he("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!l.multiple},he("invalid",r);break;case"textarea":gs(r,l),he("invalid",r)}ol(n,l),o=null;for(var i in l)if(l.hasOwnProperty(i)){var c=l[i];i==="children"?typeof c=="string"?r.textContent!==c&&(l.suppressHydrationWarning!==!0&&lo(r.textContent,c,e),o=["children",c]):typeof c=="number"&&r.textContent!==""+c&&(l.suppressHydrationWarning!==!0&&lo(r.textContent,c,e),o=["children",""+c]):m.hasOwnProperty(i)&&c!=null&&i==="onScroll"&&he("scroll",r)}switch(n){case"input":Ar(r),vs(r,l,!0);break;case"textarea":Ar(r),ws(r);break;case"select":case"option":break;default:typeof l.onClick=="function"&&(r.onclick=io)}r=o,t.updateQueue=r,r!==null&&(t.flags|=4)}else{i=o.nodeType===9?o:o.ownerDocument,e==="http://www.w3.org/1999/xhtml"&&(e=xs(n)),e==="http://www.w3.org/1999/xhtml"?n==="script"?(e=i.createElement("div"),e.innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):typeof r.is=="string"?e=i.createElement(n,{is:r.is}):(e=i.createElement(n),n==="select"&&(i=e,r.multiple?i.multiple=!0:r.size&&(i.size=r.size))):e=i.createElementNS(e,n),e[kt]=t,e[pr]=r,Ma(e,t,!1,!1),t.stateNode=e;e:{switch(i=ll(n,r),n){case"dialog":he("cancel",e),he("close",e),o=r;break;case"iframe":case"object":case"embed":he("load",e),o=r;break;case"video":case"audio":for(o=0;o<cr.length;o++)he(cr[o],e);o=r;break;case"source":he("error",e),o=r;break;case"img":case"image":case"link":he("error",e),he("load",e),o=r;break;case"details":he("toggle",e),o=r;break;case"input":ms(e,r),o=qo(e,r),he("invalid",e);break;case"option":o=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},o=$({},r,{value:void 0}),he("invalid",e);break;case"textarea":gs(e,r),o=nl(e,r),he("invalid",e);break;default:o=r}ol(n,o),c=o;for(l in c)if(c.hasOwnProperty(l)){var d=c[l];l==="style"?Cs(e,d):l==="dangerouslySetInnerHTML"?(d=d?d.__html:void 0,d!=null&&ks(e,d)):l==="children"?typeof d=="string"?(n!=="textarea"||d!=="")&&Qn(e,d):typeof d=="number"&&Qn(e,""+d):l!=="suppressContentEditableWarning"&&l!=="suppressHydrationWarning"&&l!=="autoFocus"&&(m.hasOwnProperty(l)?d!=null&&l==="onScroll"&&he("scroll",e):d!=null&&b(e,l,d,i))}switch(n){case"input":Ar(e),vs(e,r,!1);break;case"textarea":Ar(e),ws(e);break;case"option":r.value!=null&&e.setAttribute("value",""+ae(r.value));break;case"select":e.multiple=!!r.multiple,l=r.value,l!=null?gn(e,!!r.multiple,l,!1):r.defaultValue!=null&&gn(e,!!r.multiple,r.defaultValue,!0);break;default:typeof o.onClick=="function"&&(e.onclick=io)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}t.ref!==null&&(t.flags|=512,t.flags|=2097152)}return Be(t),null;case 6:if(e&&t.stateNode!=null)Oa(e,t,e.memoizedProps,r);else{if(typeof r!="string"&&t.stateNode===null)throw Error(u(166));if(n=cn(yr.current),cn(St.current),mo(t)){if(r=t.stateNode,n=t.memoizedProps,r[kt]=t,(l=r.nodeValue!==n)&&(e=nt,e!==null))switch(e.tag){case 3:lo(r.nodeValue,n,(e.mode&1)!==0);break;case 5:e.memoizedProps.suppressHydrationWarning!==!0&&lo(r.nodeValue,n,(e.mode&1)!==0)}l&&(t.flags|=4)}else r=(n.nodeType===9?n:n.ownerDocument).createTextNode(r),r[kt]=t,t.stateNode=r}return Be(t),null;case 13:if(ve(ke),r=t.memoizedState,e===null||e.memoizedState!==null&&e.memoizedState.dehydrated!==null){if(we&&rt!==null&&t.mode&1&&!(t.flags&128))Du(),Mn(),t.flags|=98560,l=!1;else if(l=mo(t),r!==null&&r.dehydrated!==null){if(e===null){if(!l)throw Error(u(318));if(l=t.memoizedState,l=l!==null?l.dehydrated:null,!l)throw Error(u(317));l[kt]=t}else Mn(),!(t.flags&128)&&(t.memoizedState=null),t.flags|=4;Be(t),l=!1}else ht!==null&&(Ii(ht),ht=null),l=!0;if(!l)return t.flags&65536?t:null}return t.flags&128?(t.lanes=n,t):(r=r!==null,r!==(e!==null&&e.memoizedState!==null)&&r&&(t.child.flags|=8192,t.mode&1&&(e===null||ke.current&1?Te===0&&(Te=3):Fi())),t.updateQueue!==null&&(t.flags|=4),Be(t),null);case 4:return An(),Ei(e,t),e===null&&dr(t.stateNode.containerInfo),Be(t),null;case 10:return bl(t.type._context),Be(t),null;case 17:return Ye(t.type)&&uo(),Be(t),null;case 19:if(ve(ke),l=t.memoizedState,l===null)return Be(t),null;if(r=(t.flags&128)!==0,i=l.rendering,i===null)if(r)Cr(l,!1);else{if(Te!==0||e!==null&&e.flags&128)for(e=t.child;e!==null;){if(i=xo(e),i!==null){for(t.flags|=128,Cr(l,!1),r=i.updateQueue,r!==null&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;n!==null;)l=n,e=r,l.flags&=14680066,i=l.alternate,i===null?(l.childLanes=0,l.lanes=e,l.child=null,l.subtreeFlags=0,l.memoizedProps=null,l.memoizedState=null,l.updateQueue=null,l.dependencies=null,l.stateNode=null):(l.childLanes=i.childLanes,l.lanes=i.lanes,l.child=i.child,l.subtreeFlags=0,l.deletions=null,l.memoizedProps=i.memoizedProps,l.memoizedState=i.memoizedState,l.updateQueue=i.updateQueue,l.type=i.type,e=i.dependencies,l.dependencies=e===null?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return pe(ke,ke.current&1|2),t.child}e=e.sibling}l.tail!==null&&Ee()>Vn&&(t.flags|=128,r=!0,Cr(l,!1),t.lanes=4194304)}else{if(!r)if(e=xo(i),e!==null){if(t.flags|=128,r=!0,n=e.updateQueue,n!==null&&(t.updateQueue=n,t.flags|=4),Cr(l,!0),l.tail===null&&l.tailMode==="hidden"&&!i.alternate&&!we)return Be(t),null}else 2*Ee()-l.renderingStartTime>Vn&&n!==1073741824&&(t.flags|=128,r=!0,Cr(l,!1),t.lanes=4194304);l.isBackwards?(i.sibling=t.child,t.child=i):(n=l.last,n!==null?n.sibling=i:t.child=i,l.last=i)}return l.tail!==null?(t=l.tail,l.rendering=t,l.tail=t.sibling,l.renderingStartTime=Ee(),t.sibling=null,n=ke.current,pe(ke,r?n&1|2:n&1),t):(Be(t),null);case 22:case 23:return Di(),r=t.memoizedState!==null,e!==null&&e.memoizedState!==null!==r&&(t.flags|=8192),r&&t.mode&1?ot&1073741824&&(Be(t),t.subtreeFlags&6&&(t.flags|=8192)):Be(t),null;case 24:return null;case 25:return null}throw Error(u(156,t.tag))}function If(e,t){switch(Gl(t),t.tag){case 1:return Ye(t.type)&&uo(),e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 3:return An(),ve(Ke),ve($e),li(),e=t.flags,e&65536&&!(e&128)?(t.flags=e&-65537|128,t):null;case 5:return ri(t),null;case 13:if(ve(ke),e=t.memoizedState,e!==null&&e.dehydrated!==null){if(t.alternate===null)throw Error(u(340));Mn()}return e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 19:return ve(ke),null;case 4:return An(),null;case 10:return bl(t.type._context),null;case 22:case 23:return Di(),null;case 24:return null;default:return null}}var Ro=!1,He=!1,Af=typeof WeakSet=="function"?WeakSet:Set,W=null;function Fn(e,t){var n=e.ref;if(n!==null)if(typeof n=="function")try{n(null)}catch(r){Ce(e,t,r)}else n.current=null}function Ni(e,t,n){try{n()}catch(r){Ce(e,t,r)}}var Ia=!1;function Df(e,t){if(Dl=Yr,e=pu(),Rl(e)){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{n=(n=e.ownerDocument)&&n.defaultView||window;var r=n.getSelection&&n.getSelection();if(r&&r.rangeCount!==0){n=r.anchorNode;var o=r.anchorOffset,l=r.focusNode;r=r.focusOffset;try{n.nodeType,l.nodeType}catch{n=null;break e}var i=0,c=-1,d=-1,w=0,_=0,P=e,E=null;t:for(;;){for(var V;P!==n||o!==0&&P.nodeType!==3||(c=i+o),P!==l||r!==0&&P.nodeType!==3||(d=i+r),P.nodeType===3&&(i+=P.nodeValue.length),(V=P.firstChild)!==null;)E=P,P=V;for(;;){if(P===e)break t;if(E===n&&++w===o&&(c=i),E===l&&++_===r&&(d=i),(V=P.nextSibling)!==null)break;P=E,E=P.parentNode}P=V}n=c===-1||d===-1?null:{start:c,end:d}}else n=null}n=n||{start:0,end:0}}else n=null;for(Fl={focusedElem:e,selectionRange:n},Yr=!1,W=t;W!==null;)if(t=W,e=t.child,(t.subtreeFlags&1028)!==0&&e!==null)e.return=t,W=e;else for(;W!==null;){t=W;try{var B=t.alternate;if(t.flags&1024)switch(t.tag){case 0:case 11:case 15:break;case 1:if(B!==null){var H=B.memoizedProps,Ne=B.memoizedState,g=t.stateNode,p=g.getSnapshotBeforeUpdate(t.elementType===t.type?H:vt(t.type,H),Ne);g.__reactInternalSnapshotBeforeUpdate=p}break;case 3:var y=t.stateNode.containerInfo;y.nodeType===1?y.textContent="":y.nodeType===9&&y.documentElement&&y.removeChild(y.documentElement);break;case 5:case 6:case 4:case 17:break;default:throw Error(u(163))}}catch(T){Ce(t,t.return,T)}if(e=t.sibling,e!==null){e.return=t.return,W=e;break}W=t.return}return B=Ia,Ia=!1,B}function Er(e,t,n){var r=t.updateQueue;if(r=r!==null?r.lastEffect:null,r!==null){var o=r=r.next;do{if((o.tag&e)===e){var l=o.destroy;o.destroy=void 0,l!==void 0&&Ni(t,n,l)}o=o.next}while(o!==r)}}function zo(e,t){if(t=t.updateQueue,t=t!==null?t.lastEffect:null,t!==null){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function _i(e){var t=e.ref;if(t!==null){var n=e.stateNode;switch(e.tag){case 5:e=n;break;default:e=n}typeof t=="function"?t(e):t.current=e}}function Aa(e){var t=e.alternate;t!==null&&(e.alternate=null,Aa(t)),e.child=null,e.deletions=null,e.sibling=null,e.tag===5&&(t=e.stateNode,t!==null&&(delete t[kt],delete t[pr],delete t[Wl],delete t[wf],delete t[xf])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function Da(e){return e.tag===5||e.tag===3||e.tag===4}function Fa(e){e:for(;;){for(;e.sibling===null;){if(e.return===null||Da(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;e.tag!==5&&e.tag!==6&&e.tag!==18;){if(e.flags&2||e.child===null||e.tag===4)continue e;e.child.return=e,e=e.child}if(!(e.flags&2))return e.stateNode}}function Pi(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.nodeType===8?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(n.nodeType===8?(t=n.parentNode,t.insertBefore(e,n)):(t=n,t.appendChild(e)),n=n._reactRootContainer,n!=null||t.onclick!==null||(t.onclick=io));else if(r!==4&&(e=e.child,e!==null))for(Pi(e,t,n),e=e.sibling;e!==null;)Pi(e,t,n),e=e.sibling}function ji(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(r!==4&&(e=e.child,e!==null))for(ji(e,t,n),e=e.sibling;e!==null;)ji(e,t,n),e=e.sibling}var Ie=null,gt=!1;function Yt(e,t,n){for(n=n.child;n!==null;)Ua(e,t,n),n=n.sibling}function Ua(e,t,n){if(xt&&typeof xt.onCommitFiberUnmount=="function")try{xt.onCommitFiberUnmount(Wr,n)}catch{}switch(n.tag){case 5:He||Fn(n,t);case 6:var r=Ie,o=gt;Ie=null,Yt(e,t,n),Ie=r,gt=o,Ie!==null&&(gt?(e=Ie,n=n.stateNode,e.nodeType===8?e.parentNode.removeChild(n):e.removeChild(n)):Ie.removeChild(n.stateNode));break;case 18:Ie!==null&&(gt?(e=Ie,n=n.stateNode,e.nodeType===8?$l(e.parentNode,n):e.nodeType===1&&$l(e,n),nr(e)):$l(Ie,n.stateNode));break;case 4:r=Ie,o=gt,Ie=n.stateNode.containerInfo,gt=!0,Yt(e,t,n),Ie=r,gt=o;break;case 0:case 11:case 14:case 15:if(!He&&(r=n.updateQueue,r!==null&&(r=r.lastEffect,r!==null))){o=r=r.next;do{var l=o,i=l.destroy;l=l.tag,i!==void 0&&(l&2||l&4)&&Ni(n,t,i),o=o.next}while(o!==r)}Yt(e,t,n);break;case 1:if(!He&&(Fn(n,t),r=n.stateNode,typeof r.componentWillUnmount=="function"))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(c){Ce(n,t,c)}Yt(e,t,n);break;case 21:Yt(e,t,n);break;case 22:n.mode&1?(He=(r=He)||n.memoizedState!==null,Yt(e,t,n),He=r):Yt(e,t,n);break;default:Yt(e,t,n)}}function Va(e){var t=e.updateQueue;if(t!==null){e.updateQueue=null;var n=e.stateNode;n===null&&(n=e.stateNode=new Af),t.forEach(function(r){var o=Gf.bind(null,e,r);n.has(r)||(n.add(r),r.then(o,o))})}}function yt(e,t){var n=t.deletions;if(n!==null)for(var r=0;r<n.length;r++){var o=n[r];try{var l=e,i=t,c=i;e:for(;c!==null;){switch(c.tag){case 5:Ie=c.stateNode,gt=!1;break e;case 3:Ie=c.stateNode.containerInfo,gt=!0;break e;case 4:Ie=c.stateNode.containerInfo,gt=!0;break e}c=c.return}if(Ie===null)throw Error(u(160));Ua(l,i,o),Ie=null,gt=!1;var d=o.alternate;d!==null&&(d.return=null),o.return=null}catch(w){Ce(o,t,w)}}if(t.subtreeFlags&12854)for(t=t.child;t!==null;)$a(t,e),t=t.sibling}function $a(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(yt(t,e),Et(e),r&4){try{Er(3,e,e.return),zo(3,e)}catch(H){Ce(e,e.return,H)}try{Er(5,e,e.return)}catch(H){Ce(e,e.return,H)}}break;case 1:yt(t,e),Et(e),r&512&&n!==null&&Fn(n,n.return);break;case 5:if(yt(t,e),Et(e),r&512&&n!==null&&Fn(n,n.return),e.flags&32){var o=e.stateNode;try{Qn(o,"")}catch(H){Ce(e,e.return,H)}}if(r&4&&(o=e.stateNode,o!=null)){var l=e.memoizedProps,i=n!==null?n.memoizedProps:l,c=e.type,d=e.updateQueue;if(e.updateQueue=null,d!==null)try{c==="input"&&l.type==="radio"&&l.name!=null&&hs(o,l),ll(c,i);var w=ll(c,l);for(i=0;i<d.length;i+=2){var _=d[i],P=d[i+1];_==="style"?Cs(o,P):_==="dangerouslySetInnerHTML"?ks(o,P):_==="children"?Qn(o,P):b(o,_,P,w)}switch(c){case"input":el(o,l);break;case"textarea":ys(o,l);break;case"select":var E=o._wrapperState.wasMultiple;o._wrapperState.wasMultiple=!!l.multiple;var V=l.value;V!=null?gn(o,!!l.multiple,V,!1):E!==!!l.multiple&&(l.defaultValue!=null?gn(o,!!l.multiple,l.defaultValue,!0):gn(o,!!l.multiple,l.multiple?[]:"",!1))}o[pr]=l}catch(H){Ce(e,e.return,H)}}break;case 6:if(yt(t,e),Et(e),r&4){if(e.stateNode===null)throw Error(u(162));o=e.stateNode,l=e.memoizedProps;try{o.nodeValue=l}catch(H){Ce(e,e.return,H)}}break;case 3:if(yt(t,e),Et(e),r&4&&n!==null&&n.memoizedState.isDehydrated)try{nr(t.containerInfo)}catch(H){Ce(e,e.return,H)}break;case 4:yt(t,e),Et(e);break;case 13:yt(t,e),Et(e),o=e.child,o.flags&8192&&(l=o.memoizedState!==null,o.stateNode.isHidden=l,!l||o.alternate!==null&&o.alternate.memoizedState!==null||(Ti=Ee())),r&4&&Va(e);break;case 22:if(_=n!==null&&n.memoizedState!==null,e.mode&1?(He=(w=He)||_,yt(t,e),He=w):yt(t,e),Et(e),r&8192){if(w=e.memoizedState!==null,(e.stateNode.isHidden=w)&&!_&&e.mode&1)for(W=e,_=e.child;_!==null;){for(P=W=_;W!==null;){switch(E=W,V=E.child,E.tag){case 0:case 11:case 14:case 15:Er(4,E,E.return);break;case 1:Fn(E,E.return);var B=E.stateNode;if(typeof B.componentWillUnmount=="function"){r=E,n=E.return;try{t=r,B.props=t.memoizedProps,B.state=t.memoizedState,B.componentWillUnmount()}catch(H){Ce(r,n,H)}}break;case 5:Fn(E,E.return);break;case 22:if(E.memoizedState!==null){Ha(P);continue}}V!==null?(V.return=E,W=V):Ha(P)}_=_.sibling}e:for(_=null,P=e;;){if(P.tag===5){if(_===null){_=P;try{o=P.stateNode,w?(l=o.style,typeof l.setProperty=="function"?l.setProperty("display","none","important"):l.display="none"):(c=P.stateNode,d=P.memoizedProps.style,i=d!=null&&d.hasOwnProperty("display")?d.display:null,c.style.display=Ss("display",i))}catch(H){Ce(e,e.return,H)}}}else if(P.tag===6){if(_===null)try{P.stateNode.nodeValue=w?"":P.memoizedProps}catch(H){Ce(e,e.return,H)}}else if((P.tag!==22&&P.tag!==23||P.memoizedState===null||P===e)&&P.child!==null){P.child.return=P,P=P.child;continue}if(P===e)break e;for(;P.sibling===null;){if(P.return===null||P.return===e)break e;_===P&&(_=null),P=P.return}_===P&&(_=null),P.sibling.return=P.return,P=P.sibling}}break;case 19:yt(t,e),Et(e),r&4&&Va(e);break;case 21:break;default:yt(t,e),Et(e)}}function Et(e){var t=e.flags;if(t&2){try{e:{for(var n=e.return;n!==null;){if(Da(n)){var r=n;break e}n=n.return}throw Error(u(160))}switch(r.tag){case 5:var o=r.stateNode;r.flags&32&&(Qn(o,""),r.flags&=-33);var l=Fa(e);ji(e,l,o);break;case 3:case 4:var i=r.stateNode.containerInfo,c=Fa(e);Pi(e,c,i);break;default:throw Error(u(161))}}catch(d){Ce(e,e.return,d)}e.flags&=-3}t&4096&&(e.flags&=-4097)}function Ff(e,t,n){W=e,Wa(e)}function Wa(e,t,n){for(var r=(e.mode&1)!==0;W!==null;){var o=W,l=o.child;if(o.tag===22&&r){var i=o.memoizedState!==null||Ro;if(!i){var c=o.alternate,d=c!==null&&c.memoizedState!==null||He;c=Ro;var w=He;if(Ro=i,(He=d)&&!w)for(W=o;W!==null;)i=W,d=i.child,i.tag===22&&i.memoizedState!==null?Qa(o):d!==null?(d.return=i,W=d):Qa(o);for(;l!==null;)W=l,Wa(l),l=l.sibling;W=o,Ro=c,He=w}Ba(e)}else o.subtreeFlags&8772&&l!==null?(l.return=o,W=l):Ba(e)}}function Ba(e){for(;W!==null;){var t=W;if(t.flags&8772){var n=t.alternate;try{if(t.flags&8772)switch(t.tag){case 0:case 11:case 15:He||zo(5,t);break;case 1:var r=t.stateNode;if(t.flags&4&&!He)if(n===null)r.componentDidMount();else{var o=t.elementType===t.type?n.memoizedProps:vt(t.type,n.memoizedProps);r.componentDidUpdate(o,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var l=t.updateQueue;l!==null&&Hu(t,l,r);break;case 3:var i=t.updateQueue;if(i!==null){if(n=null,t.child!==null)switch(t.child.tag){case 5:n=t.child.stateNode;break;case 1:n=t.child.stateNode}Hu(t,i,n)}break;case 5:var c=t.stateNode;if(n===null&&t.flags&4){n=c;var d=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":d.autoFocus&&n.focus();break;case"img":d.src&&(n.src=d.src)}}break;case 6:break;case 4:break;case 12:break;case 13:if(t.memoizedState===null){var w=t.alternate;if(w!==null){var _=w.memoizedState;if(_!==null){var P=_.dehydrated;P!==null&&nr(P)}}}break;case 19:case 17:case 21:case 22:case 23:case 25:break;default:throw Error(u(163))}He||t.flags&512&&_i(t)}catch(E){Ce(t,t.return,E)}}if(t===e){W=null;break}if(n=t.sibling,n!==null){n.return=t.return,W=n;break}W=t.return}}function Ha(e){for(;W!==null;){var t=W;if(t===e){W=null;break}var n=t.sibling;if(n!==null){n.return=t.return,W=n;break}W=t.return}}function Qa(e){for(;W!==null;){var t=W;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{zo(4,t)}catch(d){Ce(t,n,d)}break;case 1:var r=t.stateNode;if(typeof r.componentDidMount=="function"){var o=t.return;try{r.componentDidMount()}catch(d){Ce(t,o,d)}}var l=t.return;try{_i(t)}catch(d){Ce(t,l,d)}break;case 5:var i=t.return;try{_i(t)}catch(d){Ce(t,i,d)}}}catch(d){Ce(t,t.return,d)}if(t===e){W=null;break}var c=t.sibling;if(c!==null){c.return=t.return,W=c;break}W=t.return}}var Uf=Math.ceil,To=J.ReactCurrentDispatcher,Ri=J.ReactCurrentOwner,at=J.ReactCurrentBatchConfig,oe=0,Le=null,Pe=null,Ae=0,ot=0,Un=Bt(0),Te=0,Nr=null,fn=0,Mo=0,zi=0,_r=null,Ze=null,Ti=0,Vn=1/0,Mt=null,Lo=!1,Mi=null,Xt=null,Oo=!1,Zt=null,Io=0,Pr=0,Li=null,Ao=-1,Do=0;function Ge(){return oe&6?Ee():Ao!==-1?Ao:Ao=Ee()}function Jt(e){return e.mode&1?oe&2&&Ae!==0?Ae&-Ae:Sf.transition!==null?(Do===0&&(Do=Fs()),Do):(e=ce,e!==0||(e=window.event,e=e===void 0?16:Ks(e.type)),e):1}function wt(e,t,n,r){if(50<Pr)throw Pr=0,Li=null,Error(u(185));Jn(e,n,r),(!(oe&2)||e!==Le)&&(e===Le&&(!(oe&2)&&(Mo|=n),Te===4&&bt(e,Ae)),Je(e,r),n===1&&oe===0&&!(t.mode&1)&&(Vn=Ee()+500,co&&Qt()))}function Je(e,t){var n=e.callbackNode;kd(e,t);var r=Qr(e,e===Le?Ae:0);if(r===0)n!==null&&Is(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(n!=null&&Is(n),t===1)e.tag===0?kf(Ka.bind(null,e)):Mu(Ka.bind(null,e)),gf(function(){!(oe&6)&&Qt()}),n=null;else{switch(Us(r)){case 1:n=fl;break;case 4:n=As;break;case 16:n=$r;break;case 536870912:n=Ds;break;default:n=$r}n=tc(n,Ga.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function Ga(e,t){if(Ao=-1,Do=0,oe&6)throw Error(u(327));var n=e.callbackNode;if($n()&&e.callbackNode!==n)return null;var r=Qr(e,e===Le?Ae:0);if(r===0)return null;if(r&30||r&e.expiredLanes||t)t=Fo(e,r);else{t=r;var o=oe;oe|=2;var l=Xa();(Le!==e||Ae!==t)&&(Mt=null,Vn=Ee()+500,mn(e,t));do try{Wf();break}catch(c){Ya(e,c)}while(!0);Jl(),To.current=l,oe=o,Pe!==null?t=0:(Le=null,Ae=0,t=Te)}if(t!==0){if(t===2&&(o=pl(e),o!==0&&(r=o,t=Oi(e,o))),t===1)throw n=Nr,mn(e,0),bt(e,r),Je(e,Ee()),n;if(t===6)bt(e,r);else{if(o=e.current.alternate,!(r&30)&&!Vf(o)&&(t=Fo(e,r),t===2&&(l=pl(e),l!==0&&(r=l,t=Oi(e,l))),t===1))throw n=Nr,mn(e,0),bt(e,r),Je(e,Ee()),n;switch(e.finishedWork=o,e.finishedLanes=r,t){case 0:case 1:throw Error(u(345));case 2:hn(e,Ze,Mt);break;case 3:if(bt(e,r),(r&130023424)===r&&(t=Ti+500-Ee(),10<t)){if(Qr(e,0)!==0)break;if(o=e.suspendedLanes,(o&r)!==r){Ge(),e.pingedLanes|=e.suspendedLanes&o;break}e.timeoutHandle=Vl(hn.bind(null,e,Ze,Mt),t);break}hn(e,Ze,Mt);break;case 4:if(bt(e,r),(r&4194240)===r)break;for(t=e.eventTimes,o=-1;0<r;){var i=31-pt(r);l=1<<i,i=t[i],i>o&&(o=i),r&=~l}if(r=o,r=Ee()-r,r=(120>r?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*Uf(r/1960))-r,10<r){e.timeoutHandle=Vl(hn.bind(null,e,Ze,Mt),r);break}hn(e,Ze,Mt);break;case 5:hn(e,Ze,Mt);break;default:throw Error(u(329))}}}return Je(e,Ee()),e.callbackNode===n?Ga.bind(null,e):null}function Oi(e,t){var n=_r;return e.current.memoizedState.isDehydrated&&(mn(e,t).flags|=256),e=Fo(e,t),e!==2&&(t=Ze,Ze=n,t!==null&&Ii(t)),e}function Ii(e){Ze===null?Ze=e:Ze.push.apply(Ze,e)}function Vf(e){for(var t=e;;){if(t.flags&16384){var n=t.updateQueue;if(n!==null&&(n=n.stores,n!==null))for(var r=0;r<n.length;r++){var o=n[r],l=o.getSnapshot;o=o.value;try{if(!mt(l(),o))return!1}catch{return!1}}}if(n=t.child,t.subtreeFlags&16384&&n!==null)n.return=t,t=n;else{if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function bt(e,t){for(t&=~zi,t&=~Mo,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-pt(t),r=1<<n;e[n]=-1,t&=~r}}function Ka(e){if(oe&6)throw Error(u(327));$n();var t=Qr(e,0);if(!(t&1))return Je(e,Ee()),null;var n=Fo(e,t);if(e.tag!==0&&n===2){var r=pl(e);r!==0&&(t=r,n=Oi(e,r))}if(n===1)throw n=Nr,mn(e,0),bt(e,t),Je(e,Ee()),n;if(n===6)throw Error(u(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,hn(e,Ze,Mt),Je(e,Ee()),null}function Ai(e,t){var n=oe;oe|=1;try{return e(t)}finally{oe=n,oe===0&&(Vn=Ee()+500,co&&Qt())}}function pn(e){Zt!==null&&Zt.tag===0&&!(oe&6)&&$n();var t=oe;oe|=1;var n=at.transition,r=ce;try{if(at.transition=null,ce=1,e)return e()}finally{ce=r,at.transition=n,oe=t,!(oe&6)&&Qt()}}function Di(){ot=Un.current,ve(Un)}function mn(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(n!==-1&&(e.timeoutHandle=-1,vf(n)),Pe!==null)for(n=Pe.return;n!==null;){var r=n;switch(Gl(r),r.tag){case 1:r=r.type.childContextTypes,r!=null&&uo();break;case 3:An(),ve(Ke),ve($e),li();break;case 5:ri(r);break;case 4:An();break;case 13:ve(ke);break;case 19:ve(ke);break;case 10:bl(r.type._context);break;case 22:case 23:Di()}n=n.return}if(Le=e,Pe=e=qt(e.current,null),Ae=ot=t,Te=0,Nr=null,zi=Mo=fn=0,Ze=_r=null,an!==null){for(t=0;t<an.length;t++)if(n=an[t],r=n.interleaved,r!==null){n.interleaved=null;var o=r.next,l=n.pending;if(l!==null){var i=l.next;l.next=o,r.next=i}n.pending=r}an=null}return e}function Ya(e,t){do{var n=Pe;try{if(Jl(),ko.current=No,So){for(var r=Se.memoizedState;r!==null;){var o=r.queue;o!==null&&(o.pending=null),r=r.next}So=!1}if(dn=0,Me=ze=Se=null,wr=!1,xr=0,Ri.current=null,n===null||n.return===null){Te=1,Nr=t,Pe=null;break}e:{var l=e,i=n.return,c=n,d=t;if(t=Ae,c.flags|=32768,d!==null&&typeof d=="object"&&typeof d.then=="function"){var w=d,_=c,P=_.tag;if(!(_.mode&1)&&(P===0||P===11||P===15)){var E=_.alternate;E?(_.updateQueue=E.updateQueue,_.memoizedState=E.memoizedState,_.lanes=E.lanes):(_.updateQueue=null,_.memoizedState=null)}var V=wa(i);if(V!==null){V.flags&=-257,xa(V,i,c,l,t),V.mode&1&&ya(l,w,t),t=V,d=w;var B=t.updateQueue;if(B===null){var H=new Set;H.add(d),t.updateQueue=H}else B.add(d);break e}else{if(!(t&1)){ya(l,w,t),Fi();break e}d=Error(u(426))}}else if(we&&c.mode&1){var Ne=wa(i);if(Ne!==null){!(Ne.flags&65536)&&(Ne.flags|=256),xa(Ne,i,c,l,t),Xl(Dn(d,c));break e}}l=d=Dn(d,c),Te!==4&&(Te=2),_r===null?_r=[l]:_r.push(l),l=i;do{switch(l.tag){case 3:l.flags|=65536,t&=-t,l.lanes|=t;var g=va(l,d,t);Bu(l,g);break e;case 1:c=d;var p=l.type,y=l.stateNode;if(!(l.flags&128)&&(typeof p.getDerivedStateFromError=="function"||y!==null&&typeof y.componentDidCatch=="function"&&(Xt===null||!Xt.has(y)))){l.flags|=65536,t&=-t,l.lanes|=t;var T=ga(l,c,t);Bu(l,T);break e}}l=l.return}while(l!==null)}Ja(n)}catch(G){t=G,Pe===n&&n!==null&&(Pe=n=n.return);continue}break}while(!0)}function Xa(){var e=To.current;return To.current=No,e===null?No:e}function Fi(){(Te===0||Te===3||Te===2)&&(Te=4),Le===null||!(fn&268435455)&&!(Mo&268435455)||bt(Le,Ae)}function Fo(e,t){var n=oe;oe|=2;var r=Xa();(Le!==e||Ae!==t)&&(Mt=null,mn(e,t));do try{$f();break}catch(o){Ya(e,o)}while(!0);if(Jl(),oe=n,To.current=r,Pe!==null)throw Error(u(261));return Le=null,Ae=0,Te}function $f(){for(;Pe!==null;)Za(Pe)}function Wf(){for(;Pe!==null&&!fd();)Za(Pe)}function Za(e){var t=ec(e.alternate,e,ot);e.memoizedProps=e.pendingProps,t===null?Ja(e):Pe=t,Ri.current=null}function Ja(e){var t=e;do{var n=t.alternate;if(e=t.return,t.flags&32768){if(n=If(n,t),n!==null){n.flags&=32767,Pe=n;return}if(e!==null)e.flags|=32768,e.subtreeFlags=0,e.deletions=null;else{Te=6,Pe=null;return}}else if(n=Of(n,t,ot),n!==null){Pe=n;return}if(t=t.sibling,t!==null){Pe=t;return}Pe=t=e}while(t!==null);Te===0&&(Te=5)}function hn(e,t,n){var r=ce,o=at.transition;try{at.transition=null,ce=1,Bf(e,t,n,r)}finally{at.transition=o,ce=r}return null}function Bf(e,t,n,r){do $n();while(Zt!==null);if(oe&6)throw Error(u(327));n=e.finishedWork;var o=e.finishedLanes;if(n===null)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(u(177));e.callbackNode=null,e.callbackPriority=0;var l=n.lanes|n.childLanes;if(Sd(e,l),e===Le&&(Pe=Le=null,Ae=0),!(n.subtreeFlags&2064)&&!(n.flags&2064)||Oo||(Oo=!0,tc($r,function(){return $n(),null})),l=(n.flags&15990)!==0,n.subtreeFlags&15990||l){l=at.transition,at.transition=null;var i=ce;ce=1;var c=oe;oe|=4,Ri.current=null,Df(e,n),$a(n,e),af(Fl),Yr=!!Dl,Fl=Dl=null,e.current=n,Ff(n),pd(),oe=c,ce=i,at.transition=l}else e.current=n;if(Oo&&(Oo=!1,Zt=e,Io=o),l=e.pendingLanes,l===0&&(Xt=null),vd(n.stateNode),Je(e,Ee()),t!==null)for(r=e.onRecoverableError,n=0;n<t.length;n++)o=t[n],r(o.value,{componentStack:o.stack,digest:o.digest});if(Lo)throw Lo=!1,e=Mi,Mi=null,e;return Io&1&&e.tag!==0&&$n(),l=e.pendingLanes,l&1?e===Li?Pr++:(Pr=0,Li=e):Pr=0,Qt(),null}function $n(){if(Zt!==null){var e=Us(Io),t=at.transition,n=ce;try{if(at.transition=null,ce=16>e?16:e,Zt===null)var r=!1;else{if(e=Zt,Zt=null,Io=0,oe&6)throw Error(u(331));var o=oe;for(oe|=4,W=e.current;W!==null;){var l=W,i=l.child;if(W.flags&16){var c=l.deletions;if(c!==null){for(var d=0;d<c.length;d++){var w=c[d];for(W=w;W!==null;){var _=W;switch(_.tag){case 0:case 11:case 15:Er(8,_,l)}var P=_.child;if(P!==null)P.return=_,W=P;else for(;W!==null;){_=W;var E=_.sibling,V=_.return;if(Aa(_),_===w){W=null;break}if(E!==null){E.return=V,W=E;break}W=V}}}var B=l.alternate;if(B!==null){var H=B.child;if(H!==null){B.child=null;do{var Ne=H.sibling;H.sibling=null,H=Ne}while(H!==null)}}W=l}}if(l.subtreeFlags&2064&&i!==null)i.return=l,W=i;else e:for(;W!==null;){if(l=W,l.flags&2048)switch(l.tag){case 0:case 11:case 15:Er(9,l,l.return)}var g=l.sibling;if(g!==null){g.return=l.return,W=g;break e}W=l.return}}var p=e.current;for(W=p;W!==null;){i=W;var y=i.child;if(i.subtreeFlags&2064&&y!==null)y.return=i,W=y;else e:for(i=p;W!==null;){if(c=W,c.flags&2048)try{switch(c.tag){case 0:case 11:case 15:zo(9,c)}}catch(G){Ce(c,c.return,G)}if(c===i){W=null;break e}var T=c.sibling;if(T!==null){T.return=c.return,W=T;break e}W=c.return}}if(oe=o,Qt(),xt&&typeof xt.onPostCommitFiberRoot=="function")try{xt.onPostCommitFiberRoot(Wr,e)}catch{}r=!0}return r}finally{ce=n,at.transition=t}}return!1}function ba(e,t,n){t=Dn(n,t),t=va(e,t,1),e=Kt(e,t,1),t=Ge(),e!==null&&(Jn(e,1,t),Je(e,t))}function Ce(e,t,n){if(e.tag===3)ba(e,e,n);else for(;t!==null;){if(t.tag===3){ba(t,e,n);break}else if(t.tag===1){var r=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof r.componentDidCatch=="function"&&(Xt===null||!Xt.has(r))){e=Dn(n,e),e=ga(t,e,1),t=Kt(t,e,1),e=Ge(),t!==null&&(Jn(t,1,e),Je(t,e));break}}t=t.return}}function Hf(e,t,n){var r=e.pingCache;r!==null&&r.delete(t),t=Ge(),e.pingedLanes|=e.suspendedLanes&n,Le===e&&(Ae&n)===n&&(Te===4||Te===3&&(Ae&130023424)===Ae&&500>Ee()-Ti?mn(e,0):zi|=n),Je(e,t)}function qa(e,t){t===0&&(e.mode&1?(t=Hr,Hr<<=1,!(Hr&130023424)&&(Hr=4194304)):t=1);var n=Ge();e=Rt(e,t),e!==null&&(Jn(e,t,n),Je(e,n))}function Qf(e){var t=e.memoizedState,n=0;t!==null&&(n=t.retryLane),qa(e,n)}function Gf(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,o=e.memoizedState;o!==null&&(n=o.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(u(314))}r!==null&&r.delete(t),qa(e,n)}var ec;ec=function(e,t,n){if(e!==null)if(e.memoizedProps!==t.pendingProps||Ke.current)Xe=!0;else{if(!(e.lanes&n)&&!(t.flags&128))return Xe=!1,Lf(e,t,n);Xe=!!(e.flags&131072)}else Xe=!1,we&&t.flags&1048576&&Lu(t,po,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;jo(e,t),e=t.pendingProps;var o=Rn(t,$e.current);In(t,n),o=ui(null,t,r,e,o,n);var l=ai();return t.flags|=1,typeof o=="object"&&o!==null&&typeof o.render=="function"&&o.$$typeof===void 0?(t.tag=1,t.memoizedState=null,t.updateQueue=null,Ye(r)?(l=!0,ao(t)):l=!1,t.memoizedState=o.state!==null&&o.state!==void 0?o.state:null,ti(t),o.updater=_o,t.stateNode=o,o._reactInternals=t,hi(t,r,e,n),t=wi(null,t,r,!0,l,n)):(t.tag=0,we&&l&&Ql(t),Qe(null,t,o,n),t=t.child),t;case 16:r=t.elementType;e:{switch(jo(e,t),e=t.pendingProps,o=r._init,r=o(r._payload),t.type=r,o=t.tag=Yf(r),e=vt(r,e),o){case 0:t=yi(null,t,r,e,n);break e;case 1:t=_a(null,t,r,e,n);break e;case 11:t=ka(null,t,r,e,n);break e;case 14:t=Sa(null,t,r,vt(r.type,e),n);break e}throw Error(u(306,r,""))}return t;case 0:return r=t.type,o=t.pendingProps,o=t.elementType===r?o:vt(r,o),yi(e,t,r,o,n);case 1:return r=t.type,o=t.pendingProps,o=t.elementType===r?o:vt(r,o),_a(e,t,r,o,n);case 3:e:{if(Pa(t),e===null)throw Error(u(387));r=t.pendingProps,l=t.memoizedState,o=l.element,Wu(e,t),wo(t,r,null,n);var i=t.memoizedState;if(r=i.element,l.isDehydrated)if(l={element:r,isDehydrated:!1,cache:i.cache,pendingSuspenseBoundaries:i.pendingSuspenseBoundaries,transitions:i.transitions},t.updateQueue.baseState=l,t.memoizedState=l,t.flags&256){o=Dn(Error(u(423)),t),t=ja(e,t,r,n,o);break e}else if(r!==o){o=Dn(Error(u(424)),t),t=ja(e,t,r,n,o);break e}else for(rt=Wt(t.stateNode.containerInfo.firstChild),nt=t,we=!0,ht=null,n=Vu(t,null,r,n),t.child=n;n;)n.flags=n.flags&-3|4096,n=n.sibling;else{if(Mn(),r===o){t=Tt(e,t,n);break e}Qe(e,t,r,n)}t=t.child}return t;case 5:return Qu(t),e===null&&Yl(t),r=t.type,o=t.pendingProps,l=e!==null?e.memoizedProps:null,i=o.children,Ul(r,o)?i=null:l!==null&&Ul(r,l)&&(t.flags|=32),Na(e,t),Qe(e,t,i,n),t.child;case 6:return e===null&&Yl(t),null;case 13:return Ra(e,t,n);case 4:return ni(t,t.stateNode.containerInfo),r=t.pendingProps,e===null?t.child=Ln(t,null,r,n):Qe(e,t,r,n),t.child;case 11:return r=t.type,o=t.pendingProps,o=t.elementType===r?o:vt(r,o),ka(e,t,r,o,n);case 7:return Qe(e,t,t.pendingProps,n),t.child;case 8:return Qe(e,t,t.pendingProps.children,n),t.child;case 12:return Qe(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,o=t.pendingProps,l=t.memoizedProps,i=o.value,pe(vo,r._currentValue),r._currentValue=i,l!==null)if(mt(l.value,i)){if(l.children===o.children&&!Ke.current){t=Tt(e,t,n);break e}}else for(l=t.child,l!==null&&(l.return=t);l!==null;){var c=l.dependencies;if(c!==null){i=l.child;for(var d=c.firstContext;d!==null;){if(d.context===r){if(l.tag===1){d=zt(-1,n&-n),d.tag=2;var w=l.updateQueue;if(w!==null){w=w.shared;var _=w.pending;_===null?d.next=d:(d.next=_.next,_.next=d),w.pending=d}}l.lanes|=n,d=l.alternate,d!==null&&(d.lanes|=n),ql(l.return,n,t),c.lanes|=n;break}d=d.next}}else if(l.tag===10)i=l.type===t.type?null:l.child;else if(l.tag===18){if(i=l.return,i===null)throw Error(u(341));i.lanes|=n,c=i.alternate,c!==null&&(c.lanes|=n),ql(i,n,t),i=l.sibling}else i=l.child;if(i!==null)i.return=l;else for(i=l;i!==null;){if(i===t){i=null;break}if(l=i.sibling,l!==null){l.return=i.return,i=l;break}i=i.return}l=i}Qe(e,t,o.children,n),t=t.child}return t;case 9:return o=t.type,r=t.pendingProps.children,In(t,n),o=st(o),r=r(o),t.flags|=1,Qe(e,t,r,n),t.child;case 14:return r=t.type,o=vt(r,t.pendingProps),o=vt(r.type,o),Sa(e,t,r,o,n);case 15:return Ca(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,o=t.pendingProps,o=t.elementType===r?o:vt(r,o),jo(e,t),t.tag=1,Ye(r)?(e=!0,ao(t)):e=!1,In(t,n),ma(t,r,o),hi(t,r,o,n),wi(null,t,r,!0,e,n);case 19:return Ta(e,t,n);case 22:return Ea(e,t,n)}throw Error(u(156,t.tag))};function tc(e,t){return Os(e,t)}function Kf(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function ct(e,t,n,r){return new Kf(e,t,n,r)}function Ui(e){return e=e.prototype,!(!e||!e.isReactComponent)}function Yf(e){if(typeof e=="function")return Ui(e)?1:0;if(e!=null){if(e=e.$$typeof,e===re)return 11;if(e===qe)return 14}return 2}function qt(e,t){var n=e.alternate;return n===null?(n=ct(e.tag,t,e.key,e.mode),n.elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=e.flags&14680064,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Uo(e,t,n,r,o,l){var i=2;if(r=e,typeof e=="function")Ui(e)&&(i=1);else if(typeof e=="string")i=5;else e:switch(e){case xe:return vn(n.children,o,l,t);case _e:i=8,o|=8;break;case De:return e=ct(12,n,t,o|2),e.elementType=De,e.lanes=l,e;case Ue:return e=ct(13,n,t,o),e.elementType=Ue,e.lanes=l,e;case Ve:return e=ct(19,n,t,o),e.elementType=Ve,e.lanes=l,e;case me:return Vo(n,o,l,t);default:if(typeof e=="object"&&e!==null)switch(e.$$typeof){case Fe:i=10;break e;case ft:i=9;break e;case re:i=11;break e;case qe:i=14;break e;case Re:i=16,r=null;break e}throw Error(u(130,e==null?e:typeof e,""))}return t=ct(i,n,t,o),t.elementType=e,t.type=r,t.lanes=l,t}function vn(e,t,n,r){return e=ct(7,e,r,t),e.lanes=n,e}function Vo(e,t,n,r){return e=ct(22,e,r,t),e.elementType=me,e.lanes=n,e.stateNode={isHidden:!1},e}function Vi(e,t,n){return e=ct(6,e,null,t),e.lanes=n,e}function $i(e,t,n){return t=ct(4,e.children!==null?e.children:[],e.key,t),t.lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function Xf(e,t,n,r,o){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=ml(0),this.expirationTimes=ml(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=ml(0),this.identifierPrefix=r,this.onRecoverableError=o,this.mutableSourceEagerHydrationData=null}function Wi(e,t,n,r,o,l,i,c,d){return e=new Xf(e,t,n,c,d),t===1?(t=1,l===!0&&(t|=8)):t=0,l=ct(3,null,null,t),e.current=l,l.stateNode=e,l.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},ti(l),e}function Zf(e,t,n){var r=3<arguments.length&&arguments[3]!==void 0?arguments[3]:null;return{$$typeof:de,key:r==null?null:""+r,children:e,containerInfo:t,implementation:n}}function nc(e){if(!e)return Ht;e=e._reactInternals;e:{if(rn(e)!==e||e.tag!==1)throw Error(u(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(Ye(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(t!==null);throw Error(u(171))}if(e.tag===1){var n=e.type;if(Ye(n))return zu(e,n,t)}return t}function rc(e,t,n,r,o,l,i,c,d){return e=Wi(n,r,!0,e,o,l,i,c,d),e.context=nc(null),n=e.current,r=Ge(),o=Jt(n),l=zt(r,o),l.callback=t??null,Kt(n,l,o),e.current.lanes=o,Jn(e,o,r),Je(e,r),e}function $o(e,t,n,r){var o=t.current,l=Ge(),i=Jt(o);return n=nc(n),t.context===null?t.context=n:t.pendingContext=n,t=zt(l,i),t.payload={element:e},r=r===void 0?null:r,r!==null&&(t.callback=r),e=Kt(o,t,i),e!==null&&(wt(e,o,i,l),yo(e,o,i)),i}function Wo(e){if(e=e.current,!e.child)return null;switch(e.child.tag){case 5:return e.child.stateNode;default:return e.child.stateNode}}function oc(e,t){if(e=e.memoizedState,e!==null&&e.dehydrated!==null){var n=e.retryLane;e.retryLane=n!==0&&n<t?n:t}}function Bi(e,t){oc(e,t),(e=e.alternate)&&oc(e,t)}var lc=typeof reportError=="function"?reportError:function(e){console.error(e)};function Hi(e){this._internalRoot=e}Bo.prototype.render=Hi.prototype.render=function(e){var t=this._internalRoot;if(t===null)throw Error(u(409));$o(e,t,null,null)},Bo.prototype.unmount=Hi.prototype.unmount=function(){var e=this._internalRoot;if(e!==null){this._internalRoot=null;var t=e.containerInfo;pn(function(){$o(null,e,null,null)}),t[Nt]=null}};function Bo(e){this._internalRoot=e}Bo.prototype.unstable_scheduleHydration=function(e){if(e){var t=Ws();e={blockedOn:null,target:e,priority:t};for(var n=0;n<Ut.length&&t!==0&&t<Ut[n].priority;n++);Ut.splice(n,0,e),n===0&&Qs(e)}};function Qi(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11)}function Ho(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11&&(e.nodeType!==8||e.nodeValue!==" react-mount-point-unstable "))}function ic(){}function Jf(e,t,n,r,o){if(o){if(typeof r=="function"){var l=r;r=function(){var w=Wo(i);l.call(w)}}var i=rc(t,r,e,0,null,!1,!1,"",ic);return e._reactRootContainer=i,e[Nt]=i.current,dr(e.nodeType===8?e.parentNode:e),pn(),i}for(;o=e.lastChild;)e.removeChild(o);if(typeof r=="function"){var c=r;r=function(){var w=Wo(d);c.call(w)}}var d=Wi(e,0,!1,null,null,!1,!1,"",ic);return e._reactRootContainer=d,e[Nt]=d.current,dr(e.nodeType===8?e.parentNode:e),pn(function(){$o(t,d,n,r)}),d}function Qo(e,t,n,r,o){var l=n._reactRootContainer;if(l){var i=l;if(typeof o=="function"){var c=o;o=function(){var d=Wo(i);c.call(d)}}$o(t,i,e,o)}else i=Jf(n,t,e,o,r);return Wo(i)}Vs=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=Zn(t.pendingLanes);n!==0&&(hl(t,n|1),Je(t,Ee()),!(oe&6)&&(Vn=Ee()+500,Qt()))}break;case 13:pn(function(){var r=Rt(e,1);if(r!==null){var o=Ge();wt(r,e,1,o)}}),Bi(e,1)}},vl=function(e){if(e.tag===13){var t=Rt(e,134217728);if(t!==null){var n=Ge();wt(t,e,134217728,n)}Bi(e,134217728)}},$s=function(e){if(e.tag===13){var t=Jt(e),n=Rt(e,t);if(n!==null){var r=Ge();wt(n,e,t,r)}Bi(e,t)}},Ws=function(){return ce},Bs=function(e,t){var n=ce;try{return ce=e,t()}finally{ce=n}},ul=function(e,t,n){switch(t){case"input":if(el(e,n),t=n.name,n.type==="radio"&&t!=null){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var o=so(r);if(!o)throw Error(u(90));ps(r),el(r,o)}}}break;case"textarea":ys(e,n);break;case"select":t=n.value,t!=null&&gn(e,!!n.multiple,t,!1)}},Ps=Ai,js=pn;var bf={usingClientEntryPoint:!1,Events:[mr,Pn,so,Ns,_s,Ai]},jr={findFiberByHostInstance:on,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},qf={bundleType:jr.bundleType,version:jr.version,rendererPackageName:jr.rendererPackageName,rendererConfig:jr.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:J.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return e=Ms(e),e===null?null:e.stateNode},findFiberByHostInstance:jr.findFiberByHostInstance,findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var Go=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!Go.isDisabled&&Go.supportsFiber)try{Wr=Go.inject(qf),xt=Go}catch{}}return be.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=bf,be.createPortal=function(e,t){var n=2<arguments.length&&arguments[2]!==void 0?arguments[2]:null;if(!Qi(t))throw Error(u(200));return Zf(e,t,null,n)},be.createRoot=function(e,t){if(!Qi(e))throw Error(u(299));var n=!1,r="",o=lc;return t!=null&&(t.unstable_strictMode===!0&&(n=!0),t.identifierPrefix!==void 0&&(r=t.identifierPrefix),t.onRecoverableError!==void 0&&(o=t.onRecoverableError)),t=Wi(e,1,!1,null,null,n,!1,r,o),e[Nt]=t.current,dr(e.nodeType===8?e.parentNode:e),new Hi(t)},be.findDOMNode=function(e){if(e==null)return null;if(e.nodeType===1)return e;var t=e._reactInternals;if(t===void 0)throw typeof e.render=="function"?Error(u(188)):(e=Object.keys(e).join(","),Error(u(268,e)));return e=Ms(t),e=e===null?null:e.stateNode,e},be.flushSync=function(e){return pn(e)},be.hydrate=function(e,t,n){if(!Ho(t))throw Error(u(200));return Qo(null,e,t,!0,n)},be.hydrateRoot=function(e,t,n){if(!Qi(e))throw Error(u(405));var r=n!=null&&n.hydratedSources||null,o=!1,l="",i=lc;if(n!=null&&(n.unstable_strictMode===!0&&(o=!0),n.identifierPrefix!==void 0&&(l=n.identifierPrefix),n.onRecoverableError!==void 0&&(i=n.onRecoverableError)),t=rc(t,null,e,1,n??null,o,!1,l,i),e[Nt]=t.current,dr(e),r)for(e=0;e<r.length;e++)n=r[e],o=n._getVersion,o=o(n._source),t.mutableSourceEagerHydrationData==null?t.mutableSourceEagerHydrationData=[n,o]:t.mutableSourceEagerHydrationData.push(n,o);return new Bo(t)},be.render=function(e,t,n){if(!Ho(t))throw Error(u(200));return Qo(null,e,t,!1,n)},be.unmountComponentAtNode=function(e){if(!Ho(e))throw Error(u(40));return e._reactRootContainer?(pn(function(){Qo(null,null,e,!1,function(){e._reactRootContainer=null,e[Nt]=null})}),!0):!1},be.unstable_batchedUpdates=Ai,be.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!Ho(n))throw Error(u(200));if(e==null||e._reactInternals===void 0)throw Error(u(38));return Qo(e,t,n,!1,r)},be.version="18.3.1-next-f1338f8080-20240426",be}var mc;function Sc(){if(mc)return Yi.exports;mc=1;function s(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(s)}catch(a){console.error(a)}}return s(),Yi.exports=ap(),Yi.exports}var hc;function cp(){if(hc)return Ko;hc=1;var s=Sc();return Ko.createRoot=s.createRoot,Ko.hydrateRoot=s.hydrateRoot,Ko}var dp=cp();function rs(s,a=[]){let u=[];function f(h,k){const x=O.createContext(k),C=u.length;u=[...u,k];const j=A=>{var U;const{scope:F,children:Q,...z}=A,M=((U=F==null?void 0:F[s])==null?void 0:U[C])||x,R=O.useMemo(()=>z,Object.values(z));return S.jsx(M.Provider,{value:R,children:Q})};j.displayName=h+"Provider";function I(A,F){var M;const Q=((M=F==null?void 0:F[s])==null?void 0:M[C])||x,z=O.useContext(Q);if(z)return z;if(k!==void 0)return k;throw new Error(`\`${A}\` must be used within \`${h}\``)}return[j,I]}const m=()=>{const h=u.map(k=>O.createContext(k));return function(x){const C=(x==null?void 0:x[s])||h;return O.useMemo(()=>({[`__scope${s}`]:{...x,[s]:C}}),[x,C])}};return m.scopeName=s,[f,fp(m,...a)]}function fp(...s){const a=s[0];if(s.length===1)return a;const u=()=>{const f=s.map(m=>({useScope:m(),scopeName:m.scopeName}));return function(h){const k=f.reduce((x,{useScope:C,scopeName:j})=>{const A=C(h)[`__scope${j}`];return{...x,...A}},{});return O.useMemo(()=>({[`__scope${a.scopeName}`]:k}),[k])}};return u.scopeName=a.scopeName,u}function vc(s,a){if(typeof s=="function")return s(a);s!=null&&(s.current=a)}function Cc(...s){return a=>{let u=!1;const f=s.map(m=>{const h=vc(m,a);return!u&&typeof h=="function"&&(u=!0),h});if(u)return()=>{for(let m=0;m<f.length;m++){const h=f[m];typeof h=="function"?h():vc(s[m],null)}}}}function Lr(...s){return O.useCallback(Cc(...s),s)}var Or=O.forwardRef((s,a)=>{const{children:u,...f}=s,m=O.Children.toArray(u),h=m.find(mp);if(h){const k=h.props.children,x=m.map(C=>C===h?O.Children.count(k)>1?O.Children.only(null):O.isValidElement(k)?k.props.children:null:C);return S.jsx(bi,{...f,ref:a,children:O.isValidElement(k)?O.cloneElement(k,void 0,x):null})}return S.jsx(bi,{...f,ref:a,children:u})});Or.displayName="Slot";var bi=O.forwardRef((s,a)=>{const{children:u,...f}=s;if(O.isValidElement(u)){const m=vp(u);return O.cloneElement(u,{...hp(f,u.props),ref:a?Cc(a,m):m})}return O.Children.count(u)>1?O.Children.only(null):null});bi.displayName="SlotClone";var pp=({children:s})=>S.jsx(S.Fragment,{children:s});function mp(s){return O.isValidElement(s)&&s.type===pp}function hp(s,a){const u={...a};for(const f in a){const m=s[f],h=a[f];/^on[A-Z]/.test(f)?m&&h?u[f]=(...x)=>{h(...x),m(...x)}:m&&(u[f]=m):f==="style"?u[f]={...m,...h}:f==="className"&&(u[f]=[m,h].filter(Boolean).join(" "))}return{...s,...u}}function vp(s){var f,m;let a=(f=Object.getOwnPropertyDescriptor(s.props,"ref"))==null?void 0:f.get,u=a&&"isReactWarning"in a&&a.isReactWarning;return u?s.ref:(a=(m=Object.getOwnPropertyDescriptor(s,"ref"))==null?void 0:m.get,u=a&&"isReactWarning"in a&&a.isReactWarning,u?s.props.ref:s.props.ref||s.ref)}function gp(s){const a=s+"CollectionProvider",[u,f]=rs(a),[m,h]=u(a,{collectionRef:{current:null},itemMap:new Map}),k=Q=>{const{scope:z,children:M}=Q,R=je.useRef(null),U=je.useRef(new Map).current;return S.jsx(m,{scope:z,itemMap:U,collectionRef:R,children:M})};k.displayName=a;const x=s+"CollectionSlot",C=je.forwardRef((Q,z)=>{const{scope:M,children:R}=Q,U=h(x,M),L=Lr(z,U.collectionRef);return S.jsx(Or,{ref:L,children:R})});C.displayName=x;const j=s+"CollectionItemSlot",I="data-radix-collection-item",A=je.forwardRef((Q,z)=>{const{scope:M,children:R,...U}=Q,L=je.useRef(null),b=Lr(z,L),J=h(j,M);return je.useEffect(()=>(J.itemMap.set(L,{ref:L,...U}),()=>void J.itemMap.delete(L))),S.jsx(Or,{[I]:"",ref:b,children:R})});A.displayName=j;function F(Q){const z=h(s+"CollectionConsumer",Q);return je.useCallback(()=>{const R=z.collectionRef.current;if(!R)return[];const U=Array.from(R.querySelectorAll(`[${I}]`));return Array.from(z.itemMap.values()).sort((J,le)=>U.indexOf(J.ref.current)-U.indexOf(le.ref.current))},[z.collectionRef,z.itemMap])}return[{Provider:k,Slot:C,ItemSlot:A},F,f]}function Ec(s,a,{checkForDefaultPrevented:u=!0}={}){return function(m){if(s==null||s(m),u===!1||!m.defaultPrevented)return a==null?void 0:a(m)}}function Nc(s){const a=O.useRef(s);return O.useEffect(()=>{a.current=s}),O.useMemo(()=>(...u)=>{var f;return(f=a.current)==null?void 0:f.call(a,...u)},[])}function os({prop:s,defaultProp:a,onChange:u=()=>{}}){const[f,m]=yp({defaultProp:a,onChange:u}),h=s!==void 0,k=h?s:f,x=Nc(u),C=O.useCallback(j=>{if(h){const A=typeof j=="function"?j(s):j;A!==s&&x(A)}else m(j)},[h,s,m,x]);return[k,C]}function yp({defaultProp:s,onChange:a}){const u=O.useState(s),[f]=u,m=O.useRef(f),h=Nc(a);return O.useEffect(()=>{m.current!==f&&(h(f),m.current=f)},[f,m,h]),u}Sc();var wp=["a","button","div","form","h2","h3","img","input","label","li","nav","ol","p","span","svg","ul"],Ir=wp.reduce((s,a)=>{const u=O.forwardRef((f,m)=>{const{asChild:h,...k}=f,x=h?Or:a;return typeof window<"u"&&(window[Symbol.for("radix-ui")]=!0),S.jsx(x,{...k,ref:m})});return u.displayName=`Primitive.${a}`,{...s,[a]:u}},{}),Xo=globalThis!=null&&globalThis.document?O.useLayoutEffect:()=>{};function xp(s,a){return O.useReducer((u,f)=>a[u][f]??u,s)}var _c=s=>{const{present:a,children:u}=s,f=kp(a),m=typeof u=="function"?u({present:f.isPresent}):O.Children.only(u),h=Lr(f.ref,Sp(m));return typeof u=="function"||f.isPresent?O.cloneElement(m,{ref:h}):null};_c.displayName="Presence";function kp(s){const[a,u]=O.useState(),f=O.useRef({}),m=O.useRef(s),h=O.useRef("none"),k=s?"mounted":"unmounted",[x,C]=xp(k,{mounted:{UNMOUNT:"unmounted",ANIMATION_OUT:"unmountSuspended"},unmountSuspended:{MOUNT:"mounted",ANIMATION_END:"unmounted"},unmounted:{MOUNT:"mounted"}});return O.useEffect(()=>{const j=Yo(f.current);h.current=x==="mounted"?j:"none"},[x]),Xo(()=>{const j=f.current,I=m.current;if(I!==s){const F=h.current,Q=Yo(j);s?C("MOUNT"):Q==="none"||(j==null?void 0:j.display)==="none"?C("UNMOUNT"):C(I&&F!==Q?"ANIMATION_OUT":"UNMOUNT"),m.current=s}},[s,C]),Xo(()=>{if(a){let j;const I=a.ownerDocument.defaultView??window,A=Q=>{const M=Yo(f.current).includes(Q.animationName);if(Q.target===a&&M&&(C("ANIMATION_END"),!m.current)){const R=a.style.animationFillMode;a.style.animationFillMode="forwards",j=I.setTimeout(()=>{a.style.animationFillMode==="forwards"&&(a.style.animationFillMode=R)})}},F=Q=>{Q.target===a&&(h.current=Yo(f.current))};return a.addEventListener("animationstart",F),a.addEventListener("animationcancel",A),a.addEventListener("animationend",A),()=>{I.clearTimeout(j),a.removeEventListener("animationstart",F),a.removeEventListener("animationcancel",A),a.removeEventListener("animationend",A)}}else C("ANIMATION_END")},[a,C]),{isPresent:["mounted","unmountSuspended"].includes(x),ref:O.useCallback(j=>{j&&(f.current=getComputedStyle(j)),u(j)},[])}}function Yo(s){return(s==null?void 0:s.animationName)||"none"}function Sp(s){var f,m;let a=(f=Object.getOwnPropertyDescriptor(s.props,"ref"))==null?void 0:f.get,u=a&&"isReactWarning"in a&&a.isReactWarning;return u?s.ref:(a=(m=Object.getOwnPropertyDescriptor(s,"ref"))==null?void 0:m.get,u=a&&"isReactWarning"in a&&a.isReactWarning,u?s.props.ref:s.props.ref||s.ref)}var Cp=ip.useId||(()=>{}),Ep=0;function Pc(s){const[a,u]=O.useState(Cp());return Xo(()=>{u(f=>f??String(Ep++))},[s]),a?`radix-${a}`:""}var ls="Collapsible",[Np,jc]=rs(ls),[_p,is]=Np(ls),Rc=O.forwardRef((s,a)=>{const{__scopeCollapsible:u,open:f,defaultOpen:m,disabled:h,onOpenChange:k,...x}=s,[C=!1,j]=os({prop:f,defaultProp:m,onChange:k});return S.jsx(_p,{scope:u,disabled:h,contentId:Pc(),open:C,onOpenToggle:O.useCallback(()=>j(I=>!I),[j]),children:S.jsx(Ir.div,{"data-state":us(C),"data-disabled":h?"":void 0,...x,ref:a})})});Rc.displayName=ls;var zc="CollapsibleTrigger",Tc=O.forwardRef((s,a)=>{const{__scopeCollapsible:u,...f}=s,m=is(zc,u);return S.jsx(Ir.button,{type:"button","aria-controls":m.contentId,"aria-expanded":m.open||!1,"data-state":us(m.open),"data-disabled":m.disabled?"":void 0,disabled:m.disabled,...f,ref:a,onClick:Ec(s.onClick,m.onOpenToggle)})});Tc.displayName=zc;var ss="CollapsibleContent",Mc=O.forwardRef((s,a)=>{const{forceMount:u,...f}=s,m=is(ss,s.__scopeCollapsible);return S.jsx(_c,{present:u||m.open,children:({present:h})=>S.jsx(Pp,{...f,ref:a,present:h})})});Mc.displayName=ss;var Pp=O.forwardRef((s,a)=>{const{__scopeCollapsible:u,present:f,children:m,...h}=s,k=is(ss,u),[x,C]=O.useState(f),j=O.useRef(null),I=Lr(a,j),A=O.useRef(0),F=A.current,Q=O.useRef(0),z=Q.current,M=k.open||x,R=O.useRef(M),U=O.useRef(void 0);return O.useEffect(()=>{const L=requestAnimationFrame(()=>R.current=!1);return()=>cancelAnimationFrame(L)},[]),Xo(()=>{const L=j.current;if(L){U.current=U.current||{transitionDuration:L.style.transitionDuration,animationName:L.style.animationName},L.style.transitionDuration="0s",L.style.animationName="none";const b=L.getBoundingClientRect();A.current=b.height,Q.current=b.width,R.current||(L.style.transitionDuration=U.current.transitionDuration,L.style.animationName=U.current.animationName),C(f)}},[k.open,f]),S.jsx(Ir.div,{"data-state":us(k.open),"data-disabled":k.disabled?"":void 0,id:k.contentId,hidden:!M,...h,ref:I,style:{"--radix-collapsible-content-height":F?`${F}px`:void 0,"--radix-collapsible-content-width":z?`${z}px`:void 0,...s.style},children:M&&m})});function us(s){return s?"open":"closed"}var jp=Rc,Rp=Tc,zp=Mc,Tp=O.createContext(void 0);function Mp(s){const a=O.useContext(Tp);return s||a||"ltr"}var It="Accordion",Lp=["Home","End","ArrowDown","ArrowUp","ArrowLeft","ArrowRight"],[as,Op,Ip]=gp(It),[Jo,Hm]=rs(It,[Ip,jc]),cs=jc(),Lc=je.forwardRef((s,a)=>{const{type:u,...f}=s,m=f,h=f;return S.jsx(as.Provider,{scope:s.__scopeAccordion,children:u==="multiple"?S.jsx(Up,{...h,ref:a}):S.jsx(Fp,{...m,ref:a})})});Lc.displayName=It;var[Oc,Ap]=Jo(It),[Ic,Dp]=Jo(It,{collapsible:!1}),Fp=je.forwardRef((s,a)=>{const{value:u,defaultValue:f,onValueChange:m=()=>{},collapsible:h=!1,...k}=s,[x,C]=os({prop:u,defaultProp:f,onChange:m});return S.jsx(Oc,{scope:s.__scopeAccordion,value:x?[x]:[],onItemOpen:C,onItemClose:je.useCallback(()=>h&&C(""),[h,C]),children:S.jsx(Ic,{scope:s.__scopeAccordion,collapsible:h,children:S.jsx(Ac,{...k,ref:a})})})}),Up=je.forwardRef((s,a)=>{const{value:u,defaultValue:f,onValueChange:m=()=>{},...h}=s,[k=[],x]=os({prop:u,defaultProp:f,onChange:m}),C=je.useCallback(I=>x((A=[])=>[...A,I]),[x]),j=je.useCallback(I=>x((A=[])=>A.filter(F=>F!==I)),[x]);return S.jsx(Oc,{scope:s.__scopeAccordion,value:k,onItemOpen:C,onItemClose:j,children:S.jsx(Ic,{scope:s.__scopeAccordion,collapsible:!0,children:S.jsx(Ac,{...h,ref:a})})})}),[Vp,bo]=Jo(It),Ac=je.forwardRef((s,a)=>{const{__scopeAccordion:u,disabled:f,dir:m,orientation:h="vertical",...k}=s,x=je.useRef(null),C=Lr(x,a),j=Op(u),A=Mp(m)==="ltr",F=Ec(s.onKeyDown,Q=>{var _e;if(!Lp.includes(Q.key))return;const z=Q.target,M=j().filter(De=>{var Fe;return!((Fe=De.ref.current)!=null&&Fe.disabled)}),R=M.findIndex(De=>De.ref.current===z),U=M.length;if(R===-1)return;Q.preventDefault();let L=R;const b=0,J=U-1,le=()=>{L=R+1,L>J&&(L=b)},de=()=>{L=R-1,L<b&&(L=J)};switch(Q.key){case"Home":L=b;break;case"End":L=J;break;case"ArrowRight":h==="horizontal"&&(A?le():de());break;case"ArrowDown":h==="vertical"&&le();break;case"ArrowLeft":h==="horizontal"&&(A?de():le());break;case"ArrowUp":h==="vertical"&&de();break}const xe=L%U;(_e=M[xe].ref.current)==null||_e.focus()});return S.jsx(Vp,{scope:u,disabled:f,direction:m,orientation:h,children:S.jsx(as.Slot,{scope:u,children:S.jsx(Ir.div,{...k,"data-orientation":h,ref:C,onKeyDown:f?void 0:F})})})}),Zo="AccordionItem",[$p,ds]=Jo(Zo),Dc=je.forwardRef((s,a)=>{const{__scopeAccordion:u,value:f,...m}=s,h=bo(Zo,u),k=Ap(Zo,u),x=cs(u),C=Pc(),j=f&&k.value.includes(f)||!1,I=h.disabled||s.disabled;return S.jsx($p,{scope:u,open:j,disabled:I,triggerId:C,children:S.jsx(jp,{"data-orientation":h.orientation,"data-state":Bc(j),...x,...m,ref:a,disabled:I,open:j,onOpenChange:A=>{A?k.onItemOpen(f):k.onItemClose(f)}})})});Dc.displayName=Zo;var Fc="AccordionHeader",Uc=je.forwardRef((s,a)=>{const{__scopeAccordion:u,...f}=s,m=bo(It,u),h=ds(Fc,u);return S.jsx(Ir.h3,{"data-orientation":m.orientation,"data-state":Bc(h.open),"data-disabled":h.disabled?"":void 0,...f,ref:a})});Uc.displayName=Fc;var qi="AccordionTrigger",Vc=je.forwardRef((s,a)=>{const{__scopeAccordion:u,...f}=s,m=bo(It,u),h=ds(qi,u),k=Dp(qi,u),x=cs(u);return S.jsx(as.ItemSlot,{scope:u,children:S.jsx(Rp,{"aria-disabled":h.open&&!k.collapsible||void 0,"data-orientation":m.orientation,id:h.triggerId,...x,...f,ref:a})})});Vc.displayName=qi;var $c="AccordionContent",Wc=je.forwardRef((s,a)=>{const{__scopeAccordion:u,...f}=s,m=bo(It,u),h=ds($c,u),k=cs(u);return S.jsx(zp,{role:"region","aria-labelledby":h.triggerId,"data-orientation":m.orientation,...k,...f,ref:a,style:{"--radix-accordion-content-height":"var(--radix-collapsible-content-height)","--radix-accordion-content-width":"var(--radix-collapsible-content-width)",...s.style}})});Wc.displayName=$c;function Bc(s){return s?"open":"closed"}var Wp=Lc,Bp=Dc,Hp=Uc,Hc=Vc,Qc=Wc;/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const Qp=s=>s.replace(/([a-z0-9])([A-Z])/g,"$1-$2").toLowerCase(),Gc=(...s)=>s.filter((a,u,f)=>!!a&&a.trim()!==""&&f.indexOf(a)===u).join(" ").trim();/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */var Gp={xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:2,strokeLinecap:"round",strokeLinejoin:"round"};/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const Kp=O.forwardRef(({color:s="currentColor",size:a=24,strokeWidth:u=2,absoluteStrokeWidth:f,className:m="",children:h,iconNode:k,...x},C)=>O.createElement("svg",{ref:C,...Gp,width:a,height:a,stroke:s,strokeWidth:f?Number(u)*24/Number(a):u,className:Gc("lucide",m),...x},[...k.map(([j,I])=>O.createElement(j,I)),...Array.isArray(h)?h:[h]]));/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const dt=(s,a)=>{const u=O.forwardRef(({className:f,...m},h)=>O.createElement(Kp,{ref:h,iconNode:a,className:Gc(`lucide-${Qp(s)}`,f),...m}));return u.displayName=`${s}`,u};/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const Kc=dt("Check",[["path",{d:"M20 6 9 17l-5-5",key:"1gmf2c"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const Yp=dt("ChevronDown",[["path",{d:"m6 9 6 6 6-6",key:"qrunsl"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const Xp=dt("Download",[["path",{d:"M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4",key:"ih7n3h"}],["polyline",{points:"7 10 12 15 17 10",key:"2ggqvy"}],["line",{x1:"12",x2:"12",y1:"15",y2:"3",key:"1vk2je"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const Zp=dt("EyeOff",[["path",{d:"M10.733 5.076a10.744 10.744 0 0 1 11.205 6.575 1 1 0 0 1 0 .696 10.747 10.747 0 0 1-1.444 2.49",key:"ct8e1f"}],["path",{d:"M14.084 14.158a3 3 0 0 1-4.242-4.242",key:"151rxh"}],["path",{d:"M17.479 17.499a10.75 10.75 0 0 1-15.417-5.151 1 1 0 0 1 0-.696 10.75 10.75 0 0 1 4.446-5.143",key:"13bj9a"}],["path",{d:"m2 2 20 20",key:"1ooewy"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const Jp=dt("Eye",[["path",{d:"M2.062 12.348a1 1 0 0 1 0-.696 10.75 10.75 0 0 1 19.876 0 1 1 0 0 1 0 .696 10.75 10.75 0 0 1-19.876 0",key:"1nclc0"}],["circle",{cx:"12",cy:"12",r:"3",key:"1v7zrd"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const bp=dt("Moon",[["path",{d:"M12 3a6 6 0 0 0 9 9 9 9 0 1 1-9-9Z",key:"a7tn18"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const qp=dt("Pencil",[["path",{d:"M21.174 6.812a1 1 0 0 0-3.986-3.987L3.842 16.174a2 2 0 0 0-.5.83l-1.321 4.352a.5.5 0 0 0 .623.622l4.353-1.32a2 2 0 0 0 .83-.497z",key:"1a8usu"}],["path",{d:"m15 5 4 4",key:"1mk7zo"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const em=dt("Plus",[["path",{d:"M5 12h14",key:"1ays0h"}],["path",{d:"M12 5v14",key:"s699le"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const gc=dt("Settings",[["path",{d:"M12.22 2h-.44a2 2 0 0 0-2 2v.18a2 2 0 0 1-1 1.73l-.43.25a2 2 0 0 1-2 0l-.15-.08a2 2 0 0 0-2.73.73l-.22.38a2 2 0 0 0 .73 2.73l.15.1a2 2 0 0 1 1 1.72v.51a2 2 0 0 1-1 1.74l-.15.09a2 2 0 0 0-.73 2.73l.22.38a2 2 0 0 0 2.73.73l.15-.08a2 2 0 0 1 2 0l.43.25a2 2 0 0 1 1 1.73V20a2 2 0 0 0 2 2h.44a2 2 0 0 0 2-2v-.18a2 2 0 0 1 1-1.73l.43-.25a2 2 0 0 1 2 0l.15.08a2 2 0 0 0 2.73-.73l.22-.39a2 2 0 0 0-.73-2.73l-.15-.08a2 2 0 0 1-1-1.74v-.5a2 2 0 0 1 1-1.74l.15-.09a2 2 0 0 0 .73-2.73l-.22-.38a2 2 0 0 0-2.73-.73l-.15.08a2 2 0 0 1-2 0l-.43-.25a2 2 0 0 1-1-1.73V4a2 2 0 0 0-2-2z",key:"1qme2f"}],["circle",{cx:"12",cy:"12",r:"3",key:"1v7zrd"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const tm=dt("Sun",[["circle",{cx:"12",cy:"12",r:"4",key:"4exip2"}],["path",{d:"M12 2v2",key:"tus03m"}],["path",{d:"M12 20v2",key:"1lh1kg"}],["path",{d:"m4.93 4.93 1.41 1.41",key:"149t6j"}],["path",{d:"m17.66 17.66 1.41 1.41",key:"ptbguv"}],["path",{d:"M2 12h2",key:"1t8f8n"}],["path",{d:"M20 12h2",key:"1q8mjw"}],["path",{d:"m6.34 17.66-1.41 1.41",key:"1m8zz5"}],["path",{d:"m19.07 4.93-1.41 1.41",key:"1shlcs"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const nm=dt("Trash2",[["path",{d:"M3 6h18",key:"d0wm0j"}],["path",{d:"M19 6v14c0 1-1 2-2 2H7c-1 0-2-1-2-2V6",key:"4alrt4"}],["path",{d:"M8 6V4c0-1 1-2 2-2h4c1 0 2 1 2 2v2",key:"v07s0e"}],["line",{x1:"10",x2:"10",y1:"11",y2:"17",key:"1uufr5"}],["line",{x1:"14",x2:"14",y1:"11",y2:"17",key:"xtxkd"}]]);/**
- * @license lucide-react v0.469.0 - ISC
- *
- * This source code is licensed under the ISC license.
- * See the LICENSE file in the root directory of this source tree.
- */const Yc=dt("X",[["path",{d:"M18 6 6 18",key:"1bl5f8"}],["path",{d:"m6 6 12 12",key:"d8bk6v"}]]);function Xc(s){var a,u,f="";if(typeof s=="string"||typeof s=="number")f+=s;else if(typeof s=="object")if(Array.isArray(s)){var m=s.length;for(a=0;a<m;a++)s[a]&&(u=Xc(s[a]))&&(f&&(f+=" "),f+=u)}else for(u in s)s[u]&&(f&&(f+=" "),f+=u);return f}function Zc(){for(var s,a,u=0,f="",m=arguments.length;u<m;u++)(s=arguments[u])&&(a=Xc(s))&&(f&&(f+=" "),f+=a);return f}const fs="-",rm=s=>{const a=lm(s),{conflictingClassGroups:u,conflictingClassGroupModifiers:f}=s;return{getClassGroupId:k=>{const x=k.split(fs);return x[0]===""&&x.length!==1&&x.shift(),Jc(x,a)||om(k)},getConflictingClassGroupIds:(k,x)=>{const C=u[k]||[];return x&&f[k]?[...C,...f[k]]:C}}},Jc=(s,a)=>{var k;if(s.length===0)return a.classGroupId;const u=s[0],f=a.nextPart.get(u),m=f?Jc(s.slice(1),f):void 0;if(m)return m;if(a.validators.length===0)return;const h=s.join(fs);return(k=a.validators.find(({validator:x})=>x(h)))==null?void 0:k.classGroupId},yc=/^\[(.+)\]$/,om=s=>{if(yc.test(s)){const a=yc.exec(s)[1],u=a==null?void 0:a.substring(0,a.indexOf(":"));if(u)return"arbitrary.."+u}},lm=s=>{const{theme:a,prefix:u}=s,f={nextPart:new Map,validators:[]};return sm(Object.entries(s.classGroups),u).forEach(([h,k])=>{es(k,f,h,a)}),f},es=(s,a,u,f)=>{s.forEach(m=>{if(typeof m=="string"){const h=m===""?a:wc(a,m);h.classGroupId=u;return}if(typeof m=="function"){if(im(m)){es(m(f),a,u,f);return}a.validators.push({validator:m,classGroupId:u});return}Object.entries(m).forEach(([h,k])=>{es(k,wc(a,h),u,f)})})},wc=(s,a)=>{let u=s;return a.split(fs).forEach(f=>{u.nextPart.has(f)||u.nextPart.set(f,{nextPart:new Map,validators:[]}),u=u.nextPart.get(f)}),u},im=s=>s.isThemeGetter,sm=(s,a)=>a?s.map(([u,f])=>{const m=f.map(h=>typeof h=="string"?a+h:typeof h=="object"?Object.fromEntries(Object.entries(h).map(([k,x])=>[a+k,x])):h);return[u,m]}):s,um=s=>{if(s<1)return{get:()=>{},set:()=>{}};let a=0,u=new Map,f=new Map;const m=(h,k)=>{u.set(h,k),a++,a>s&&(a=0,f=u,u=new Map)};return{get(h){let k=u.get(h);if(k!==void 0)return k;if((k=f.get(h))!==void 0)return m(h,k),k},set(h,k){u.has(h)?u.set(h,k):m(h,k)}}},bc="!",am=s=>{const{separator:a,experimentalParseClassName:u}=s,f=a.length===1,m=a[0],h=a.length,k=x=>{const C=[];let j=0,I=0,A;for(let R=0;R<x.length;R++){let U=x[R];if(j===0){if(U===m&&(f||x.slice(R,R+h)===a)){C.push(x.slice(I,R)),I=R+h;continue}if(U==="/"){A=R;continue}}U==="["?j++:U==="]"&&j--}const F=C.length===0?x:x.substring(I),Q=F.startsWith(bc),z=Q?F.substring(1):F,M=A&&A>I?A-I:void 0;return{modifiers:C,hasImportantModifier:Q,baseClassName:z,maybePostfixModifierPosition:M}};return u?x=>u({className:x,parseClassName:k}):k},cm=s=>{if(s.length<=1)return s;const a=[];let u=[];return s.forEach(f=>{f[0]==="["?(a.push(...u.sort(),f),u=[]):u.push(f)}),a.push(...u.sort()),a},dm=s=>({cache:um(s.cacheSize),parseClassName:am(s),...rm(s)}),fm=/\s+/,pm=(s,a)=>{const{parseClassName:u,getClassGroupId:f,getConflictingClassGroupIds:m}=a,h=[],k=s.trim().split(fm);let x="";for(let C=k.length-1;C>=0;C-=1){const j=k[C],{modifiers:I,hasImportantModifier:A,baseClassName:F,maybePostfixModifierPosition:Q}=u(j);let z=!!Q,M=f(z?F.substring(0,Q):F);if(!M){if(!z){x=j+(x.length>0?" "+x:x);continue}if(M=f(F),!M){x=j+(x.length>0?" "+x:x);continue}z=!1}const R=cm(I).join(":"),U=A?R+bc:R,L=U+M;if(h.includes(L))continue;h.push(L);const b=m(M,z);for(let J=0;J<b.length;++J){const le=b[J];h.push(U+le)}x=j+(x.length>0?" "+x:x)}return x};function mm(){let s=0,a,u,f="";for(;s<arguments.length;)(a=arguments[s++])&&(u=qc(a))&&(f&&(f+=" "),f+=u);return f}const qc=s=>{if(typeof s=="string")return s;let a,u="";for(let f=0;f<s.length;f++)s[f]&&(a=qc(s[f]))&&(u&&(u+=" "),u+=a);return u};function hm(s,...a){let u,f,m,h=k;function k(C){const j=a.reduce((I,A)=>A(I),s());return u=dm(j),f=u.cache.get,m=u.cache.set,h=x,x(C)}function x(C){const j=f(C);if(j)return j;const I=pm(C,u);return m(C,I),I}return function(){return h(mm.apply(null,arguments))}}const ge=s=>{const a=u=>u[s]||[];return a.isThemeGetter=!0,a},ed=/^\[(?:([a-z-]+):)?(.+)\]$/i,vm=/^\d+\/\d+$/,gm=new Set(["px","full","screen"]),ym=/^(\d+(\.\d+)?)?(xs|sm|md|lg|xl)$/,wm=/\d+(%|px|r?em|[sdl]?v([hwib]|min|max)|pt|pc|in|cm|mm|cap|ch|ex|r?lh|cq(w|h|i|b|min|max))|\b(calc|min|max|clamp)\(.+\)|^0$/,xm=/^(rgba?|hsla?|hwb|(ok)?(lab|lch))\(.+\)$/,km=/^(inset_)?-?((\d+)?\.?(\d+)[a-z]+|0)_-?((\d+)?\.?(\d+)[a-z]+|0)/,Sm=/^(url|image|image-set|cross-fade|element|(repeating-)?(linear|radial|conic)-gradient)\(.+\)$/,Lt=s=>Wn(s)||gm.has(s)||vm.test(s),tn=s=>Bn(s,"length",zm),Wn=s=>!!s&&!Number.isNaN(Number(s)),Ji=s=>Bn(s,"number",Wn),zr=s=>!!s&&Number.isInteger(Number(s)),Cm=s=>s.endsWith("%")&&Wn(s.slice(0,-1)),q=s=>ed.test(s),nn=s=>ym.test(s),Em=new Set(["length","size","percentage"]),Nm=s=>Bn(s,Em,td),_m=s=>Bn(s,"position",td),Pm=new Set(["image","url"]),jm=s=>Bn(s,Pm,Mm),Rm=s=>Bn(s,"",Tm),Tr=()=>!0,Bn=(s,a,u)=>{const f=ed.exec(s);return f?f[1]?typeof a=="string"?f[1]===a:a.has(f[1]):u(f[2]):!1},zm=s=>wm.test(s)&&!xm.test(s),td=()=>!1,Tm=s=>km.test(s),Mm=s=>Sm.test(s),Lm=()=>{const s=ge("colors"),a=ge("spacing"),u=ge("blur"),f=ge("brightness"),m=ge("borderColor"),h=ge("borderRadius"),k=ge("borderSpacing"),x=ge("borderWidth"),C=ge("contrast"),j=ge("grayscale"),I=ge("hueRotate"),A=ge("invert"),F=ge("gap"),Q=ge("gradientColorStops"),z=ge("gradientColorStopPositions"),M=ge("inset"),R=ge("margin"),U=ge("opacity"),L=ge("padding"),b=ge("saturate"),J=ge("scale"),le=ge("sepia"),de=ge("skew"),xe=ge("space"),_e=ge("translate"),De=()=>["auto","contain","none"],Fe=()=>["auto","hidden","clip","visible","scroll"],ft=()=>["auto",q,a],re=()=>[q,a],Ue=()=>["",Lt,tn],Ve=()=>["auto",Wn,q],qe=()=>["bottom","center","left","left-bottom","left-top","right","right-bottom","right-top","top"],Re=()=>["solid","dashed","dotted","double","none"],me=()=>["normal","multiply","screen","overlay","darken","lighten","color-dodge","color-burn","hard-light","soft-light","difference","exclusion","hue","saturation","color","luminosity"],D=()=>["start","end","center","between","around","evenly","stretch"],K=()=>["","0",q],$=()=>["auto","avoid","all","avoid-page","page","left","right","column"],v=()=>[Wn,q];return{cacheSize:500,separator:":",theme:{colors:[Tr],spacing:[Lt,tn],blur:["none","",nn,q],brightness:v(),borderColor:[s],borderRadius:["none","","full",nn,q],borderSpacing:re(),borderWidth:Ue(),contrast:v(),grayscale:K(),hueRotate:v(),invert:K(),gap:re(),gradientColorStops:[s],gradientColorStopPositions:[Cm,tn],inset:ft(),margin:ft(),opacity:v(),padding:re(),saturate:v(),scale:v(),sepia:K(),skew:v(),space:re(),translate:re()},classGroups:{aspect:[{aspect:["auto","square","video",q]}],container:["container"],columns:[{columns:[nn]}],"break-after":[{"break-after":$()}],"break-before":[{"break-before":$()}],"break-inside":[{"break-inside":["auto","avoid","avoid-page","avoid-column"]}],"box-decoration":[{"box-decoration":["slice","clone"]}],box:[{box:["border","content"]}],display:["block","inline-block","inline","flex","inline-flex","table","inline-table","table-caption","table-cell","table-column","table-column-group","table-footer-group","table-header-group","table-row-group","table-row","flow-root","grid","inline-grid","contents","list-item","hidden"],float:[{float:["right","left","none","start","end"]}],clear:[{clear:["left","right","both","none","start","end"]}],isolation:["isolate","isolation-auto"],"object-fit":[{object:["contain","cover","fill","none","scale-down"]}],"object-position":[{object:[...qe(),q]}],overflow:[{overflow:Fe()}],"overflow-x":[{"overflow-x":Fe()}],"overflow-y":[{"overflow-y":Fe()}],overscroll:[{overscroll:De()}],"overscroll-x":[{"overscroll-x":De()}],"overscroll-y":[{"overscroll-y":De()}],position:["static","fixed","absolute","relative","sticky"],inset:[{inset:[M]}],"inset-x":[{"inset-x":[M]}],"inset-y":[{"inset-y":[M]}],start:[{start:[M]}],end:[{end:[M]}],top:[{top:[M]}],right:[{right:[M]}],bottom:[{bottom:[M]}],left:[{left:[M]}],visibility:["visible","invisible","collapse"],z:[{z:["auto",zr,q]}],basis:[{basis:ft()}],"flex-direction":[{flex:["row","row-reverse","col","col-reverse"]}],"flex-wrap":[{flex:["wrap","wrap-reverse","nowrap"]}],flex:[{flex:["1","auto","initial","none",q]}],grow:[{grow:K()}],shrink:[{shrink:K()}],order:[{order:["first","last","none",zr,q]}],"grid-cols":[{"grid-cols":[Tr]}],"col-start-end":[{col:["auto",{span:["full",zr,q]},q]}],"col-start":[{"col-start":Ve()}],"col-end":[{"col-end":Ve()}],"grid-rows":[{"grid-rows":[Tr]}],"row-start-end":[{row:["auto",{span:[zr,q]},q]}],"row-start":[{"row-start":Ve()}],"row-end":[{"row-end":Ve()}],"grid-flow":[{"grid-flow":["row","col","dense","row-dense","col-dense"]}],"auto-cols":[{"auto-cols":["auto","min","max","fr",q]}],"auto-rows":[{"auto-rows":["auto","min","max","fr",q]}],gap:[{gap:[F]}],"gap-x":[{"gap-x":[F]}],"gap-y":[{"gap-y":[F]}],"justify-content":[{justify:["normal",...D()]}],"justify-items":[{"justify-items":["start","end","center","stretch"]}],"justify-self":[{"justify-self":["auto","start","end","center","stretch"]}],"align-content":[{content:["normal",...D(),"baseline"]}],"align-items":[{items:["start","end","center","baseline","stretch"]}],"align-self":[{self:["auto","start","end","center","stretch","baseline"]}],"place-content":[{"place-content":[...D(),"baseline"]}],"place-items":[{"place-items":["start","end","center","baseline","stretch"]}],"place-self":[{"place-self":["auto","start","end","center","stretch"]}],p:[{p:[L]}],px:[{px:[L]}],py:[{py:[L]}],ps:[{ps:[L]}],pe:[{pe:[L]}],pt:[{pt:[L]}],pr:[{pr:[L]}],pb:[{pb:[L]}],pl:[{pl:[L]}],m:[{m:[R]}],mx:[{mx:[R]}],my:[{my:[R]}],ms:[{ms:[R]}],me:[{me:[R]}],mt:[{mt:[R]}],mr:[{mr:[R]}],mb:[{mb:[R]}],ml:[{ml:[R]}],"space-x":[{"space-x":[xe]}],"space-x-reverse":["space-x-reverse"],"space-y":[{"space-y":[xe]}],"space-y-reverse":["space-y-reverse"],w:[{w:["auto","min","max","fit","svw","lvw","dvw",q,a]}],"min-w":[{"min-w":[q,a,"min","max","fit"]}],"max-w":[{"max-w":[q,a,"none","full","min","max","fit","prose",{screen:[nn]},nn]}],h:[{h:[q,a,"auto","min","max","fit","svh","lvh","dvh"]}],"min-h":[{"min-h":[q,a,"min","max","fit","svh","lvh","dvh"]}],"max-h":[{"max-h":[q,a,"min","max","fit","svh","lvh","dvh"]}],size:[{size:[q,a,"auto","min","max","fit"]}],"font-size":[{text:["base",nn,tn]}],"font-smoothing":["antialiased","subpixel-antialiased"],"font-style":["italic","not-italic"],"font-weight":[{font:["thin","extralight","light","normal","medium","semibold","bold","extrabold","black",Ji]}],"font-family":[{font:[Tr]}],"fvn-normal":["normal-nums"],"fvn-ordinal":["ordinal"],"fvn-slashed-zero":["slashed-zero"],"fvn-figure":["lining-nums","oldstyle-nums"],"fvn-spacing":["proportional-nums","tabular-nums"],"fvn-fraction":["diagonal-fractions","stacked-fractions"],tracking:[{tracking:["tighter","tight","normal","wide","wider","widest",q]}],"line-clamp":[{"line-clamp":["none",Wn,Ji]}],leading:[{leading:["none","tight","snug","normal","relaxed","loose",Lt,q]}],"list-image":[{"list-image":["none",q]}],"list-style-type":[{list:["none","disc","decimal",q]}],"list-style-position":[{list:["inside","outside"]}],"placeholder-color":[{placeholder:[s]}],"placeholder-opacity":[{"placeholder-opacity":[U]}],"text-alignment":[{text:["left","center","right","justify","start","end"]}],"text-color":[{text:[s]}],"text-opacity":[{"text-opacity":[U]}],"text-decoration":["underline","overline","line-through","no-underline"],"text-decoration-style":[{decoration:[...Re(),"wavy"]}],"text-decoration-thickness":[{decoration:["auto","from-font",Lt,tn]}],"underline-offset":[{"underline-offset":["auto",Lt,q]}],"text-decoration-color":[{decoration:[s]}],"text-transform":["uppercase","lowercase","capitalize","normal-case"],"text-overflow":["truncate","text-ellipsis","text-clip"],"text-wrap":[{text:["wrap","nowrap","balance","pretty"]}],indent:[{indent:re()}],"vertical-align":[{align:["baseline","top","middle","bottom","text-top","text-bottom","sub","super",q]}],whitespace:[{whitespace:["normal","nowrap","pre","pre-line","pre-wrap","break-spaces"]}],break:[{break:["normal","words","all","keep"]}],hyphens:[{hyphens:["none","manual","auto"]}],content:[{content:["none",q]}],"bg-attachment":[{bg:["fixed","local","scroll"]}],"bg-clip":[{"bg-clip":["border","padding","content","text"]}],"bg-opacity":[{"bg-opacity":[U]}],"bg-origin":[{"bg-origin":["border","padding","content"]}],"bg-position":[{bg:[...qe(),_m]}],"bg-repeat":[{bg:["no-repeat",{repeat:["","x","y","round","space"]}]}],"bg-size":[{bg:["auto","cover","contain",Nm]}],"bg-image":[{bg:["none",{"gradient-to":["t","tr","r","br","b","bl","l","tl"]},jm]}],"bg-color":[{bg:[s]}],"gradient-from-pos":[{from:[z]}],"gradient-via-pos":[{via:[z]}],"gradient-to-pos":[{to:[z]}],"gradient-from":[{from:[Q]}],"gradient-via":[{via:[Q]}],"gradient-to":[{to:[Q]}],rounded:[{rounded:[h]}],"rounded-s":[{"rounded-s":[h]}],"rounded-e":[{"rounded-e":[h]}],"rounded-t":[{"rounded-t":[h]}],"rounded-r":[{"rounded-r":[h]}],"rounded-b":[{"rounded-b":[h]}],"rounded-l":[{"rounded-l":[h]}],"rounded-ss":[{"rounded-ss":[h]}],"rounded-se":[{"rounded-se":[h]}],"rounded-ee":[{"rounded-ee":[h]}],"rounded-es":[{"rounded-es":[h]}],"rounded-tl":[{"rounded-tl":[h]}],"rounded-tr":[{"rounded-tr":[h]}],"rounded-br":[{"rounded-br":[h]}],"rounded-bl":[{"rounded-bl":[h]}],"border-w":[{border:[x]}],"border-w-x":[{"border-x":[x]}],"border-w-y":[{"border-y":[x]}],"border-w-s":[{"border-s":[x]}],"border-w-e":[{"border-e":[x]}],"border-w-t":[{"border-t":[x]}],"border-w-r":[{"border-r":[x]}],"border-w-b":[{"border-b":[x]}],"border-w-l":[{"border-l":[x]}],"border-opacity":[{"border-opacity":[U]}],"border-style":[{border:[...Re(),"hidden"]}],"divide-x":[{"divide-x":[x]}],"divide-x-reverse":["divide-x-reverse"],"divide-y":[{"divide-y":[x]}],"divide-y-reverse":["divide-y-reverse"],"divide-opacity":[{"divide-opacity":[U]}],"divide-style":[{divide:Re()}],"border-color":[{border:[m]}],"border-color-x":[{"border-x":[m]}],"border-color-y":[{"border-y":[m]}],"border-color-s":[{"border-s":[m]}],"border-color-e":[{"border-e":[m]}],"border-color-t":[{"border-t":[m]}],"border-color-r":[{"border-r":[m]}],"border-color-b":[{"border-b":[m]}],"border-color-l":[{"border-l":[m]}],"divide-color":[{divide:[m]}],"outline-style":[{outline:["",...Re()]}],"outline-offset":[{"outline-offset":[Lt,q]}],"outline-w":[{outline:[Lt,tn]}],"outline-color":[{outline:[s]}],"ring-w":[{ring:Ue()}],"ring-w-inset":["ring-inset"],"ring-color":[{ring:[s]}],"ring-opacity":[{"ring-opacity":[U]}],"ring-offset-w":[{"ring-offset":[Lt,tn]}],"ring-offset-color":[{"ring-offset":[s]}],shadow:[{shadow:["","inner","none",nn,Rm]}],"shadow-color":[{shadow:[Tr]}],opacity:[{opacity:[U]}],"mix-blend":[{"mix-blend":[...me(),"plus-lighter","plus-darker"]}],"bg-blend":[{"bg-blend":me()}],filter:[{filter:["","none"]}],blur:[{blur:[u]}],brightness:[{brightness:[f]}],contrast:[{contrast:[C]}],"drop-shadow":[{"drop-shadow":["","none",nn,q]}],grayscale:[{grayscale:[j]}],"hue-rotate":[{"hue-rotate":[I]}],invert:[{invert:[A]}],saturate:[{saturate:[b]}],sepia:[{sepia:[le]}],"backdrop-filter":[{"backdrop-filter":["","none"]}],"backdrop-blur":[{"backdrop-blur":[u]}],"backdrop-brightness":[{"backdrop-brightness":[f]}],"backdrop-contrast":[{"backdrop-contrast":[C]}],"backdrop-grayscale":[{"backdrop-grayscale":[j]}],"backdrop-hue-rotate":[{"backdrop-hue-rotate":[I]}],"backdrop-invert":[{"backdrop-invert":[A]}],"backdrop-opacity":[{"backdrop-opacity":[U]}],"backdrop-saturate":[{"backdrop-saturate":[b]}],"backdrop-sepia":[{"backdrop-sepia":[le]}],"border-collapse":[{border:["collapse","separate"]}],"border-spacing":[{"border-spacing":[k]}],"border-spacing-x":[{"border-spacing-x":[k]}],"border-spacing-y":[{"border-spacing-y":[k]}],"table-layout":[{table:["auto","fixed"]}],caption:[{caption:["top","bottom"]}],transition:[{transition:["none","all","","colors","opacity","shadow","transform",q]}],duration:[{duration:v()}],ease:[{ease:["linear","in","out","in-out",q]}],delay:[{delay:v()}],animate:[{animate:["none","spin","ping","pulse","bounce",q]}],transform:[{transform:["","gpu","none"]}],scale:[{scale:[J]}],"scale-x":[{"scale-x":[J]}],"scale-y":[{"scale-y":[J]}],rotate:[{rotate:[zr,q]}],"translate-x":[{"translate-x":[_e]}],"translate-y":[{"translate-y":[_e]}],"skew-x":[{"skew-x":[de]}],"skew-y":[{"skew-y":[de]}],"transform-origin":[{origin:["center","top","top-right","right","bottom-right","bottom","bottom-left","left","top-left",q]}],accent:[{accent:["auto",s]}],appearance:[{appearance:["none","auto"]}],cursor:[{cursor:["auto","default","pointer","wait","text","move","help","not-allowed","none","context-menu","progress","cell","crosshair","vertical-text","alias","copy","no-drop","grab","grabbing","all-scroll","col-resize","row-resize","n-resize","e-resize","s-resize","w-resize","ne-resize","nw-resize","se-resize","sw-resize","ew-resize","ns-resize","nesw-resize","nwse-resize","zoom-in","zoom-out",q]}],"caret-color":[{caret:[s]}],"pointer-events":[{"pointer-events":["none","auto"]}],resize:[{resize:["none","y","x",""]}],"scroll-behavior":[{scroll:["auto","smooth"]}],"scroll-m":[{"scroll-m":re()}],"scroll-mx":[{"scroll-mx":re()}],"scroll-my":[{"scroll-my":re()}],"scroll-ms":[{"scroll-ms":re()}],"scroll-me":[{"scroll-me":re()}],"scroll-mt":[{"scroll-mt":re()}],"scroll-mr":[{"scroll-mr":re()}],"scroll-mb":[{"scroll-mb":re()}],"scroll-ml":[{"scroll-ml":re()}],"scroll-p":[{"scroll-p":re()}],"scroll-px":[{"scroll-px":re()}],"scroll-py":[{"scroll-py":re()}],"scroll-ps":[{"scroll-ps":re()}],"scroll-pe":[{"scroll-pe":re()}],"scroll-pt":[{"scroll-pt":re()}],"scroll-pr":[{"scroll-pr":re()}],"scroll-pb":[{"scroll-pb":re()}],"scroll-pl":[{"scroll-pl":re()}],"snap-align":[{snap:["start","end","center","align-none"]}],"snap-stop":[{snap:["normal","always"]}],"snap-type":[{snap:["none","x","y","both"]}],"snap-strictness":[{snap:["mandatory","proximity"]}],touch:[{touch:["auto","none","manipulation"]}],"touch-x":[{"touch-pan":["x","left","right"]}],"touch-y":[{"touch-pan":["y","up","down"]}],"touch-pz":["touch-pinch-zoom"],select:[{select:["none","text","all","auto"]}],"will-change":[{"will-change":["auto","scroll","contents","transform",q]}],fill:[{fill:[s,"none"]}],"stroke-w":[{stroke:[Lt,tn,Ji]}],stroke:[{stroke:[s,"none"]}],sr:["sr-only","not-sr-only"],"forced-color-adjust":[{"forced-color-adjust":["auto","none"]}]},conflictingClassGroups:{overflow:["overflow-x","overflow-y"],overscroll:["overscroll-x","overscroll-y"],inset:["inset-x","inset-y","start","end","top","right","bottom","left"],"inset-x":["right","left"],"inset-y":["top","bottom"],flex:["basis","grow","shrink"],gap:["gap-x","gap-y"],p:["px","py","ps","pe","pt","pr","pb","pl"],px:["pr","pl"],py:["pt","pb"],m:["mx","my","ms","me","mt","mr","mb","ml"],mx:["mr","ml"],my:["mt","mb"],size:["w","h"],"font-size":["leading"],"fvn-normal":["fvn-ordinal","fvn-slashed-zero","fvn-figure","fvn-spacing","fvn-fraction"],"fvn-ordinal":["fvn-normal"],"fvn-slashed-zero":["fvn-normal"],"fvn-figure":["fvn-normal"],"fvn-spacing":["fvn-normal"],"fvn-fraction":["fvn-normal"],"line-clamp":["display","overflow"],rounded:["rounded-s","rounded-e","rounded-t","rounded-r","rounded-b","rounded-l","rounded-ss","rounded-se","rounded-ee","rounded-es","rounded-tl","rounded-tr","rounded-br","rounded-bl"],"rounded-s":["rounded-ss","rounded-es"],"rounded-e":["rounded-se","rounded-ee"],"rounded-t":["rounded-tl","rounded-tr"],"rounded-r":["rounded-tr","rounded-br"],"rounded-b":["rounded-br","rounded-bl"],"rounded-l":["rounded-tl","rounded-bl"],"border-spacing":["border-spacing-x","border-spacing-y"],"border-w":["border-w-s","border-w-e","border-w-t","border-w-r","border-w-b","border-w-l"],"border-w-x":["border-w-r","border-w-l"],"border-w-y":["border-w-t","border-w-b"],"border-color":["border-color-s","border-color-e","border-color-t","border-color-r","border-color-b","border-color-l"],"border-color-x":["border-color-r","border-color-l"],"border-color-y":["border-color-t","border-color-b"],"scroll-m":["scroll-mx","scroll-my","scroll-ms","scroll-me","scroll-mt","scroll-mr","scroll-mb","scroll-ml"],"scroll-mx":["scroll-mr","scroll-ml"],"scroll-my":["scroll-mt","scroll-mb"],"scroll-p":["scroll-px","scroll-py","scroll-ps","scroll-pe","scroll-pt","scroll-pr","scroll-pb","scroll-pl"],"scroll-px":["scroll-pr","scroll-pl"],"scroll-py":["scroll-pt","scroll-pb"],touch:["touch-x","touch-y","touch-pz"],"touch-x":["touch"],"touch-y":["touch"],"touch-pz":["touch"]},conflictingClassGroupModifiers:{"font-size":["leading"]}}},Om=hm(Lm);function Ot(...s){return Om(Zc(s))}const Im=Wp,nd=O.forwardRef(({className:s,...a},u)=>S.jsx(Bp,{ref:u,className:Ot("w-full",s),...a}));nd.displayName="AccordionItem";const rd=O.forwardRef(({className:s,children:a,...u},f)=>S.jsx(Hp,{className:"flex",children:S.jsxs(Hc,{ref:f,className:Ot("flex flex-1 items-center justify-between transition-all","[&[data-state=open]>svg]:rotate-180",s),...u,children:[a,S.jsx(Yp,{className:"h-4 w-4 shrink-0 transition-transform duration-200 text-muted-foreground"})]})}));rd.displayName=Hc.displayName;const od=O.forwardRef(({className:s,children:a,...u},f)=>S.jsx(Qc,{ref:f,className:"overflow-hidden text-sm transition-all data-[state=closed]:animate-accordion-up data-[state=open]:animate-accordion-down",...u,children:S.jsx("div",{className:Ot(s),children:a})}));od.displayName=Qc.displayName;const xc=s=>typeof s=="boolean"?`${s}`:s===0?"0":s,kc=Zc,Am=(s,a)=>u=>{var f;if((a==null?void 0:a.variants)==null)return kc(s,u==null?void 0:u.class,u==null?void 0:u.className);const{variants:m,defaultVariants:h}=a,k=Object.keys(m).map(j=>{const I=u==null?void 0:u[j],A=h==null?void 0:h[j];if(I===null)return null;const F=xc(I)||xc(A);return m[j][F]}),x=u&&Object.entries(u).reduce((j,I)=>{let[A,F]=I;return F===void 0||(j[A]=F),j},{}),C=a==null||(f=a.compoundVariants)===null||f===void 0?void 0:f.reduce((j,I)=>{let{class:A,className:F,...Q}=I;return Object.entries(Q).every(z=>{let[M,R]=z;return Array.isArray(R)?R.includes({...h,...x}[M]):{...h,...x}[M]===R})?[...j,A,F]:j},[]);return kc(s,k,C,u==null?void 0:u.class,u==null?void 0:u.className)},Dm=Am("inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50",{variants:{variant:{default:"bg-primary text-primary-foreground shadow hover:bg-primary/90",destructive:"bg-destructive text-destructive-foreground shadow-sm hover:bg-destructive/90",outline:"border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground",secondary:"bg-secondary text-secondary-foreground shadow-sm hover:bg-secondary/80",ghost:"hover:bg-accent hover:text-accent-foreground",link:"text-primary underline-offset-4 hover:underline"},size:{default:"h-9 px-4 py-2",sm:"h-8 rounded-md px-3 text-xs",lg:"h-10 rounded-md px-8",icon:"h-9 w-9"}},defaultVariants:{variant:"default",size:"default"}}),ts=O.forwardRef(({className:s,variant:a,size:u,asChild:f=!1,...m},h)=>{const k=f?Or:"button";return S.jsx(k,{className:Ot(Dm({variant:a,size:u,className:s})),ref:h,...m})});ts.displayName="Button";const Fm=({isDark:s,onThemeToggle:a,skills:u})=>{const f=()=>{const m=JSON.stringify(u,null,2),h=new Blob([m],{type:"application/json"}),k=URL.createObjectURL(h),x=document.createElement("a");x.href=k,x.download="skill-settings.json",document.body.appendChild(x),x.click(),document.body.removeChild(x),URL.revokeObjectURL(k)};return S.jsx("div",{className:"border-b",children:S.jsx("div",{className:"max-w-7xl mx-auto p-4 md:p-6",children:S.jsxs("div",{className:"flex justify-between items-center",children:[S.jsxs("div",{children:[S.jsx("h1",{className:"text-3xl font-bold text-primary",children:"Skill Settings"}),S.jsx("p",{className:"text-sm text-muted-foreground mt-1",children:"Configure your voice assistant skills"})]}),S.jsxs("div",{className:"flex items-center gap-2",children:[S.jsx(ts,{variant:"outline",size:"icon",onClick:f,className:"h-9 w-9",title:"Export settings",children:S.jsx(Xp,{className:"h-4 w-4"})}),S.jsx(ts,{variant:"outline",size:"icon",onClick:a,className:"h-9 w-9",title:s?"Switch to light mode":"Switch to dark mode",children:s?S.jsx(tm,{className:"h-4 w-4"}):S.jsx(bp,{className:"h-4 w-4"})})]})]})})})},Um=({settingKey:s,value:a,onSave:u,onDelete:f})=>{const[m,h]=O.useState(!1),[k,x]=O.useState(a),[C]=O.useState(a),[j,I]=O.useState(!1),F=(L=>typeof L=="boolean"?"boolean":typeof L=="number"?"number":typeof L=="string"&&L.includes(`
-`)?"textarea":"text")(a),Q=async()=>{I(!0);try{let L=k;F==="number"&&(L=Number(k)),await u(s,L),h(!1)}catch(L){console.error("Failed to save:",L),x(C)}finally{I(!1)}},z=()=>{x(C),h(!1)},M=async()=>{if(window.confirm(`Are you sure you want to delete ${s}?`)){I(!0);try{await f(s)}catch(L){console.error("Failed to delete:",L)}finally{I(!1)}}},R=()=>{switch(F){case"boolean":return S.jsxs("select",{value:String(k),onChange:L=>x(L.target.value==="true"),className:"w-24 rounded border bg-background p-1 text-sm",autoFocus:!0,children:[S.jsx("option",{value:"true",children:"True"}),S.jsx("option",{value:"false",children:"False"})]});case"number":return S.jsx("input",{type:"number",value:k,onChange:L=>x(L.target.value),className:"w-full rounded border bg-background p-1 text-sm",autoFocus:!0});case"textarea":return S.jsx("textarea",{value:k,onChange:L=>x(L.target.value),className:"w-full rounded border bg-background p-1 text-sm min-h-[100px]",autoFocus:!0});default:return S.jsx("input",{type:"text",value:k,onChange:L=>x(L.target.value),className:"w-full rounded border bg-background p-1 text-sm",autoFocus:!0})}},U="p-1 rounded hover:bg-accent hover:text-accent-foreground transition-colors";return S.jsxs("div",{className:"space-y-1",children:[S.jsx("div",{className:"font-medium text-sm text-primary mb-1",children:s}),S.jsx("div",{className:"flex items-start gap-2",children:m?S.jsxs(S.Fragment,{children:[S.jsx("div",{className:"flex-grow",children:R()}),S.jsxs("div",{className:"flex items-center gap-1",children:[S.jsx("button",{onClick:Q,disabled:j,className:U,title:"Save",children:S.jsx(Kc,{className:"h-6 w-6"})}),S.jsx("button",{onClick:z,disabled:j,className:U,title:"Cancel",children:S.jsx(Yc,{className:"h-6 w-6"})})]})]}):S.jsxs(S.Fragment,{children:[S.jsx("div",{className:Ot("flex-grow text-sm text-muted-foreground font-mono break-words",typeof k=="string"&&k.length>50&&"text-xs"),children:String(k)}),S.jsxs("div",{className:"flex items-center gap-1",children:[S.jsx("button",{onClick:()=>h(!0),className:U,title:"Edit",children:S.jsx(qp,{className:"h-6 w-6"})}),S.jsx("button",{onClick:M,className:U,title:"Delete",children:S.jsx(nm,{className:"h-6 w-6 text-red-500 hover:text-red-600"})})]})]})})]})},Vm=({onSave:s})=>{const[a,u]=O.useState(!1),[f,m]=O.useState(""),[h,k]=O.useState(""),[x,C]=O.useState("string"),[j,I]=O.useState(!1),A=async()=>{if(f.trim()){I(!0);try{let z=h;switch(x){case"number":z=Number(h);break;case"boolean":z=h==="true";break}await s(f.trim(),z),u(!1),m(""),k(""),C("string")}catch(z){console.error("Failed to add setting:",z)}finally{I(!1)}}},F=()=>{u(!1),m(""),k(""),C("string")};if(!a)return S.jsxs("button",{onClick:()=>u(!0),className:"flex items-center gap-2 text-sm text-muted-foreground hover:text-foreground transition-colors",children:[S.jsx(em,{className:"h-6 w-6"}),"Add Setting"]});const Q="p-1 rounded hover:bg-accent hover:text-accent-foreground transition-colors";return S.jsxs("div",{className:"border rounded-md p-4 space-y-4",children:[S.jsxs("div",{className:"space-y-2",children:[S.jsx("input",{type:"text",placeholder:"Setting name",value:f,onChange:z=>m(z.target.value),className:"w-full rounded border bg-background p-1 text-sm",autoFocus:!0}),S.jsxs("select",{value:x,onChange:z=>C(z.target.value),className:"w-full rounded border bg-background p-1 text-sm",children:[S.jsx("option",{value:"string",children:"String"}),S.jsx("option",{value:"number",children:"Number"}),S.jsx("option",{value:"boolean",children:"Boolean"})]}),x==="boolean"?S.jsxs("select",{value:h,onChange:z=>k(z.target.value),className:"w-full rounded border bg-background p-1 text-sm",children:[S.jsx("option",{value:"true",children:"True"}),S.jsx("option",{value:"false",children:"False"})]}):S.jsx("input",{type:x==="number"?"number":"text",placeholder:"Value",value:h,onChange:z=>k(z.target.value),className:"w-full rounded border bg-background p-1 text-sm"})]}),S.jsxs("div",{className:"flex justify-end gap-2",children:[S.jsx("button",{onClick:A,disabled:j||!f.trim(),className:Ot(Q,"text-green-600 hover:text-green-700"),title:"Save",children:S.jsx(Kc,{className:"h-6 w-6"})}),S.jsx("button",{onClick:F,disabled:j,className:Q,title:"Cancel",children:S.jsx(Yc,{className:"h-6 w-6 text-red-600 hover:text-red-700"})})]})]})},$m=()=>{if(typeof window<"u"){const s=localStorage.getItem("theme-preference");return s?s==="dark":window.matchMedia("(prefers-color-scheme: dark)").matches}return!1},Wm=()=>{const[s,a]=O.useState($m),[u,f]=O.useState([]),[m,h]=O.useState(!0),[k,x]=O.useState(null),[C,j]=O.useState(!1);O.useEffect(()=>{if(!localStorage.getItem("theme-preference")){const M=window.matchMedia("(prefers-color-scheme: dark)"),R=U=>{a(U.matches),document.documentElement.classList.toggle("dark",U.matches)};return M.addEventListener("change",R),()=>M.removeEventListener("change",R)}},[]),O.useEffect(()=>{document.documentElement.classList.toggle("dark",s),localStorage.setItem("theme-preference",s?"dark":"light")},[s]),O.useEffect(()=>{(async()=>{try{const M=await fetch("/api/v1/skills");if(!M.ok)throw new Error("Failed to fetch skills");const U=(await M.json()).map(L=>({...L,settings:Object.fromEntries(Object.entries(L.settings).filter(([b])=>b!=="__mycroft_skill_firstrun").sort(([b],[J])=>b.localeCompare(J)))})).sort((L,b)=>Mr(L.id).name.localeCompare(Mr(b.id).name));f(U)}catch(M){x(M instanceof Error?M.message:"An unknown error occurred")}finally{h(!1)}})()},[]);const I=()=>a(z=>!z),A=()=>j(z=>!z),F=async(z,M)=>{try{const R=await fetch(`/api/v1/skills/${z}`);if(!R.ok)throw new Error("Failed to fetch current settings");const U=await R.json(),L=Object.fromEntries(Object.entries(U.settings).filter(([le])=>le!==M)),b=await fetch(`/api/v1/skills/${z}`,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify(L)});if(!b.ok)throw new Error("Failed to delete setting");const J=await b.json();f(le=>le.map(de=>de.id===z?{...de,settings:Object.fromEntries(Object.entries(J.settings).filter(([xe])=>xe!=="__mycroft_skill_firstrun").sort(([xe],[_e])=>xe.localeCompare(_e)))}:de))}catch(R){throw console.error("Error deleting setting:",R),R}},Q=async(z,M,R)=>{try{const U=await fetch(`/api/v1/skills/${z}/merge`,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify({[M]:R})});if(!U.ok)throw new Error("Failed to save setting");const L=await U.json();f(b=>b.map(J=>J.id===z?{...J,settings:Object.fromEntries(Object.entries(L.settings).filter(([le])=>le!=="__mycroft_skill_firstrun").sort(([le],[de])=>le.localeCompare(de)))}:J))}catch(U){throw console.error("Error saving setting:",U),U}};return m?S.jsx("div",{className:"flex items-center justify-center h-screen bg-background",children:S.jsx(gc,{className:"h-6 w-6 animate-spin text-muted-foreground"})}):k?S.jsx("div",{className:"text-destructive p-4",children:k}):S.jsxs("div",{className:Ot("min-h-screen bg-background text-foreground transition-colors duration-300",s?"dark":""),children:[S.jsx(Fm,{isDark:s,onThemeToggle:I,skills:u}),S.jsxs("main",{className:"max-w-7xl mx-auto p-4 md:p-6",children:[S.jsxs("div",{className:"flex items-center justify-between mb-4",children:[S.jsx("h2",{className:"text-2xl font-bold",children:"Skill Settings"}),S.jsxs("button",{onClick:A,className:Ot("px-4 py-2 rounded-md text-sm font-medium transition-colors","bg-primary/10 text-primary hover:bg-primary/20","flex items-center gap-2"),children:[C?S.jsx(Zp,{className:"h-4 w-4"}):S.jsx(Jp,{className:"h-4 w-4"}),C?"Show Empty Skills":"Hide Empty Skills"]})]}),S.jsx(Im,{type:"single",collapsible:!0,className:"space-y-2",children:u.filter(z=>!C||Object.keys(z.settings).length>0).sort((z,M)=>{const R=Object.keys(z.settings).length===0,U=Object.keys(M.settings).length===0;return R&&!U?1:!R&&U?-1:Mr(z.id).name.localeCompare(Mr(M.id).name)}).map(z=>{const{name:M,author:R}=Mr(z.id),U=Object.keys(z.settings).length;return S.jsxs(nd,{value:z.id,className:Ot("rounded-lg border shadow-sm","transition-colors duration-200","bg-card text-card-foreground","hover:bg-accent/50",U===0&&"opacity-60"),children:[S.jsx(rd,{className:"px-4 py-3 hover:no-underline",children:S.jsxs("div",{className:"flex items-center gap-3 w-full",children:[S.jsx("div",{className:"flex h-8 w-8 items-center justify-center rounded-md bg-primary/10",children:S.jsx(gc,{className:"h-4 w-4 text-primary"})}),S.jsxs("div",{className:"flex-1 text-left",children:[S.jsx("div",{className:"font-semibold",children:M}),S.jsxs("div",{className:"text-sm text-muted-foreground",children:["by ",R," • ",U," settings"]})]})]})}),S.jsx(od,{className:"px-4 pb-4",children:S.jsxs("div",{className:"space-y-4 pt-2",children:[Object.entries(z.settings).map(([L,b])=>S.jsx("div",{className:"border-b border-border pb-3 last:border-0",children:S.jsx(Um,{settingKey:L,value:b,onSave:async(J,le)=>{await Q(z.id,J,le)},onDelete:async J=>{await F(z.id,J)}})},L)),S.jsx("div",{className:"pt-2",children:S.jsx(Vm,{onSave:async(L,b)=>{await Q(z.id,L,b)}})})]})})]},z.id)})})]})]})},Mr=s=>{const a=s.split("."),u=a.length>1?a[a.length-1]:"unknown",f=a.slice(0,-1).join(".")||s;let m=f.replace(/^(skill-|ovos-skill-|ovos-)/,"").split(/[-_]/).map(h=>h.charAt(0).toUpperCase()+h.slice(1)).join(" ");return f.search("skill")||(m=m+" Skill"),{name:m,author:u}};function Bm(){return S.jsx(S.Fragment,{children:S.jsx(Wm,{isDark:!0})})}dp.createRoot(document.getElementById("root")).render(S.jsx(O.StrictMode,{children:S.jsx(Bm,{})}));
