+++ ovos_skill_config/static/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/x-icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>OVOS/Neon Skill Configuration</title>
-    <script type="module" crossorigin src="/assets/index-sHOxXLxz.js"></script>
+    <script type="module" crossorigin src="/assets/index-wAmFI38s.js"></script>
     <link rel="stylesheet" crossorigin href="/assets/index-GG8hYcZw.css">
   </head>
   <body>
